% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Birth}
\alias{Birth}
\title{Live birth rates in 1990 and 1998 for all states}
<<<<<<< HEAD
\format{A data frame with 51 observations on the following 3 variables.
\describe{ 
\item{State}{a factor with levels \code{Alabama}
=======
\format{A data frame with 51 observations on the following three variables.
\describe{ 
\item{state}{a character with levels \code{Alabama}
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
\code{Alaska} \code{Arizona} \code{Arkansas} \code{California}
\code{Colorado} \code{Connecticut} \code{Delaware} \code{District of
Colunbia} \code{Florida} \code{Georgia} \code{Hawaii} \code{Idaho}
\code{Illinois} \code{Indiana} \code{Iowa} \code{Kansas} \code{Kentucky}
\code{Louisiana} \code{Maine} \code{Maryland} \code{Massachusetts}
\code{Michigan} \code{Minnesota} \code{Mississippi} \code{Missour}
\code{Montana} \code{Nebraska} \code{Nevada} \code{New Hampshire} \code{New
Jersey} \code{New Mexico} \code{New York} \code{North Carolina} \code{North
Dakota} \code{Ohio} \code{Oklahoma} \code{Oregon} \code{Pennsylvania}
\code{Rhode Island} \code{South Carolina} \code{South Dakota}
\code{Tennessee} \code{Texas} \code{Utah} \code{Vermont} \code{Virginia}
\code{Washington} \code{West Virginia} \code{Wisconsin} \code{Wyoming}}
<<<<<<< HEAD
\item{X1990rate}{a numeric vector} 
\item{X1998rate}{a numeric vector} 
=======
\item{rate}{live birth rates per 1000 population} 
\item{year}{a factor with levels \code{1990} and \code{1998}} 
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}}
\usage{
Birth
}
\description{
Data for Exercise 1.10
}
\examples{

<<<<<<< HEAD
str(Birth)
attach(Birth)
stem(X1998rate)
hist(X1998rate,breaks=seq(10.9,21.9,1.0),xlab="1998 Birth Rate",
main="Figure 1.14 in BSDA",col="pink")
hist(X1998rate,breaks=seq(10.9,21.9,1.0),xlab="1998 Birth Rate",
main="Figure 1.14 in BSDA",col="pink",prob=TRUE)
lines(density(X1998rate),col="red",lwd=2)
detach(Birth)

=======
rate1998 <- subset(Birth, year == "1998", select = rate, drop = TRUE)
stem(x = rate1998, scale = 2)
hist(rate1998, breaks = seq(10.9, 21.9, 1.0), xlab = "1998 Birth Rate",
     main = "Figure 1.14 in BSDA", col = "pink")
hist(rate1998, breaks = seq(10.9, 21.9, 1.0), xlab = "1998 Birth Rate",
     main = "Figure 1.16 in BSDA", col = "pink", freq = FALSE)      
lines(density(rate1998), lwd = 3)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

