\name{Books}
\alias{Books}
\docType{data}
\title{Number of books read and final spelling scores for 17 third graders}
\description{
  Data for Exercise 9.53
}
\usage{Books}
\format{
  A data frame with 17 observations on the following 2 variables.
  \describe{
    \item{\code{book}}{a numeric vector}
    \item{\code{spelling}}{a numeric vector}
  }
}
\source{Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}. Duxbury}

\examples{
data(Books)
}
\keyword{datasets}
