\name{BSDA-package}
\alias{BSDA-package}
\alias{BSDA}
\docType{package}
\title{Basic Statistics and Data Analysis}
\description{
Data and functions for the book \emph{Basic Statistics and Data Analysis}
}
\details{
\tabular{ll}{
Package: \tab PASWR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-04-16\cr
License: \tab GPL (>=2) \cr
}
The package BSDA provides data and functions for the book
\emph{Basic Statistics and Data Analysis}
}
\author{
Alan T. Arnholt

Maintainer:  <arnholtat@appstate.edu>
}
\references{
Kitchens, L. J (2003)
\emph{Basic Statistics and Data Analysis}. Brooks/Cole.}

\keyword{package}