% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combinations.R
\name{Combinations}
\alias{Combinations}
\title{Combinations}
\usage{
Combinations(n, k)
}
\arguments{
\item{n}{a number.}

\item{k}{a number less than or equal to \code{n}.}
}
\value{
Returns a matrix containing the possible combinations of \code{n}
objects taken \code{k} at a time.
}
\description{
Computes all possible combinations of \code{n} objects taken \code{k} at a
time.
}
\examples{

Combinations(5,2)
    # The columns in the matrix list the values of the 10 possible
    # combinations of 5 things taken 2 at a time.

}
\seealso{
\code{\link{SRS}}
}
\keyword{distribution}

