% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Diesel}
\alias{Diesel}
\title{Diesel fuel prices in 1999-2000 in nine regions of the country}
\format{A data frame with 650 observations on the following 3 variables.
\describe{ 
\item{Date}{date when price was recorded}
\item{PricePerGallon}{price per gallon (in dollars)}
\item{Location}{a factor with levels \code{California}, \code{CentralAtlantic},
\code{Coast}, \code{EastCoast}, \code{Gulf}, \code{LowerAtlantic}, \code{NatAvg},
\code{NorthEast}, \code{Rocky}, and \code{WesternMountain}}
}}
\usage{
Diesel
}
\description{
Data for Exercise 2.8
}
\examples{

str(Diesel)
par(las = 2)
boxplot(PricePerGallon ~ Location, data = Diesel)
boxplot(PricePerGallon ~ Location, 
         data = droplevels(Diesel[Diesel$Location == "EastCoast" | 
         Diesel$Location == "Gulf" | Diesel$Location == "NatAvg" | 
         Diesel$Location == "Rocky" | Diesel$Location == "California", ]), 
         col = "pink", main = "Exercise 2.8")
par(las = 1) 
\dontrun{
ggplot2::ggplot(data = Diesel, aes(x = Date, y = PricePerGallon, 
               color = Location)) + 
               geom_point() + 
               geom_smooth(se = FALSE) + 
               theme_bw() + 
               labs(y = "Price per Gallon (in dollars)")
}         
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

