% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Dogs}
\alias{Dogs}
\title{Rankings of the favorite breeds of dogs}
\format{A data frame with 20 observations on the following 5 variables.
\describe{ 
\item{Dog}{a factor with levels \code{Beagle}
\code{Boxer} \code{Chihuahua} \code{Chow} \code{Dachshund} \code{Dalmatian}
\code{Doberman} \code{Huskie} \code{Labrador} \code{Pomeranian}
\code{Poodle} \code{Retriever} \code{Rotweiler} \code{Schnauzer}
\code{Shepherd} \code{Shetland} \code{ShihTzu} \code{Spaniel}
\code{Springer} \code{Yorkshire}}
\item{X1992}{a numeric vector}
\item{X1993}{a numeric vector} 
\item{X1997}{a numeric vector} 
\item{X1998}{a numeric vector} 
}}
\description{
Data for Exercise 2.88
}
\examples{

str(Dogs)
attach(Dogs)
cor(Dogs[,2:5])
detach(Dogs)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

