% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Eggs}
\alias{Eggs}
\title{Number of eggs versus amounts of feed supplement}
\format{A data frame with 12 observations on the following 2 variables.
\describe{ 
\item{feed}{a numeric vector} 
\item{eggs}{a numeric vector} 
}}
\usage{
Eggs
}
\description{
Data for Exercise 9.22
}
\examples{

str(Eggs)
model <- lm(Eggs$eggs~Eggs$feed)
plot(Eggs$feed,Eggs$eggs)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

