% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Energy}
\alias{Energy}
\title{Amount of energy consumed by homes versus their sizes}
\format{A data frame with 12 observations on the following 2 variables.
\describe{ 
\item{Size}{a numeric vector} 
\item{kilowatt}{a numeric vector} 
}}
\usage{
Energy
}
\description{
Data for Exercises 2.5, 2.24, and 2.55
}
\examples{

str(Energy)
plot(Energy$Size,Energy$kilowatt)
cor(Energy$Size,Energy$kilowatt)
model <- lm(Energy$kilowatt~Energy$Size)
plot(Energy$Size,resid(model))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

