% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Engineer}
\alias{Engineer}
\title{Salaries after 10 years for graduates of three different universities}
\format{An object of class \code{data.frame} with 51 rows and 3 columns.}
\usage{
Engineer
}
\description{
Data for Example 10.7
}
\examples{

str(Engineer)
boxplot(Engineer$Salary~Engineer$University)
kruskal.test(Engineer$Salary~Engineer$University)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

