% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Faithful}
\alias{Faithful}
\title{Waiting times between successive eruptions of the Old Faithful geyser}
\format{A data frame with 299 observations on the following 2 variables.
\describe{ 
\item{Time}{a numeric vector} 
\item{Eruption}{a numeric vector} 
}}
\usage{
Faithful
}
\description{
Data for Exercise 5.12 and 5.111
}
\examples{

str(Faithful)
hist(Faithful$Time,prob=TRUE,xlab="Waiting time between eruptions",col="tomato")
lines(density(Faithful$Time),col="red",lwd=3)
t.test(Faithful$Time)$conf

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

