% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Ferraro1}
\alias{Ferraro1}
\title{Choice of presidental ticket in 1984 by gender}
\format{An object of class \code{data.frame} with 1000 rows and 2 columns.}
\usage{
Ferraro1
}
\description{
Data for Exercise 8.23
}
\examples{

str(Ferraro1)
T1 <- xtabs(~gender + vote, data = Ferraro1)
T1
chisq.test(T1)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

