% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Ferraro2}
\alias{Ferraro2}
\title{Choice of vice presidental candidate in 1984 by gender}
\format{An object of class \code{data.frame} with 1000 rows and 2 columns.}
\usage{
Ferraro2
}
\description{
Data for Exercise 8.23
}
\examples{

str(Ferraro2)
T2 <- xtabs(~gender + vote, data = Ferraro2)
T2
chisq.test(T2)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

