% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Fish}
\alias{Fish}
\title{Length and number of fish caught with small and large mesh codend}
\format{An object of class \code{data.frame} with 1506 rows and 2 columns.}
\usage{
Fish
}
\description{
Data for Exercises 5.83, 5.119, and 7.29
}
\examples{

str(Fish)
TF <- xtabs(~mesh + Length, data = Fish)
TF
median(TF[2,],na.rm=TRUE) #is.na applied to null error
median(TF[1,])
IQR(TF[2,],na.rm=TRUE)
IQR(TF[1,])
SIGN.test(TF[2,],conf.level=.99)
SIGN.test(TF[1,],conf.level=.99)
t.test(TF[2,],TF[1,])

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

