% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{German}
\alias{German}
\title{Number of errors in copying a German passage before and after an
experimental course in German}
\format{A data frame with 10 observations on the following 2 variables.
\describe{ 
\item{Before}{a numeric vector} 
\item{After}{a numeric vector} 
}}
\usage{
German
}
\description{
Data for Exercise 7.60
}
\examples{

str(German)
dif <- (German$Before - German$After)
qqnorm(dif)
qqline(dif)
shapiro.test(dif)
wilcox.test(German$Before,German$After,paired=TRUE)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

