% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Gpa}
\alias{Gpa}
\title{High school GPA versus college GPA}
\format{A data frame with 10 observations on the following 2 variables.
\describe{ 
\item{HSGPA}{a numeric vector} 
\item{CollGPA}{a numeric vector} 
}}
\usage{
Gpa
}
\description{
Data for Exercise 2.13
}
\examples{

str(Gpa)
plot(Gpa$HSGPA,Gpa$CollGPA)
model <- lm(Gpa$CollGPA~Gpa$HSGPA)
abline(model)
model
r <- resid(model)
yhat <- fitted(model)
Table2.1 <- cbind(Gpa$HSGPA,Gpa$CollGPA,yhat,r)
Table2.1

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

