% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Groupabc}
\alias{Groupabc}
\title{Group data to illustrate analysis of variance}
\format{A data frame with 45 observations on the following 2 variables.
\describe{ 
\item{Group}{a factor with levels \code{GroupA} \code{GroupB} \code{GroupC}
}
\item{Value}{a numeric vector} 
}}
\usage{
Groupabc
}
\description{
Data for Exercise 10.42
}
\examples{

str(Groupabc)
boxplot(Value~Group,col=c("red","blue","green"),data=Groupabc)
anova(lm(Value~Group,data=Groupabc))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

