% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Groups}
\alias{Groups}
\title{An illustration of analysis of variance}
\format{A data frame with 26 observations on the following 3 variables.
\describe{ 
\item{Group}{a factor with levels \code{GroupA} \code{GroupB} \code{GroupC}
}
\item{Value}{a numeric vector} 
}}
\usage{
Groups
}
\description{
Data for Exercise 10.4
}
\examples{

str(Groups)
boxplot(Score~Group,col=c("red","blue","green"),data=Groups)
anova(lm(Score~Group,data=Groups)) 

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

