% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Gym}
\alias{Gym}
\title{Children's age versus number of completed gymnastic activities}
\format{A data frame with 8 observations on the following 2 variables.
\describe{
\item{age}{a numeric vector} 
\item{number}{a numeric vector} 
}}
\usage{
Gym
}
\description{
Data for Exercises 2.21, 9.14, and 9.32
}
\examples{

str(Gym)
plot(Gym$age,Gym$number)
model <- lm(Gym$number~Gym$age)
abline(model)
cor(Gym$age,Gym$number)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

