% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Habits}
\alias{Habits}
\title{Study habits of students in two matched school districts}
\format{A data frame with 11 observations on the following 2 variables.
\describe{ 
\item{A}{a numeric vector} 
\item{B}{a numeric vector} 
}}
\usage{
Habits
}
\description{
Data for Exercise 7.57
}
\examples{

str(Habits)
difference <- (Habits$B - Habits$A)
qqnorm(difference)
qqline(difference)
shapiro.test(difference)
t.test(Habits$B,Habits$A,paired=TRUE,alternative="less")
wilcox.test(Habits$B,Habits$A,paired=TRUE,alternative="less")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

