% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Haptoglo}
\alias{Haptoglo}
\title{Haptoglobin concentration in blood serum of 8 healthy adults}
\format{A data frame with 8 observations on the following variable.
\describe{ 
\item{concent}{a numeric vector} 
}}
\usage{
Haptoglo
}
\description{
Data for Example 6.9
}
\examples{

str(Haptoglo)
qqnorm(Haptoglo$concent,col="blue")
qqline(Haptoglo$concent,col="red")
shapiro.test(Haptoglo$concent)
t.test(Haptoglo$concent,mu=2,alternative="less")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

