% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Hardwood}
\alias{Hardwood}
\title{Tensile strength of Kraft paper for different percentages of hardwood in the
batches of pulp}
\format{A data frame with 19 observations on 2 variables.
\describe{ 
\item{tensile}{a numeric vector}
\item{hardwood}{a numeric vector} 
}}
\usage{
Hardwood
}
\description{
Data for Exercise 9.33
}
\examples{

str(Hardwood)



}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

