% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Heat}
\alias{Heat}
\title{Primary heating sources of homes on indian reservations versus all
households}
\format{A data frame with 6 observations on the following 4 variables.
\describe{ 
\item{Fuel}{a factor with levels \code{Electricity}
\code{Fuel Oil} \code{LP bottled gas} \code{Other} \code{Utility gas}
\code{Wood}} 
\item{Reserv}{a numeric vector} 
\item{All.US}{a numeric vector} 
\item{Not.Rese}{a numeric vector} 
}}
\usage{
Heat
}
\description{
Data for Exercise 1.29
}
\examples{

#barplot(t(Heat),beside=TRUE,legend=TRUE,main="Heating of American Indian Homes")
sum(Heat$Reserv)
sum(Heat$All.US)
sum(Heat$Not.Rese)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

