% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Heating}
\alias{Heating}
\title{Fuel efficiency ratings for three types of oil heaters}
\format{A data frame with 30 observations on the following 3 variables.
\describe{
\item{Type}{a factor with levels \code{TypeA} \code{TypeB} \code{TypeC}
} 
\item{Rating}{a numeric vector} 
\item{Ranks}{a numeric vector} 
}}
\usage{
Heating
}
\description{
Data for Exercise 10.32
}
\examples{

str(Heating)
boxplot(Heating$Rating~Heating$Type)
kruskal.test(Heating$Rating~as.factor(Heating$Type))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

