% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Honda}
\alias{Honda}
\title{Miles per gallon for a Honda Civic on 35 different occasions}
\format{A data frame with 35 observations on the following variable.
\describe{ 
\item{mileage}{a numeric vector} 
}}
\usage{
Honda
}
\description{
Data for Statistical Insight Chapter 6
}
\examples{

str(Honda)
t.test(Honda$mileage,mu=40,alternative="less")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

