% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Hostile}
\alias{Hostile}
\title{Hostility levels of high school students from rural, suburban, and urban
areas}
\format{A data frame with 45 observations on the following 3 variables.
\describe{ 
\item{Type}{a factor with levels \code{Suburban} \code{Urban} 
} 
\item{HLT}{a numeric vector} 
\item{Ranks}{a numeric vector} 
}}
\usage{
Hostile
}
\description{
Data for Example 10.6
}
\examples{

str(Hostile)
boxplot(Hostile$HLT~Hostile$Type)
kruskal.test(Hostile$HLT~as.factor(Hostile$Type))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

