% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Jobsat}
\alias{Jobsat}
\title{Job satisfaction and stress level for 9 school teachers}
\format{A data frame with 9 observations on the following 2 variables.
\describe{ 
\item{WSPT}{a numeric vector} 
\item{satisfac}{a numeric vector} 
}}
\usage{
Jobsat
}
\description{
Data for Exercise 9.60
}
\examples{

str(Jobsat)
plot(Jobsat$WSPT,Jobsat$satisfac)
model <- lm(Jobsat$satisfac~Jobsat$WSPT)
abline(model)
summary(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

