% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Laminect}
\alias{Laminect}
\title{Median costs of laminectomies at hospitals across North Carolina in 1992}
\format{A data frame with 46 observations on the following 2 variables.
\describe{
 \item{cost}{a numeric vector} 
 \item{class}{a factor with levels \code{Rural} \code{Regional} \code{Metropol}
 } 
 }}
\usage{
Laminect
}
\description{
Data for Exercise 10.18
}
\examples{

str(Laminect)
boxplot(Laminect$cost~Laminect$class)
anova(lm(Laminect$cost~as.factor(Laminect$class)))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

