% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Life}
\alias{Life}
\title{Life expectancy of men and women in U.S.}
\format{A data frame with 8 observations on the following 3 variables.
\describe{ 
\item{year}{a numeric vector} 
\item{Men}{a numeric vector} 
\item{Women}{a numeric vector} 
}}
\usage{
Life
}
\description{
Data for Exercise 1.31
}
\examples{

str(Life)
plot(Life$year,Life$Men,type="l",ylim=c(min(Life$Men,Life$Women),max(Life$Men,Life$Women)),col="blue",
main="Life Expectancy versus Year",ylab="Age",xlab="Year")
lines(Life$year,Life$Women,col="red")
text(1955,65,"Men",col="blue")
text(1955,70,"Women",col="red")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

