% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Lifespan}
\alias{Lifespan}
\title{Life span of electronic components used in a spacecraft versus heat}
\format{A data frame with 6 observations on the following 2 variables.
\describe{ 
\item{heat}{a numeric vector} 
\item{life}{a numeric vector} 
}}
\usage{
Lifespan
}
\description{
Data for Exercise 2.4, 2.37, and 2.49
}
\examples{

str(Lifespan)
plot(Lifespan$heat,Lifespan$life)
model <- lm(Lifespan$life~Lifespan$heat)
model
resid(model)
sum((resid(model))^2)
anova(model)
# plot(model)  # Used for diagnostic purposes

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

