% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Lodge}
\alias{Lodge}
\title{Measured traffic at three prospective locations for a motor lodge}
\format{A data frame with 45 observations on the following 3 variables.
\describe{ 
\item{Site}{a factor with levels \code{SiteA} \code{SiteB} \code{SiteC}
}
\item{Traffic}{a numeric vector} 
\item{Ranks}{a numeric vector} 
}}
\usage{
Lodge
}
\description{
Data for Exercise 10.33
}
\examples{

str(Lodge)
boxplot(Lodge$Traffic~Lodge$Site)
anova(lm(Lodge$Traffic~as.factor(Lodge$Site)))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

