% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Longtail}
\alias{Longtail}
\title{Long-tailed distributions to illustrate Kruskal Wallis test}
\format{A data frame with 60 observations on the following 3 variables.
\describe{ 
\item{Group}{A factor with levels \code{GroupA} \code{GroupB} \code{GroupC}
}
\item{score}{a numeric vector} 
\item{Ranks}{a numeric vector} 
}}
\usage{
Longtail
}
\description{
Data for Exercise 10.45
}
\examples{

str(Longtail)
boxplot(Longtail$Score~Longtail$Group)
kruskal.test(Longtail$Score~as.factor(Longtail$Group))
anova(lm(Longtail$Score~as.factor(Longtail$Group)))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

