% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Lowabil}
\alias{Lowabil}
\title{Reading skills of 24 matched low ability students}
\format{A data frame with 12 observations on the following 2 variables.
\describe{ 
\item{Experimt}{a numeric vector} 
\item{Control}{a numeric vector} 
}}
\usage{
Lowabil
}
\description{
Data for Example 7.18
}
\examples{

str(Lowabil)
DIF <- Lowabil$experiment - Lowabil$control
qqnorm(DIF) #error has no y 
qqline(DIF)
shapiro.test(DIF)
t.test(Lowabil$experiment,Lowabil$control,paired=TRUE)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

