% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Median}
\alias{Median}
\title{Illustrates test of equality of medians with the Kruskal Wallis test}
\format{A data frame/tibble with 45 observations on the following 2 variables.
\describe{ 
\item{sample}{a vector with values \code{Sample1}, \code{Sample 2}, and \code{Sample 3}} 
\item{value}{a numeric vector} 
}}
\usage{
Median
}
\description{
Data for Exercise 10.52
}
\examples{

boxplot(value ~ sample, data = Median, col = rainbow(3))
anova(lm(value ~ sample, data = Median))
kruskal.test(value ~ factor(sample), data = Median)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
