% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Name}
\alias{Name}
\title{Estimated value of a brand name product and the conpany's revenue}
\format{
A data frame/tibble with 42 observations on three variables
\describe{ 
\item{brand}{a factor with levels \code{Band-Aid},
\code{Barbie}, \code{Birds Eye}, \code{Budweiser}, \code{Camel}, \code{Campbell},
\code{Carlsberg}, \code{Coca-Cola}, \code{Colgate}, \code{Del Monte},
\code{Fisher-Price}, \verb{Gordon's}, \code{Green Giant}, \code{Guinness},
\code{Haagen-Dazs}, \code{Heineken}, \code{Heinz}, \code{Hennessy},
\code{Hermes}, \code{Hershey}, \code{Ivory}, \code{Jell-o}, \code{Johnnie
Walker}, \code{Kellogg}, \code{Kleenex}, \code{Kraft}, \code{Louis Vuitton},
\code{Marlboro}, \code{Nescafe}, \code{Nestle}, \code{Nivea}, \code{Oil of Olay},
\code{Pampers}, \code{Pepsi-Cola}, \code{Planters}, \code{Quaker}, \code{Sara
Lee}, \code{Schweppes}, \code{Smirnoff}, \code{Tampax}, \code{Winston}, and
\verb{Wrigley's}}
\item{value}{value in billions of dollars}
\item{revenue}{revenue in billions of dollars} 
}
}
\source{
Financial World.
}
\usage{
Name
}
\description{
Data for Exercises 2.28, 9.19, and Example 2.8
}
\examples{

plot(value ~ revenue, data = Name)
model <- lm(value ~ revenue, data = Name)
abline(model, col = "red")
cor(Name$value, Name$revenue)
summary(model)
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
