% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normarea.R
\name{normarea}
\alias{normarea}
\title{Normal Area}
\usage{
normarea(lower = -Inf, upper = Inf, m, sig)
}
\arguments{
\item{lower}{the lower value}

\item{upper}{the upper value}

\item{m}{the mean for the population}

\item{sig}{the standard deviation of the population}
}
\description{
Function that computes and draws the area between two user specified values
in a user specified normal distribution with a given mean and standard
deviation
}
\examples{
normarea(70, 130, 100, 15)
    # Finds and P(70 < X < 130) given X is N(100,15).
}
\author{
Alan T. Arnholt
}
\keyword{distribution}

