% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Prejudic}
\alias{Prejudic}
\title{Racial prejudice measured on a sample of 25 high school students}
\format{A data frame with 25 observations on one variable
\describe{ 
\item{prejud}{racial prejudice score} 
}}
\usage{
Prejudic
}
\description{
Data for Exercise 5.10 and 5.22
}
\examples{

stem(Prejudic$prejud)
EDA(Prejudic$prejud)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
