% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Program}
\alias{Program}
\title{Effects of four different methods of programmed learning for statistics
students}
\format{A data frame/tibble with 44 observations on two variables
\describe{ 
\item{method}{a character variable with values \code{method1}, \code{method2}, 
\code{method3}, and \code{method4}} 
\item{score}{standardized test score} 
}}
\usage{
Program
}
\description{
Data for Exercise 10.17
}
\examples{

boxplot(score ~ method, col = c("red", "blue", "green", "yellow"), data = Program)
anova(lm(score ~ method, data = Program))
TukeyHSD(aov(score ~ method, data = Program))
par(mar = c(5.1, 4.1 + 4, 4.1, 2.1))
plot(TukeyHSD(aov(score ~ method, data = Program)), las = 1)
par(mar = c(5.1, 4.1, 4.1, 2.1))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
