% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Situp}
\alias{Situp}
\title{Maximum number of situps by participants in an exercise class}
\format{A data frame/tibble with 20 observations on one variable
\describe{ 
\item{number}{maximum number of situps completed in an exercise class 
after 1 month in the program} 
}}
\usage{
Situp
}
\description{
Data for Exercise 1.47
}
\examples{

stem(Situp$number)
hist(Situp$number, breaks = seq(0, 70, 10), right = FALSE)
hist(Situp$number, breaks = seq(0, 70, 10), right = FALSE, 
     freq = FALSE, col = "pink", main = "Problem 1.47", 
     xlab = "Maximum number of situps")
lines(density(Situp$number), col = "red")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
