% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Tiaa}
\alias{Tiaa}
\title{1999-2000 closing prices for TIAA-CREF stocks}
\format{A data frame/tibble with 365 observations on four variables
\describe{ 
\item{crefstk}{closing price (in dollars)} 
\item{crefgwt}{closing price (in dollars)} 
\item{tiaa}{closing price (in dollars)} 
\item{date}{day of the year} 
}}
\usage{
Tiaa
}
\description{
Data for your enjoyment
}
\examples{

data(Tiaa)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
