% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Abbey}
\alias{Abbey}
\title{Daily price returns (in pence) of Abbey National shares between 7/31/91 and
10/8/91}
\format{A data frame with 50 observations on the following variable.
\describe{ 
\item{C1}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data used in problem 6.39
}
\examples{

attach(Abbey)
EDA(C1)
t.test(C1,mu=300)
detach(Abbey)

}
\keyword{datasets}

