% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Abbey}
\alias{Abbey}
\title{Daily price returns (in pence) of Abbey National shares between 7/31/91 and
10/8/91}
\format{A data frame with 50 observations on the following variable.
\describe{ 
<<<<<<< HEAD
\item{C1}{a numeric vector} 
=======
\item{price}{daily price returns (in pence) of Abbey National shares} 
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}}
\usage{
Abbey
}
\description{
Data used in problem 6.39
}
\examples{

<<<<<<< HEAD
attach(Abbey)
EDA(C1)
t.test(C1,mu=300)
detach(Abbey)

=======
qqnorm(Abbey$price)
qqline(Abbey$price)
t.test(Abbey$price, mu = 300)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

