% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Abbey}
\alias{Abbey}
\title{Daily price returns (in pence) of Abbey National shares between 7/31/91 and
10/8/91}
\format{A data frame with 50 observations on the following variable.
\describe{
\item{price}{daily price returns (in pence) of Abbey National shares}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\usage{
Abbey
}
\description{
Data used in problem 6.39
}
\examples{
qqnorm(Abbey$price)
qqline(Abbey$price)
t.test(Abbey$price, mu = 300)
}
\keyword{datasets}

