% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Abc}
\alias{Abc}
\title{Three samples to illustrate analysis of variance}
\format{A data frame with 18 observations on the following 3 variables.
\describe{ 
\item{GroupA}{a numeric vector}
\item{GroupB}{a numeric vector}
\item{GroupC}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data used in Exercise 10.1
}
\examples{

attach(Abc)
STACKED <-stack(Abc)
STACKED[1:5,]
boxplot(values~ind,col=c("red","blue","green"),data=STACKED)
anova(lm(values~ind,data=STACKED))
remove(STACKED)
detach(Abc)

}
\keyword{datasets}

