% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Abc}
\alias{Abc}
\title{Three samples to illustrate analysis of variance}
\format{
A data frame/tibble with 54 observations on two variables
\describe{ 
\item{response}{a numeric vector}
\item{group}{a character vector \code{A}, \code{B}, and \code{C}}
}
}
\usage{
Abc
}
\description{
Data used in Exercise 10.1
}
\examples{

boxplot(response ~ group, col=c("red", "blue", "green"), data = Abc )
anova(lm(response ~ group, data = Abc))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
