% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Ability}
\alias{Ability}
\title{Perceived math ability for 13-year olds by gender}
\format{
A data frame/tibble with 400 observations on two variables
\describe{ 
\item{gender}{a factor with levels \code{girls} and \code{boys}} 
\item{ability}{a factor with levels  \code{hopeless},  \code{belowavg}, \code{average}, \code{aboveavg}, and \code{superior}}
}
}
\usage{
Ability
}
\description{
Data used in Exercise 8.57
}
\examples{

CT <- xtabs(~gender + ability, data = Ability)
CT
chisq.test(CT)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
