% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Ability}
\alias{Ability}
\title{Perceived math ability for 13-year olds by gender}
\format{A data frame with 2 observations on the following 6 variables.
\describe{
\item{gender}{a factor with levels \code{boys} \code{girls}}
\item{hopeless}{a numeric vector}
\item{belowavg}{a numeric vector}
\item{average}{a numeric vector}
\item{aboveavg}{a numeric vector}
\item{superior}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data used in Exercise 8.57
}
\examples{
X <- as.matrix(Ability[1:2,2:6])
chisq.test(X)
}
\keyword{datasets}

