% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Abortion}
\alias{Abortion}
\title{Abortion rate by region of country}
\format{A data frame with 51 observations on the following 10 variables.
\describe{
\item{state}{a factor with levels \code{alabama}
\code{alaska} \code{arizona} \code{arkansas} \code{california}
\code{colorado} \code{connecticut} \code{delaware} \code{dist of columbia}
\code{florida} \code{georgia} \code{hawaii} \code{idaho} \code{illinois}
\code{indiana} \code{iowa} \code{kansas} \code{kentucky} \code{louisiana}
\code{maine} \code{maryland} \code{massachusetts} \code{michigan}
\code{minnesota} \code{mississippi} \code{missouri} \code{montana}
\code{nebraska} \code{nevada} \code{new hampshire} \code{new jersey}
\code{new mexico} \code{new york} \code{north carolina} \code{north dakota}
\code{ohio} \code{oklahoma} \code{oregon} \code{pennsylvania} \code{rhode
island} \code{south carolina} \code{south dakota} \code{tennessee}
\code{texas} \code{utah} \code{vermont} \code{virginia} \code{washington}
\code{west virginia} \code{wisconsin} \code{wyoming}}
\item{region}{a factor with levels \code{midwest} \code{northeast}
\code{south} \code{west}}
\item{regcode}{a numeric vector}
\item{rate1988}{a numeric vector}
\item{rate1992}{a numeric vector}
\item{rate1996}{a numeric vector}
\item{provide1988}{a numeric vector}
\item{provide1992}{a numeric vector}
\item{lowhigh}{a numeric vector}
\item{rate}{a factor with levels \code{Low} and \code{High}}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\usage{
Abortion
}
\description{
Data used in Exercise 8.51
}
\examples{
T1 <- xtabs(~region + rate, data = Abortion)
T1
chisq.test(T1)
}
\keyword{datasets}

