% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Abortion}
\alias{Abortion}
\title{Abortion rate by region of country}
<<<<<<< HEAD
\format{A data frame with 51 observations on the following 9 variables.
=======
\format{A data frame with 51 observations on the following 10 variables.
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
\describe{ 
\item{state}{a factor with levels \code{alabama}
\code{alaska} \code{arizona} \code{arkansas} \code{california}
\code{colorado} \code{connecticut} \code{delaware} \code{dist of columbia}
\code{florida} \code{georgia} \code{hawaii} \code{idaho} \code{illinois}
\code{indiana} \code{iowa} \code{kansas} \code{kentucky} \code{louisiana}
\code{maine} \code{maryland} \code{massachusetts} \code{michigan}
\code{minnesota} \code{mississippi} \code{missouri} \code{montana}
\code{nebraska} \code{nevada} \code{new hampshire} \code{new jersey}
\code{new mexico} \code{new york} \code{north carolina} \code{north dakota}
\code{ohio} \code{oklahoma} \code{oregon} \code{pennsylvania} \code{rhode
island} \code{south carolina} \code{south dakota} \code{tennessee}
\code{texas} \code{utah} \code{vermont} \code{virginia} \code{washington}
\code{west virginia} \code{wisconsin} \code{wyoming}}
\item{region}{a factor with levels \code{midwest} \code{northeast}
\code{south} \code{west}} 
\item{regcode}{a numeric vector}
<<<<<<< HEAD
\item{X88rate}{a numeric vector} 
\item{X92rate}{a numeric vector} 
\item{X96rate}{a numeric vector} 
\item{X88provid}{a numeric vector} 
\item{X92provid}{a numeric vector}
\item{lowhigh}{a numeric vector} 
=======
\item{rate1988}{a numeric vector} 
\item{rate1992}{a numeric vector} 
\item{rate1996}{a numeric vector} 
\item{provide1988}{a numeric vector} 
\item{provide1992}{a numeric vector}
\item{lowhigh}{a numeric vector} 
\item{rate}{a factor with levels \code{Low} and \code{High}}
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}}
\usage{
Abortion
}
\description{
Data used in Exercise 8.51
}
\examples{

<<<<<<< HEAD
attach(Abortion)
AbortionRate <- cut(X96rate,breaks=c(0,20,10000) )
levels(AbortionRate) <- c("Low","High")
table(region,AbortionRate)
chisq.test(table(region,AbortionRate))
detach(Abortion)

=======
T1 <- xtabs(~region + rate, data = Abortion)
T1
chisq.test(T1)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

