% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Absent}
\alias{Absent}
\title{Number of absent days for 20 employees}
\format{A data frame with 20 observations on the following 4 variables.
\describe{
\item{days}{a numeric vector}
\item{days_1}{a numeric vector}
\item{Count}{a numeric vector}
\item{Percent}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data used in Exercise 1.28
}
\examples{
data(Absent)
attach(Absent)
table(days)
barplot(table(days),col="pink")
detach(Absent)
}
\keyword{datasets}

