% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Absent}
\alias{Absent}
\title{Number of absent days for 20 employees}
\format{A data frame/tibble with 20 observations on one variable
\describe{ 
\item{days}{a numeric vector} 
}}
\usage{
Absent
}
\description{
Data used in Exercise 1.28
}
\examples{

CT <- xtabs(~ days, data = Absent)
CT
barplot(CT, col = "pink")
plot(ecdf(Absent$days), main = "ECDF")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
