% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Achieve}
\alias{Achieve}
\title{Math achievement test scores by gender for 25 high school students}
\format{A data frame with 25 observations on the following 4 variables.
\describe{
\item{Score}{a numeric vector}
\item{Gender}{a numeric vector}
\item{Female}{a numeric vector}
\item{Male}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data used in Example 7.14 and Exercise 10.7
}
\examples{
str(Achieve)
attach(Achieve)
anova(lm(Score~Gender))
t.test(Female,Male,var.equal=TRUE)
detach(Achieve)
}
\keyword{datasets}

