% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Achieve}
\alias{Achieve}
\title{Math achievement test scores by gender for 25 high school students}
\format{A data frame/tibble with 25 observations on two variables
\describe{ 
\item{score}{mathematics achiement score} 
\item{gender}{a factor with 2 levels \code{boys} and \code{girls}} 
}}
\usage{
Achieve
}
\description{
Data used in Example 7.14 and Exercise 10.7
}
\examples{

anova(lm(score ~ gender, data = Achieve))
t.test(score ~ gender, var.equal = TRUE, data = Achieve)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
