% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Adsales}
\alias{Adsales}
\title{Number of ads versus number of sales for a retailer of satellite dishes}
\format{A data frame with 6 observations on the following 2 variables.
\describe{ 
\item{ads}{a numeric vector} 
\item{sales}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data used in Exercise 9.15
}
\examples{

attach(Adsales)
plot(ads,sales)
linmod <- lm(sales~ads)
abline(linmod)
summary(linmod)
detach(Adsales)

}
\keyword{datasets}

