% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Adsales}
\alias{Adsales}
\title{Number of ads versus number of sales for a retailer of satellite dishes}
\format{A data frame/tibble with six observations on three variables
\describe{ 
\item{month}{a character vector listing month}
\item{ads}{a numeric vector containing number of ads} 
\item{sales}{a numeric vector containing number of sales} 
}}
\usage{
Adsales
}
\description{
Data used in Exercise 9.15
}
\examples{

plot(sales ~ ads, data = Adsales, main = "Exercise 9.15")
mod <- lm(sales ~ ads, data = Adsales)
abline(mod, col = "red")
summary(mod)
predict(mod, newdata = data.frame(ads = 6), interval = "conf", level = 0.99)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
