% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Aggress}
\alias{Aggress}
\title{Agressive tendency scores for a group of teenage members of a street gang}
\format{
A data frame/tibble with 28 observations on one variable
\describe{ 
\item{aggres}{measure of aggresive tendency, ranging from 10-50} 
}
}
\usage{
Aggress
}
\description{
Data used in Exercises 1.66 and 1.81
}
\examples{

with(data = Aggress,
     EDA(aggres))
# OR
IQR(Aggress$aggres)
diff(range(Aggress$aggres))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
