% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Aggress}
\alias{Aggress}
\title{Agressive tendency scores for a group of teenage members of a street gang}
\format{A data frame with 28 observations on the following variable.
\describe{ 
\item{aggres}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\usage{
Aggress
}
\description{
Data used in Exercises 1.61 and 1.81
}
\examples{

str(Aggress)
with(data = Aggress,
     EDA(aggres))
# OR
IQR(Aggress$aggres)
diff(range(Aggress$aggres))

}
\keyword{datasets}

