% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Airline}
\alias{Airline}
\title{Percentage of on-time arrivals and number of complaints for 11 airlines}
\format{A data frame with 11 observations on the following 3 variables.
\describe{ 
\item{airline}{a factor with levels \code{Alaska}
\code{Amer West} \code{American} \code{Continental} \code{Delta}
\code{Northwest} \code{Pan Am} \code{Southwest} \code{TWA} \code{United}
\code{USAir}} 
\item{ontime}{a numeric vector}
\item{complnt}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 2.9
}
\examples{

str(Airline)
attach(Airline)
barplot(complnt,names.arg=airline,col="lightblue")
plot(ontime,complnt)
detach(Airline)

}
\keyword{datasets}

