% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Alcohol}
\alias{Alcohol}
\title{Ages at which 14 female alcoholics began drinking}
\format{A data frame with 14 observations on the following variable.
\describe{ 
\item{age}{a numeric vector} }}
\usage{
Alcohol
}
\description{
Data used in Exercise 5.79
}
\examples{

qqnorm(Alcohol$age)
qqline(Alcohol$age)
SIGN.test(Alcohol$age, md = 20, conf.level = 0.99)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

