% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Allergy}
\alias{Allergy}
\title{Allergy medicines by adverse events}
\format{A data frame with 406 observations on 2 variables.
\describe{ 
\item{event}{a factor with levels \code{insomnia}
\code{headache} \code{drowsiness}} 
\item{medication}{a factor with levels \code{seldane-d}
\code{pseudoephedrine} \code{placebo}}
 }}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\usage{
Allergy
}
\description{
Data used in Exercise 8.22
}
\examples{

T1 <- xtabs(~event + medication, data = Allergy)
T1
chisq.test(T1)

}
\keyword{datasets}

