% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Allergy}
\alias{Allergy}
\title{Allergy medicines by adverse events}
\format{A data frame with 3 observations on the following 4 variables.
\describe{ 
\item{C1.T}{a factor with levels \code{Drowsiness}
\code{Headache} \code{Insomnia}} 
\item{Seldane}{a numeric vector}
\item{Pseudoep}{a numeric vector} 
\item{Placebo}{a numeric vector}
 }}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data used in Exercise 8.22
}
\examples{

attach(Allergy)
X <- as.matrix(Allergy[1:3, 2:4])
chisq.test(X)
Xr <- as.matrix(Allergy[2:3, 2:4])
chisq.test(Xr)
detach(Allergy)

}
\keyword{datasets}

