% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Anesthet}
\alias{Anesthet}
\title{Recovery times for anesthetized patients}
\format{A data frame with 10 observations on the following variable.
\describe{ 
<<<<<<< HEAD
\item{recover}{a numeric vector} 
=======
\item{recover}{recovery time (in hours)} 
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}}
\usage{
Anesthet
}
\description{
Data used in Exercise 5.58
}
\examples{

<<<<<<< HEAD
attach(Anesthet)
str(Anesthet)
qqnorm(recover)
qqline(recover)
t.test(recover,conf.level=0.90)$conf
detach(Anesthet)

=======
qqnorm(Anesthet$recover)
qqline(Anesthet$recover)
with(data = Anesthet,
t.test(recover, conf.level = 0.90)$conf
)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

