% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Anesthet}
\alias{Anesthet}
\title{Recovery times for anesthetized patients}
\format{A data frame with 10 observations on the following variable.
\describe{ 
\item{recover}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data used in Exercise 5.58
}
\examples{

attach(Anesthet)
str(Anesthet)
qqnorm(recover)
qqline(recover)
t.test(recover,conf.level=0.90)$conf
detach(Anesthet)

}
\keyword{datasets}

