% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Anxiety}
\alias{Anxiety}
\title{Math test scores versus anxiety scores before the test}
\format{A data frame/tibble  with 20 observations on the following two variables.
\describe{ 
\item{anxiety}{anxiety score before a major math test} 
\item{math}{math test score} 
}}
\usage{
Anxiety
}
\description{
Data used in Exercise 2.96
}
\examples{

plot(math ~ anxiety, data = Anxiety)
with(data = Anxiety,
cor(math, anxiety)
)
linmod <- lm(math ~ anxiety, data = Anxiety)
abline(linmod)
summary(linmod)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
