% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Apolipop}
\alias{Apolipop}
\title{Level of apolipoprotein B and number of cups of coffee consumed per day for
15 adult males}
\format{A data frame with 15 observations on the following 4 variables.
\describe{
\item{coffee}{a numeric vector}
\item{apolipB}{a numeric vector}
\item{SRES1}{a numeric vector}
\item{FITS1}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data used in Examples 9.2 and 9.9
}
\examples{
attach(Apolipop)
str(Apolipop)
plot(coffee,apolipB)
linmod <- lm(apolipB~coffee)
summary(linmod)
# plot(linmod)
detach(Apolipop)
}
\keyword{datasets}

