% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Append}
\alias{Append}
\title{Median costs of an appendectomy at 20 hospitals in North Carolina}
\format{A data frame with 20 observations on the following variable.
\describe{ 
\item{fee}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 1.119
}
\examples{

str(Append)
attach(Append)
ll <- mean(fee)-2*sd(fee)
ul <- mean(fee)+2*sd(fee)
limits <-c(ll,ul)
limits
fee[fee<ll | fee>ul]
detach(Append)

}
\keyword{datasets}

