% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Appendec}
\alias{Appendec}
\title{Median costs of appendectomies at three different types of North Carolina
hospitals}
\format{A data frame with 59 observations on the following 5 variables.
\describe{
\item{Cost}{a numeric vector}
\item{Region}{a numeric vector}
\item{Rural}{a numeric vector}
\item{Regional}{a numeric vector}
\item{Metropol}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 10.60
}
\examples{
str(Appendec)
attach(Appendec)
boxplot(Cost~Region)
anova(lm(Cost~as.factor(Region)))
detach(Appendec)
}
\keyword{datasets}

