% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Aptitude}
\alias{Aptitude}
\title{Aptitude test scores versus productivity in a factory}
\format{A data frame with 8 observations on the following 4 variables.
\describe{
\item{aptitude}{a numeric vector}
\item{product}{a numeric vector}
\item{SRES1}{a numeric vector}
\item{FITS1}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercises 2.1, 2.35 and 2.51
}
\examples{
str(Aptitude)
attach(Aptitude)
plot(aptitude,product,main="Exercise 2.1")
model1 <- lm(product~aptitude)
model1
abline(model1,col="red",lwd=3)
resid(model1)
fitted(model1)
cor(product,aptitude)
detach(Aptitude)
}
\keyword{datasets}

