% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Archaeo}
\alias{Archaeo}
\title{Radiocarbon ages of observations taken from an archaeological site}
<<<<<<< HEAD
\format{A data frame with 60 observations on the following 6 variables.
\describe{ 
\item{phase1}{a numeric vector} 
\item{phase2}{a numeric vector} 
\item{phase3}{a numeric vector}
\item{phase4}{a numeric vector} 
\item{age}{a numeric vector}
\item{phase}{a numeric vector} 
=======
\format{A data frame with 60 observations on the following two variables.
\describe{ 
\item{age}{number of years before 1983 - the year the data were obtained}
\item{phase}{Ceramic Phase numbers} 
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}}
\usage{
Archaeo
}
\description{
Data for Exercises 5.120, 10.20 and Example 1.16
}
\examples{

<<<<<<< HEAD
str(Archaeo)
attach(Archaeo)
boxplot(age~phase,col="yellow",main="Example 1.16",xlab="Phase",ylab="Age")
anova(lm(age~as.factor(phase)))
detach(Archaeo)

=======
boxplot(age ~ phase, data = Archaeo, col = "yellow", 
        main = "Example 1.16", xlab = "Ceramic Phase", ylab = "Age")
anova(lm(age ~ as.factor(phase), data= Archaeo))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

