% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Archaeo}
\alias{Archaeo}
\title{Radiocarbon ages of observations taken from an archaeological site}
\format{A data frame with 60 observations on the following 6 variables.
\describe{ 
\item{phase1}{a numeric vector} 
\item{phase2}{a numeric vector} 
\item{phase3}{a numeric vector}
\item{phase4}{a numeric vector} 
\item{age}{a numeric vector}
\item{phase}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercises 5.120, 10.20 and Example 1.16
}
\examples{

str(Archaeo)
attach(Archaeo)
boxplot(age~phase,col="yellow",main="Example 1.16",xlab="Phase",ylab="Age")
anova(lm(age~as.factor(phase)))
detach(Archaeo)

}
\keyword{datasets}

