% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Arthriti}
\alias{Arthriti}
\title{Time of relief for three treatments of arthritis}
\format{A data frame with 51 observations on the following 5 variables.
\describe{
\item{TreatA}{a numeric vector}
\item{TreatB}{a numeric vector}
\item{TreatC}{a numeric vector}
\item{Time}{a numeric vector}
\item{Treatmnt}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise
}
\examples{
str(Arthriti)
attach(Arthriti)
boxplot(Time~Treatmnt)
anova(lm(Time~as.factor(Treatmnt)))
detach(Arthriti)
}
\keyword{datasets}

