% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Arthriti}
\alias{Arthriti}
\title{Time of relief for three treatments of arthritis}
\format{A data frame with 51 observations on the following 2 variables.
\describe{ 
\item{time}{time (days?) until an arthritis sufferer experienced relief} 
\item{treatment}{a factor with levels \code{A}, \code{B}, and \code{C}} 
}}
\usage{
Arthriti
}
\description{
Data for Exercise 10.58
}
\examples{

boxplot(time ~ treatment, data = Arthriti)
anova(lm(time ~ treatment, data = Arthriti))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

