% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Artifici}
\alias{Artifici}
\title{Durations of operation for 15 artificial heart transplants}
\format{A data frame with 15 observations on the following variable.
\describe{
\item{duration}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 1.107
}
\examples{
str(Artifici)
attach(Artifici)
stem(duration)
summary(duration)
values <- duration[duration<6.5]
values
summary(values)
detach(Artifici)
remove(values)
}
\keyword{datasets}

