% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Artifici}
\alias{Artifici}
\title{Durations of operation for 15 artificial heart transplants}
\format{A data frame with 15 observations on the following variable.
\describe{ 
\item{duration}{duration (in hours) for transplant} 
}}
\usage{
Artifici
}
\description{
Data for Exercise 1.107
}
\examples{

stem(Artifici$duration, 2)
summary(Artifici$duration)
values <- Artifici$duration[Artifici$duration < 6.5]
values
summary(values)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
