% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Asprin}
\alias{Asprin}
\title{Dissolving time versus level of impurities in aspirin tablets}
\format{A data frame with 5 observations on the following 3 variables.
\describe{
\item{X1.}{a numeric vector}
\item{X5.}{a numeric vector}
\item{X10.}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 10.51
}
\examples{
str(Asprin)
attach(Asprin)
STACKED <-stack(Asprin)
STACKED[1:5,]
boxplot(values~ind,col=c("red","blue","green"),data=STACKED)
anova(lm(values~ind,data=STACKED))
remove(STACKED)
detach(Asprin)
}
\keyword{datasets}

