% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Asprin}
\alias{Aspirin}
\alias{Asprin}
\title{Dissolving time versus level of impurities in aspirin tablets}
\format{A data frame with 15 observations on 2 variables.
\describe{ 
\item{time}{time (in seconds) for aspirin to dissolve} 
\item{impurity}{impurity of an ingredient with levels \code{1\%}, \code{5\%}, and \code{10\%}} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\usage{
Aspirin
}
\description{
Data for Exercise 10.51
}
\examples{

boxplot(time ~ impurity, data = Aspirin, 
        col = c("red", "blue", "green"))

}
\keyword{datasets}

