% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Asthmati}
\alias{Asthmati}
\title{Asthmatic relief index on 9 subjects given a drug and a placebo}
\format{A data frame with 9 observations on the following 3 variables.
\describe{
\item{Drug}{a numeric vector}
\item{Placebo}{a numeric vector}
\item{differ}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.52
}
\examples{
str(Asthmati)
attach(Asthmati)
qqnorm(differ)
qqline(differ)
shapiro.test(differ)
t.test(Placebo,Drug,paired=TRUE,mu=0,alternative="greater")
detach(Asthmati)
}
\keyword{datasets}

