% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Asthmati}
\alias{Asthmati}
\title{Asthmatic relief index on nine subjects given a drug and a placebo}
\format{A data frame/tibble  with nine observations on three variables
\describe{
\item{drug}{asthmatic relief index for patients given a drug} 
\item{placebo}{asthmatic relief index for patients given a placebo} 
\item{difference}{difference between the \code{placebo} and \code{drug}} 
}}
\usage{
Asthmati
}
\description{
Data for Exercise 7.52
}
\examples{

qqnorm(Asthmati$difference)
qqline(Asthmati$difference)
shapiro.test(Asthmati$difference)
with(data = Asthmati,
     t.test(placebo, drug, paired = TRUE, mu = 0, alternative = "greater")
)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
