% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Autogear}
\alias{Autogear}
\title{Number of defective auto gears produced by two manufacturers}
\format{A data frame with 20 observations on the following 2 variables.
\describe{ 
\item{A}{a numeric vector} 
\item{B}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.46
}
\examples{

str(Autogear)
attach(Autogear)
t.test(A,B)
wilcox.test(A,B)
t.test(A,B,var.equal=TRUE)
detach(Autogear)

}
\keyword{datasets}

