% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Backtoback}
\alias{Backtoback}
\title{Illustrates inferences based on pooled t-test versus Wilcoxon rank sum test}
\format{A data frame with 24 observations on the following 3 variables.
\describe{ 
\item{score}{a numeric vector} 
\item{group}{a numeric vector} 
\item{ranks}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.40
}
\examples{

str(Backtoback)
attach(Backtoback)
wilcox.test(score~group)
detach(Backtoback)

}
\keyword{datasets}

