% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Backtoback}
\alias{Backtoback}
\title{Illustrates inferences based on pooled t-test versus Wilcoxon rank sum test}
\format{A data frame with 24 observations on the following 3 variables.
\describe{ 
\item{score}{a numeric vector} 
\item{group}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\usage{
Backtoback
}
\description{
Data for Exercise 7.40
}
\examples{

wilcox.test(score ~ group, data = Backtoback)
t.test(score ~ group, data = Backtoback)

}
\keyword{datasets}

