% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Bbsalaries}
\alias{Bbsalaries}
\title{Baseball salaries for members of five major league teams}
\format{A data frame with 33 observations on the following 5 variables.
\describe{
\item{ANGLES}{a numeric vector}
\item{ORIOLES}{a numeric vector}
\item{REDSOXS}{a numeric vector}
\item{WHITESOXS}{a numeric vector}
\item{INDIANS}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 1.11
}
\examples{
str(Bbsalaries)
attach(Bbsalaries)
stripchart(x=list(INDIANS,WHITESOXS,REDSOXS,ORIOLES,ANGLES),xlab="Salary",
method="stack",main="",pch=1,col="blue", group.names=c("Indians","White Sox",
"Red Sox", "Orioles","Angels"))
title(main="Major League Salaries")
detach(Bbsalaries)
}
\keyword{datasets}

