% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Bbsalaries}
\alias{Bbsalaries}
\title{Baseball salaries for members of five major league teams}
\format{A data frame with 142 observations on 2 variables.
\describe{ 
\item{salary}{1999 salary} 
\item{team}{a factor with levels \code{Angels}, \code{Indians}, \code{Orioles}, \code{Redsoxs}, and \code{Whitesoxs}} 
}}
\usage{
Bbsalaries
}
\description{
Data for Exercise 1.11
}
\examples{

stripchart(salary ~ team, data = Bbsalaries, method = "stack", 
           pch = 19, col = "blue", cex = 0.75)
title(main = "Major League Salaries")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

