% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Biology}
\alias{Biology}
\title{Test scores on first exam in biology class}
\format{A data frame with 30 observations on the following variable.
\describe{ 
\item{score}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 1.49
}
\examples{

str(Biology)
attach(Biology)
hist(score,breaks="scott",col="brown",prob=TRUE,main="Problem 1.49")
lines(density(score),lwd=3)
detach(Biology)

}
\keyword{datasets}

