% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Biology}
\alias{Biology}
\title{Test scores on first exam in biology class}
\format{A data frame/tibble with 30 observations on one variable
\describe{ 
\item{score}{test scores on the first test in a beginning biology class} 
}}
\usage{
Biology
}
\description{
Data for Exercise 1.49
}
\examples{

hist(Biology$score, breaks = "scott", col = "brown", freq = FALSE, 
main = "Problem 1.49", xlab = "Test Score")
lines(density(Biology$score), lwd=3)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
