% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Blackedu}
\alias{Blackedu}
\title{Education level of blacks by gender}
\format{A data frame with 5 observations on the following 3 variables.
\describe{
\item{education}{a factor with levels \code{bachelor deg}
\code{graduate deg} \code{high sch dropout} \code{high sch graduate}
\code{some college}}
\item{female}{a numeric vector}
\item{male}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 8.55
}
\examples{
str(Blackedu)
attach(Blackedu)
Blackedu
chisq.test(Blackedu[,2:3])
detach(Blackedu)
}
\keyword{datasets}

