% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Blackedu}
\alias{Blackedu}
\title{Education level of blacks by gender}
<<<<<<< HEAD
\format{A data frame with 5 observations on the following 3 variables.
\describe{ 
\item{education}{a factor with levels \code{bachelor deg}
\code{graduate deg} \code{high sch dropout} \code{high sch graduate}
\code{some college}} 
\item{female}{a numeric vector}
\item{male}{a numeric vector} 
=======
\format{A data frame with 3800 observations on two variables.
\describe{
\item{gender}{a factor with levels \code{Female} and \code{Male}}
\item{education}{a factor with levels \code{High school dropout},
\code{High school graudate}, \code{Some college}, \code{Bachelor}'\code{s degree}, and
\code{Graduate degree}} 
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}}
\usage{
Blackedu
}
\description{
Data for Exercise 8.55
}
\examples{

<<<<<<< HEAD
str(Blackedu)
attach(Blackedu)
Blackedu
chisq.test(Blackedu[,2:3])
detach(Blackedu)

=======
T1 <- xtabs(~gender + education, data = Blackedu)
T1
chisq.test(T1)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

