% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Blackedu}
\alias{Blackedu}
\title{Education level of blacks by gender}
\format{
A data frame/tibble  with 3800 observations on two variables
\describe{
\item{gender}{a factor with levels \code{Female} and \code{Male}}
\item{education}{a factor with levels \code{High school dropout},
\code{High school graudate}, \code{Some college}, \code{Bachelor}'\code{s degree}, and
\code{Graduate degree}} 
}
}
\source{
Bureau of Census data.
}
\usage{
Blackedu
}
\description{
Data for Exercise 8.55
}
\examples{

T1 <- xtabs(~gender + education, data = Blackedu)
T1
chisq.test(T1)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
