% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Blood}
\alias{Blood}
\title{Blood pressure of 15 adult males taken by machine and by an expert}
<<<<<<< HEAD
\format{A data frame with 15 observations on the following 2 variables.
\describe{ 
\item{Machine}{a numeric vector} 
\item{Expert}{a numeric vector} 
=======
\format{A data frame with 15 observations on the following two variables.
\describe{ 
\item{machine}{blood pressure recorded from an automated blood pressure machine} 
\item{expert}{blood pressure recorded by an expert using an at-home device} 
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}}
\usage{
Blood
}
\description{
Data for Exercise 7.84
}
\examples{

<<<<<<< HEAD
str(Blood)
attach(Blood)
DIF <- Machine - Expert
qqnorm(DIF)
qqline(DIF)
shapiro.test(DIF)
t.test(Machine,Expert,paired=TRUE)
detach(Blood)
remove(DIF)

=======
DIFF <- Blood$machine - Blood$expert
shapiro.test(DIFF)
qqnorm(DIFF)
qqline(DIFF)
rm(DIFF)
t.test(Blood$machine, Blood$expert, paired = TRUE)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

