% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Blood}
\alias{Blood}
\title{Blood pressure of 15 adult males taken by machine and by an expert}
\format{A data frame with 15 observations on the following 2 variables.
\describe{
\item{Machine}{a numeric vector}
\item{Expert}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.84
}
\examples{
str(Blood)
attach(Blood)
DIF <- Machine - Expert
qqnorm(DIF)
qqline(DIF)
shapiro.test(DIF)
t.test(Machine,Expert,paired=TRUE)
detach(Blood)
remove(DIF)
}
\keyword{datasets}

