% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Blood}
\alias{Blood}
\title{Blood pressure of 15 adult males taken by machine and by an expert}
\format{A data frame with 15 observations on the following 2 variables.
\describe{ 
\item{machine}{blood pressure recorded from an automated blood pressure machine} 
\item{expert}{blood pressure recorded by an expert using an at-home device} 
}}
\usage{
Blood
}
\description{
Data for Exercise 7.84
}
\examples{

DIFF <- Blood$machine - Blood$expert
shapiro.test(DIFF)
qqnorm(DIFF)
qqline(DIFF)
rm(DIFF)
t.test(Blood$machine, Blood$expert, paired = TRUE)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

