% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Board}
\alias{Board}
\title{Incomes of board members from three different universities}
\format{A data frame with 7 observations on the following 3 variables.
\describe{ 
\item{UnivA}{a numeric vector} 
\item{UnivB}{a numeric vector} 
\item{UnivC}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 10.14
}
\examples{

str(Board)
attach(Board)
STACKED <-stack(Board)
STACKED[1:5,]
boxplot(values~ind,col=c("red","blue","green"),data=STACKED)
remove(STACKED)
detach(Board)

}
\keyword{datasets}

