% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Board}
\alias{Board}
\title{Incomes of board members from three different universities}
<<<<<<< HEAD
\format{A data frame with 7 observations on the following 3 variables.
\describe{ 
\item{UnivA}{a numeric vector} 
\item{UnivB}{a numeric vector} 
\item{UnivC}{a numeric vector} 
=======
\format{A data frame with 7 observations on the following three variables.
\describe{ 
\item{salary}{1999 salary (in $1000) for board directors} 
\item{university}{a factor with levels \code{A}, \code{B}, and \code{C}} 
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}}
\usage{
Board
}
\description{
Data for Exercise 10.14
}
\examples{

<<<<<<< HEAD
str(Board)
attach(Board)
STACKED <-stack(Board)
STACKED[1:5,]
boxplot(values~ind,col=c("red","blue","green"),data=STACKED)
remove(STACKED)
detach(Board)

=======
boxplot(salary ~ university, data = Board, col = c("red", "blue", "green"), 
        ylab = "Income")
tapply(Board$salary, Board$university, summary)
anova(lm(salary ~ university, data = Board))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

