% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Board}
\alias{Board}
\title{Incomes of board members from three different universities}
\format{A data frame with 7 observations on the following 3 variables.
\describe{ 
\item{salary}{1999 salary (in $1000) for board directors} 
\item{university}{a factor with levels \code{A}, \code{B}, and \code{C}} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\usage{
Board
}
\description{
Data for Exercise 10.14
}
\examples{

boxplot(salary ~ university, data = Board, col = c("red", "blue", "green"), 
        ylab = "Income")
tapply(Board$salary, Board$university, summary)
anova(lm(salary ~ university, data = Board))

}
\keyword{datasets}

