% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Bones}
\alias{Bones}
\title{Bone density measurements of 35 physically active and 35 non-active women}
\format{A data frame with 70 observations on the following 5 variables.
\describe{ 
\item{Active}{a numeric vector}
\item{Nonactive}{a numeric vector} 
\item{Density}{a numeric vector} 
\item{group}{a numeric vector} 
\item{Ranks}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.22
}
\examples{

str(Bones)
attach(Bones)
t.test(Active,Nonactive,alternative="greater")
wilcox.test(Active,Nonactive,alternative="greater")
detach(Bones)

}
\keyword{datasets}

