% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Bones}
\alias{Bones}
\title{Bone density measurements of 35 physically active and 35 non-active women}
\format{A data frame with 70 observations on 2 variables.
\describe{ 
\item{density}{bone density measurements}
\item{group}{a factor with levels \code{active} and \code{nonactive}} 
}}
\usage{
Bones
}
\description{
Data for Example 7.22
}
\examples{

t.test(density ~ group, data = Bones, alternative = "greater")
t.test(rank(density) ~ group, data = Bones, alternative = "greater")
wilcox.test(density ~ group, data = Bones, alternative = "greater")


}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

