% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Books}
\alias{Books}
\title{Number of books read and final spelling scores for 17 third graders}
\format{A data frame with 17 observations on the following 2 variables.
\describe{
\item{book}{a numeric vector}
\item{spelling}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 9.53
}
\examples{
data(Books)
}
\keyword{datasets}

