% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Books}
\alias{Books}
\title{Number of books read and final spelling scores for 17 third graders}
\format{A data frame with 17 observations on the following 2 variables.
\describe{ 
\item{book}{number of books read} 
\item{spelling}{spelling score} 
}}
\usage{
Books
}
\description{
Data for Exercise 9.53
}
\examples{

plot(spelling ~ book, data = Books)
mod <- lm(spelling ~ book, data = Books)
summary(mod)
abline(mod, col = "blue", lwd = 2)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

