% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Bookstor}
\alias{Bookstor}
\title{Prices paid for used books at three different bookstores}
\format{
A data frame/tibble with 72 observations on two variables
\describe{ 
\item{dollars}{money obtained for selling textbooks} 
\item{store}{a factor with levels \code{A}, \code{B}, and \code{C}} 
}
}
\usage{
Bookstor
}
\description{
Data for Exercise 10.30 and 10.31
}
\examples{

boxplot(dollars ~ store, data = Bookstor, 
        col = c("purple", "lightblue", "cyan"))
kruskal.test(dollars ~ store, data = Bookstor)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
