% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Bookstor}
\alias{Bookstor}
\title{Prices paid for used books at three different bookstores}
\format{A data frame with 72 observations on the following 6 variables.
\describe{ 
\item{StoreA}{a numeric vector} 
\item{StoreB}{a numeric vector} 
\item{StoreC}{a numeric vector}
\item{Dollars}{a numeric vector} 
\item{Store}{a numeric vector} 
\item{Ranks}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 10.30 and 10.31
}
\examples{

str(Bookstor)
attach(Bookstor)
boxplot(Dollars~Store)
kruskal.test(Dollars~as.factor(Store))
detach(Bookstor)

}
\keyword{datasets}

