% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{package}
\name{BSDA-package}
\alias{BSDA}
\alias{BSDA-package}
\title{Basic Statistics and Data Analysis}
\description{
Data and functions for the book \emph{Basic Statistics and Data Analysis}
}
\details{
\tabular{ll}{ Package: \tab PASWR\cr Type: \tab Package\cr Version: \tab
1.02\cr Date: \tab 2016-02-25\cr License: \tab GPL (>=2) \cr } The package
BSDA provides data and functions for the book \emph{Basic Statistics and
Data Analysis}
}
\author{
Alan T. Arnholt

Maintainer: <arnholtat@appstate.edu>
}
\references{
Kitchens, L. J (2003) \emph{Basic Statistics and Data Analysis}.
Brooks/Cole.
}
\keyword{package}

