% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Bus}
\alias{Bus}
\title{Attendance of bus drivers versus attendance}
\format{A data frame with 29363 observations on two variables.
\describe{ 
\item{attendance}{a factor with levels \code{absent} and
\code{present}} 
\item{shift}{a factor with levels \code{am}, \code{noon}, \code{pm}, \code{swing}, and \code{split}} 
}}
\usage{
Bus
}
\description{
Data for Exercise 8.25
}
\examples{

T1 <- xtabs(~attendance + shift, data = Bus)
T1
chisq.test(T1)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
