% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Bus}
\alias{Bus}
\title{Attendance of bus drivers versus attendance}
\format{A data frame with 2 observations on the following 6 variables.
\describe{ 
\item{Attend}{a factor with levels \code{Absent}
\code{Present}} 
\item{AM}{a numeric vector} 
\item{Noon}{a numeric vector} 
\item{PM}{a numeric vector} 
\item{Swing}{a numeric vector} 
\item{Split}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 8.25
}
\examples{

str(Bus)
attach(Bus)
Bus
chisq.test(Bus[,2:6])
detach(Bus)

}
\keyword{datasets}

