% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Bypass}
\alias{Bypass}
\title{Median charges for coronary bypass at 17 hospitals in North Carolina}
\format{A data frame with 17 observations on the following 2 variables.
\describe{ 
\item{hospital}{a factor with levels \code{Carolinas Med
Ct} \code{Duke Med Ct} \code{Durham Regional} \code{Forsyth Memorial}
\code{Frye Regional} \code{High Point Regional} \code{Memorial Mission}
\code{Mercy} \code{Moore Regional} \code{Moses Cone Memorial} \code{NC
Baptist} \code{New Hanover Regional} \code{Pitt Co. Memorial}
\code{Presbyterian} \code{Rex} \code{Univ of North Carolina} \code{Wake
County}}
\item{charge}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercises 5.104 and 6.43
}
\examples{

str(Bypass)
attach(Bypass)
EDA(charge)
t.test(charge,conf.level=.90)$conf
t.test(charge,mu=35000)
detach(Bypass)

}
\keyword{datasets}

