% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Cabinets}
\alias{Cabinets}
\title{Estimates of costs of kitchen cabinets by two suppliers on 20 prospective
homes}
\format{A data frame with 20 observations on the following 3 variables.
\describe{ 
\item{Home}{a numeric vector} 
\item{SupplA}{estimate for kitchen cabinets from supplier A (in dollars)} 
\item{SupplB}{estimate for kitchen cabinets from supplier A (in dollars)} 
}}
\usage{
Cabinets
}
\description{
Data for Exercise 7.83
}
\examples{

DIF <- Cabinets$SupplA - Cabinets$SupplB
qqnorm(DIF)
qqline(DIF)
shapiro.test(DIF)
with(data = Cabinets, 
     t.test(SupplA, SupplB, paired = TRUE)
)
with(data = Cabinets,
     wilcox.test(SupplA, SupplB, paired = TRUE)
)
rm(DIF)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

