% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Cabinets}
\alias{Cabinets}
\title{Estimates of costs of kitchen cabinets by two suppliers on 20 prospective
homes}
\format{A data frame with 20 observations on the following 3 variables.
\describe{ 
\item{Home}{a numeric vector} 
\item{SupplA}{a numeric vector} 
\item{SupplB}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.83
}
\examples{

str(Cabinets)
attach(Cabinets)
DIF <- SupplA - SupplB
qqnorm(DIF)
qqline(DIF)
shapiro.test(DIF)
t.test(SupplA,SupplB,paired=TRUE)
wilcox.test(SupplA,SupplB,paired=TRUE)
detach(Cabinets)
remove(DIF)

}
\keyword{datasets}

