% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Cancer}
\alias{Cancer}
\title{Survival times of terminal cancer patients treated with vitamin C}
\format{A data frame with 17 observations on the following 5 variables.
\describe{
\item{stomach}{a numeric vector}
\item{bronchus}{a numeric vector}
\item{colon}{a numeric vector}
\item{ovary}{a numeric vector}
\item{breast}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercises 6.55 and 6.64
}
\examples{
str(Cancer)
attach(Cancer)
EDA(stomach)
SIGN.test(stomach,md=100,alternative="greater")
detach(Cancer)
}
\keyword{datasets}

