% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Carbon}
\alias{Carbon}
\title{Carbon monoxide level measured at three industrial sites}
<<<<<<< HEAD
\format{A data frame with 24 observations on the following 6 variables.
\describe{ 
\item{SiteA}{a numeric vector} 
\item{SiteB}{a numeric vector} 
\item{SiteC}{a numeric vector}
\item{monoxide}{a numeric vector} 
\item{Site}{a numeric vector} 
\item{Ranks}{a numeric vector}
=======
\format{A data frame with 24 observations on two variables.
\describe{ 
\item{CO}{carbon monoxide measured (in parts per million)} 
\item{site}{a factor with levels \code{SiteA}, \code{SiteB}, and \code{SiteC}} 
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}}
\usage{
Carbon
}
\description{
Data for Exercise 10.28 and 10.29
}
\examples{

str(Carbon)
<<<<<<< HEAD
attach(Carbon)
boxplot(monoxide~Site)
kruskal.test(monoxide~as.factor(Site))
detach(Carbon)

=======
boxplot(CO ~ site, data = Carbon)
kruskal.test(CO ~ site, data = Carbon)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

