% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Carbon}
\alias{Carbon}
\title{Carbon monoxide level measured at three industrial sites}
\format{A data frame with 24 observations on the following 6 variables.
\describe{ 
\item{SiteA}{a numeric vector} 
\item{SiteB}{a numeric vector} 
\item{SiteC}{a numeric vector}
\item{monoxide}{a numeric vector} 
\item{Site}{a numeric vector} 
\item{Ranks}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 10.28 and 10.29
}
\examples{

str(Carbon)
attach(Carbon)
boxplot(monoxide~Site)
kruskal.test(monoxide~as.factor(Site))
detach(Carbon)

}
\keyword{datasets}

