% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Carbon}
\alias{Carbon}
\title{Carbon monoxide level measured at three industrial sites}
\format{A data frame with 24 observations on two variables.
\describe{ 
\item{CO}{carbon monoxide measured (in parts per million)} 
\item{site}{a factor with levels \code{SiteA}, \code{SiteB}, and \code{SiteC}} 
}}
\usage{
Carbon
}
\description{
Data for Exercise 10.28 and 10.29
}
\examples{

str(Carbon)
boxplot(CO ~ site, data = Carbon)
kruskal.test(CO ~ site, data = Carbon)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
