% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Cat}
\alias{Cat}
\title{Reading scores on the California achievement test for a group of 3rd graders}
\format{A data frame with 17 observations on the following variable.
\describe{ 
\item{score}{reading score on the California Achievement Test} 
}}
\usage{
Cat
}
\description{
Data for Exercise 1.116
}
\examples{

str(Cat)
stem(Cat$score)
fivenum(Cat$score)
boxplot(Cat$score, main = "Problem 1.116", col = "green")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
