% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Censored}
\alias{Censored}
\title{Entry age and survival time of patients with small cell lung cancer under
two different treatments}
\format{A data frame with 121 observations on the following 8 variables.
\describe{ 
\item{survivA}{a numeric vector} 
\item{ageA}{a numeric vector} 
\item{censorA}{a numeric vector}
\item{survivB}{a numeric vector} 
\item{ageB}{a numeric vector} 
\item{censorB}{a numeric vector} 
\item{survival}{a numeric vector} 
\item{group}{a numeric vector} 
}}
\description{
Data for Exercises 7.34 and 7.48
}
\examples{

str(Censored)
attach(Censored)
boxplot(survival~group)
detach(Censored)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

