% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Chemist}
\alias{Chemist}
\title{Starting salaries of 50 chemistry majors}
\format{A data frame/tibble with 50 observations on one variable
\describe{ 
\item{salary}{starting salary (in dollars) for chemistry major} 
}}
\usage{
Chemist
}
\description{
Data for Example 5.3
}
\examples{

EDA(Chemist$salary)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
