% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Chemist}
\alias{Chemist}
\title{Starting salaries of 50 chemistry majors}
\format{A data frame with 50 observations on the following variable.
\describe{ 
\item{salary}{starting salary (in dollars) for chemistry major} 
}}
\usage{
Chemist
}
\description{
Data for Example 5.3
}
\examples{

str(Chemist)
EDA(Chemist$salary)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

