% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Chesapea}
\alias{Chesapea}
\title{Surface salinity measurements taken offshore from Annapolis, Maryland in
1927}
\format{A data frame with 16 observations on the following variable.
\describe{ 
<<<<<<< HEAD
\item{salinity}{a numeric vector} 
=======
\item{salinity}{surface salinity measurements (in parts per 1000) for station 11, 
offshore from Annanapolis, Maryland, on July 3-4, 1927.} 
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}}
\source{
Davis, J. (1986) \emph{Statistics and Data Analysis in Geology, Second Edition}. 
John Wiley and Sons, New York.
}
\usage{
Chesapea
}
\description{
Data for Exercise 6.41
}
\examples{

str(Chesapea)
<<<<<<< HEAD
attach(Chesapea)
qqnorm(salinity)
qqline(salinity)
shapiro.test(salinity)
t.test(salinity,mu=7)
detach(Chesapea)

=======
qqnorm(Chesapea$salinity)
qqline(Chesapea$salinity)
shapiro.test(Chesapea$salinity)
t.test(Chesapea$salinity, mu = 7)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

