% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Chesapea}
\alias{Chesapea}
\title{Surface salinity measurements taken offshore from Annapolis, Maryland in
1927}
\format{
A data frame/tibble with 16 observations on one variable
\describe{ 
\item{salinity}{surface salinity measurements (in parts per 1000) for station 11, 
offshore from Annanapolis, Maryland, on July 3-4, 1927.} 
}
}
\source{
Davis, J. (1986) \emph{Statistics and Data Analysis in Geology, Second Edition}. 
John Wiley and Sons, New York.
}
\usage{
Chesapea
}
\description{
Data for Exercise 6.41
}
\examples{

qqnorm(Chesapea$salinity)
qqline(Chesapea$salinity)
shapiro.test(Chesapea$salinity)
t.test(Chesapea$salinity, mu = 7)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
