% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Chesapea}
\alias{Chesapea}
\title{Surface salinity measurements taken offshore from Annapolis, Maryland in
1927}
\format{A data frame with 16 observations on the following variable.
\describe{ 
\item{salinity}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 6.46
}
\examples{

str(Chesapea)
attach(Chesapea)
qqnorm(salinity)
qqline(salinity)
shapiro.test(salinity)
t.test(salinity,mu=7)
detach(Chesapea)

}
\keyword{datasets}

