% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Chevy}
\alias{Chevy}
\title{Insurance injury ratings of Chevrolet vehicles for 1990 and 1993 models}
\format{A data frame with 2 observations on the following 6 variables.
\describe{
\item{year}{a factor with levels \code{88-90}
\code{91-93}}
\item{A}{a numeric vector}
\item{B}{a numeric vector}
\item{C}{a numeric vector}
\item{D}{a numeric vector}
\item{F}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 8.35
}
\examples{
str(Chevy)
attach(Chevy)
Chevy
chisq.test(Chevy[,2:6])
detach(Chevy)
}
\keyword{datasets}

