% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Chicken}
\alias{Chicken}
\title{Weight gain of chickens fed three different rations}
\format{A data frame with 13 observations on the following 3 variables.
\describe{
\item{Ration1}{a numeric vector}
\item{Ration2}{a numeric vector}
\item{Ration3}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 10.15
}
\examples{
str(Chicken)
attach(Chicken)
STACKED <-stack(Chicken)
STACKED[1:5,]
boxplot(values~ind,col=c("red","blue","green"),data=STACKED)
anova(lm(values~ind,data=STACKED))
remove(STACKED)
detach(Chicken)
}
\keyword{datasets}

