% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Chicken}
\alias{Chicken}
\title{Weight gain of chickens fed three different rations}
\format{A data frame with 13 observations on the following 3 variables.
\describe{ 
\item{gain}{weight gain over a specified period} 
\item{feed}{a factor with levels \code{ration1}, \code{ration2}, 
and \code{ration3}} 
}}
\usage{
Chicken
}
\description{
Data for Exercise 10.15
}
\examples{

str(Chicken)
boxplot(gain ~ feed, col = c("red","blue","green"), data = Chicken)
anova(lm(gain ~ feed, data = Chicken))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

