% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Chicken}
\alias{Chicken}
\title{Weight gain of chickens fed three different rations}
\format{A data frame/tibble with 13 observations onthree variables
\describe{ 
\item{gain}{weight gain over a specified period} 
\item{feed}{a factor with levels \code{ration1}, \code{ration2}, 
and \code{ration3}} 
}}
\usage{
Chicken
}
\description{
Data for Exercise 10.15
}
\examples{

boxplot(gain ~ feed, col = c("red","blue","green"), data = Chicken)
anova(lm(gain ~ feed, data = Chicken))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
