% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Chipavg}
\alias{Chipavg}
\title{Measurements of the thickness of the oxide layer of manufactured integrated
circuits}
\format{A data frame with 30 observations on the following 3 variables.
\describe{ 
\item{wafer1}{a numeric vector} 
\item{wafer2}{a numeric vector} 
\item{thickness}{a numeric vector} 
}}
\description{
Data for Exercises 6.49 and 7.47
}
\examples{

str(Chipavg)
attach(Chipavg)
EDA(thickness)
t.test(thickness,mu=1000)
boxplot(wafer1,wafer2,name=c("Wafer 1","Wafer 2"))
shapiro.test(wafer1)
shapiro.test(wafer2)
t.test(wafer1,wafer2,var.equal=TRUE)
detach(Chipavg)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

