% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Chips}
\alias{Chips}
\title{Four measurements on a first wafer and four measurements on a second wafer
selected from 30 lots}
\format{A data frame with 30 observations on the following 8 variables.
\describe{ 
\item{wafer11}{a numeric vector} 
\item{wafer12}{a numeric vector} 
\item{wafer13}{a numeric vector}
\item{wafer14}{a numeric vector} 
\item{wafer21}{a numeric vector} 
\item{wafer22}{a numeric vector} 
\item{wafer23}{a numeric vector} 
\item{wafer24}{a numeric vector} 
}}
\description{
Data for Exercise 10.9
}
\examples{

str(Chips)
attach(Chips)
boxplot(wafer11,wafer12,wafer13,wafer14,wafer21,wafer22,wafer23,wafer24)
detach(Chips)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

