% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Cigar}
\alias{Cigar}
\title{Milligrams of tar in 25 cigarettes selected randomly from 4 different brands}
\format{A data frame with 100 observations on the following 6 variables.
\describe{
\item{brandA}{a numeric vector}
\item{brandB}{a numeric vector}
\item{brandC}{a numeric vector}
\item{brandD}{a numeric vector}
\item{tar}{a numeric vector}
\item{brand}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Example 10.4
}
\examples{
str(Cigar)
attach(Cigar)
boxplot(tar~brand)
anova(lm(tar~as.factor(brand)))
detach(Cigar)
}
\keyword{datasets}

