% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Cigarett}
\alias{Cigarett}
\title{Effect of mother's smoking on birth weight of newborn}
\format{A data frame with 16 observations on the following 2 variables.
\describe{
\item{cigarett}{a numeric vector}
\item{weight}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 2.27
}
\examples{
str(Cigarett)
attach(Cigarett)
plot(cigarett,weight)
model <- lm(weight~cigarett)
abline(model)
cor(weight,cigarett)
detach(Cigarett)
}
\keyword{datasets}

