% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Cigarett}
\alias{Cigarett}
\title{Effect of mother's smoking on birth weight of newborn}
\format{A data frame/tibble  with 16 observations on the following two variables.
\describe{ 
\item{cigarettes}{mothers' estimated average number of cigarettes smoked per day} 
\item{weight}{children's birth weights (in pounds)} 
}}
\usage{
Cigarett
}
\description{
Data for Exercise 2.27
}
\examples{

plot(weight ~ cigarettes, data = Cigarett)
model <- lm(weight ~ cigarettes, data = Cigarett)
abline(model)
with(data = Cigarett,
cor(weight, cigarettes)
)
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
