% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Citrus}
\alias{Citrus}
\title{Percent of peak bone density of different aged children}
\format{A data frame with 9 observations on the following 2 variables.
\describe{ 
\item{age}{a numeric vector} 
\item{percent}{a numeric vector} 
}}
\description{
Data for Exercise 9.7
}
\examples{

str(Citrus)
attach(Citrus)
model <- lm(percent~age)
summary(model)
anova(model)
detach(Citrus)
remove(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

