% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Citrus}
\alias{Citrus}
\title{Percent of peak bone density of different aged children}
\format{A data frame with 9 observations on the following 2 variables.
\describe{ 
\item{age}{age of children} 
\item{percent}{percent peak bone density} 
}}
\usage{
Citrus
}
\description{
Data for Exercise 9.7
}
\examples{

str(Citrus)
model <- lm(percent ~ age, data = Citrus)
summary(model)
anova(model)
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

