% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Clean}
\alias{Clean}
\title{Residual contaminant following the use of three different cleansing agents}
<<<<<<< HEAD
\format{A data frame with 45 observations on the following 5 variables.
\describe{ 
\item{A}{a numeric vector} 
\item{B}{a numeric vector} 
\item{C}{a numeric vector} 
\item{clean}{a numeric vector} 
\item{agent}{a numeric vector} 
=======
\format{A data frame with 45 observations on the following two variables.
\describe{ 
\item{clean}{residual contaminants} 
\item{agent}{a factor with levels \code{A}, \code{B}, and \code{C}} 
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}}
\usage{
Clean
}
\description{
Data for Exercise 10.16
}
\examples{

str(Clean)
<<<<<<< HEAD
attach(Clean)
boxplot(clean~agent,col=c("red","blue","green"))
anova(lm(clean~as.factor(agent)))
detach(Clean)

=======
boxplot(clean ~ agent, col = c("red", "blue", "green"), data = Clean)
anova(lm(clean ~ agent, data = Clean))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

