% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Clean}
\alias{Clean}
\title{Residual contaminant following the use of three different cleansing agents}
\format{A data frame with 45 observations on the following 5 variables.
\describe{ 
\item{A}{a numeric vector} 
\item{B}{a numeric vector} 
\item{C}{a numeric vector} 
\item{clean}{a numeric vector} 
\item{agent}{a numeric vector} 
}}
\description{
Data for Exercise 10.16
}
\examples{

str(Clean)
attach(Clean)
boxplot(clean~agent,col=c("red","blue","green"))
anova(lm(clean~as.factor(agent)))
detach(Clean)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

