% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Coaxial}
\alias{Coaxial}
\title{Signal loss from three types of coxial cable}
\format{A data frame with 45 observations on the following 5 variables.
\describe{
\item{Type.A}{a numeric vector}
\item{Type.B}{a numeric vector}
\item{Type.C}{a numeric vector}
\item{Signal}{a numeric vector}
\item{Cable}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 10.24 and 10.25
}
\examples{
str(Coaxial)
attach(Coaxial)
boxplot(Signal~Cable)
kruskal.test(Signal~as.factor(Cable))
detach(Coaxial)
}
\keyword{datasets}

