% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Coaxial}
\alias{Coaxial}
\title{Signal loss from three types of coxial cable}
\format{A data frame with 45 observations on the following 2 variables.
\describe{ 
\item{signal}{signal loss per 1000 feet} 
\item{cable}{factor with three levels of coaxial cable \code{typeA}, 
\code{typeB}, and \code{typeC}} 
}}
\usage{
Coaxial
}
\description{
Data for Exercise 10.24 and 10.25
}
\examples{

str(Coaxial)
boxplot(signal ~ cable, data = Coaxial, col = c("red", "green", "yellow"))
kruskal.test(signal ~ cable, data = Coaxial)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

