% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Coffee}
\alias{Coffee}
\title{Productivity of workers with and without a coffee break}
\format{A data frame/tibble with nine observations on three variables
\describe{ 
\item{without}{workers' productivity scores without a coffee break} 
\item{with}{workers' productivity scores with a coffee break}
\item{differences}{\code{with} minus \code{without}} 
}}
\usage{
Coffee
}
\description{
Data for Exercise 7.55
}
\examples{

qqnorm(Coffee$differences)
qqline(Coffee$differences)
shapiro.test(Coffee$differences)
t.test(Coffee$with, Coffee$without, paired = TRUE, alternative = "greater")
wilcox.test(Coffee$with, Coffee$without, paired = TRUE, 
alterantive = "greater")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
