% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Coffee}
\alias{Coffee}
\title{Productivity of workers with and without a coffee break}
\format{A data frame with 9 observations on the following 2 variables.
\describe{ 
\item{without}{workers' productivity scores without a coffee break} 
\item{with}{workers' productivity scores with a coffee break} 
}}
\usage{
Coffee
}
\description{
Data for Exercise 7.55
}
\examples{

str(Coffee)
differences <- Coffee$with - Coffee$without
qqnorm(differences)
qqline(differences)
shapiro.test(differences)
t.test(Coffee$with, Coffee$without, paired = TRUE, alternative = "greater")
wilcox.test(Coffee$with, Coffee$without, paired = TRUE, 
            alterantive = "greater")
rm(differences)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

