% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Coffee}
\alias{Coffee}
\title{Productivity of workers with and without a coffee break}
\format{A data frame with 9 observations on the following 4 variables.
\describe{ 
\item{Without}{a numeric vector} 
\item{With}{a numeric vector} 
\item{differ}{a numeric vector}
\item{sgnrnks}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.55
}
\examples{

str(Coffee)
attach(Coffee)
qqnorm(differ)
qqline(differ)
shapiro.test(differ)
t.test(With,Without,paired=TRUE,alternative="greater")
wilcox.test(With,Without,paired=TRUE,alternative="greater")
detach(Coffee)

}
\keyword{datasets}

