% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Coins}
\alias{Coins}
\title{Yearly returns on 12 investments}
\format{A data frame with 12 observations on the following variable.
\describe{ 
\item{coins}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 5.68
}
\examples{

str(Coins)
attach(Coins)
qqnorm(coins)
qqline(coins)
EDA(coins)
detach(Coins)

}
\keyword{datasets}

