% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Coins}
\alias{Coins}
\title{Yearly returns on 12 investments}
\format{A data frame/tibble with 12 observations on one variable
\describe{ 
\item{return}{yearly returns on each of 12 possible investments} 
}}
\usage{
Coins
}
\description{
Data for Exercise 5.68
}
\examples{

qqnorm(Coins$return)
qqline(Coins$return)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
