% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Commute}
\alias{Commute}
\title{Commuting times for selected cities in 1980 and 1990}
\format{A data frame with 39 observations on the following three variables.
\describe{ 
\item{city}{a factor with levels \code{Atlanta},
\code{Baltimore}, \code{Boston}, \code{Buffalo}, \code{Charlotte},
\code{Chicago}, \code{Cincinnati}, \code{Cleveland}, \code{Columbus},
\code{Dallas}, \code{Denver}, \code{Detroit}, \code{Hartford}, \code{Houston},
\code{Indianapolis}, \code{Kansas City}, \code{Los Angeles}, \code{Miami},
\code{Milwaukee}, \code{Minneapolis}, \code{New Orleans}, \code{New York},
\code{Norfolk}, \code{Orlando}, \code{Philadelphia}, \code{Phoenix},
\code{Pittsburgh}, \code{Portland}, \code{Providence}, \code{Rochester},
\code{Sacramento}, \code{Salt Lake City}, \code{San Antonio}, \code{San Diego},
\code{San Francisco}, \code{Seattle}, \code{St. Louis}, \code{Tampa}, and
\code{Washington}}
\item{year1980}{urban commute times for the year 1980}
\item{year1990}{urban commute times for the year 1990}
}}
\usage{
Commute
}
\description{
Data for Exercises 1.13, and 7.85
}
\examples{

commute <- stack(Commute)
str(commute)
stripplot(ind ~ values, data = commute, jitter = TRUE)
dotplot(ind ~ values, data = commute)
bwplot(ind ~ values, data = commute)
stripchart(values ~ ind, data = commute, method = "stack", pch = 1, cex = 2, 
          col = c("red", "blue"), group.names = c("1980", "1990"), main = "",
          xlab = "minutes")
title(main = "Commute Time")
boxplot(values ~ ind, data = commute, names=c("1980", "1990"), 
       horizontal = TRUE, las = 1)
rm(commute)


}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
