% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Concept}
\alias{Concept}
\title{Tennessee self concept scale scores for a group of teenage boys}
\format{A data frame/tibble  with 28 observations on the following variable.
\describe{ 
\item{self}{Tennessee self concept scores} 
}}
\usage{
Concept
}
\description{
Data for Exercise 16.8 and 1.82
}
\examples{

summary(Concept$self)
sd(Concept$self)
diff(range(Concept$self))
IQR(Concept$self)
summary(Concept$self/10)
IQR(Concept$self/10)
sd(Concept$self/10)
diff(range(Concept$self/10))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
