% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Concrete}
\alias{Concrete}
\title{Compressive strength of concrete blocks made by two different methods}
\format{A data frame with 20 observations on the following 3 variables.
\describe{
\item{Strength}{a numeric vector}
\item{Method}{a numeric vector}
\item{Ranks}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.17
}
\examples{
str(Concrete)
attach(Concrete)
wilcox.test(Strength~Method,alternative="greater")
detach(Concrete)
}
\keyword{datasets}

