% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Concrete}
\alias{Concrete}
\title{Compressive strength of concrete blocks made by two different methods}
\format{A data frame/tibble  with 20 observations on the following 2 variables.
\describe{ 
\item{strength}{comprehensive strength (in pounds per square inch)} 
\item{method}{factor with levels \code{new} and \code{old} indicating the 
method used to construct a concrete block} 
}}
\usage{
Concrete
}
\description{
Data for Example 7.17
}
\examples{

wilcox.test(strength ~ method, data = Concrete, alternative = "greater")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
