% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Corn}
\alias{Corn}
\title{Comparison of the yields of a new variety and a standard variety of corn
planted on 12 plots of land}
\format{A data frame with 12 observations on the following 3 variables.
\describe{ 
\item{plot}{number of plot} 
\item{new}{corn yield with new meathod} 
\item{standard}{corn yield with standard method} 
}}
\usage{
Corn
}
\description{
Data for Exercise 7.77
}
\examples{

str(Corn)
differ <- Corn$new - Corn$standard
boxplot(differ)
qqnorm(differ)
qqline(differ)
shapiro.test(differ)
t.test(Corn$new, Corn$standard, paired = TRUE, alternative = "greater")
rm(differ)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

