% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Corn}
\alias{Corn}
\title{Comparison of the yields of a new variety and a standard variety of corn
planted on 12 plots of land}
\format{A data frame with 12 observations on the following 3 variables.
\describe{
\item{New}{a numeric vector}
\item{Standard}{a numeric vector}
\item{differ}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.77
}
\examples{
str(Corn)
attach(Corn)
boxplot(differ)
qqnorm(differ)
qqline(differ)
shapiro.test(differ)
t.test(New,Standard,paired=TRUE,alternative="greater")
detach(Corn)
}
\keyword{datasets}

