% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Corn}
\alias{Corn}
\title{Comparison of the yields of a new variety and a standard variety of corn
planted on 12 plots of land}
\format{A data frame/tibble with 12 observations on three variables
\describe{ 
\item{new}{corn yield with new meathod} 
\item{standard}{corn yield with standard method}
\item{differences}{\code{new} minus \code{standard}} 
}}
\usage{
Corn
}
\description{
Data for Exercise 7.77
}
\examples{

boxplot(Corn$differences)
qqnorm(Corn$differences)
qqline(Corn$differences)
shapiro.test(Corn$differences)
t.test(Corn$new, Corn$standard, paired = TRUE, alternative = "greater")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
