% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Correlat}
\alias{Correlat}
\title{Exercise to illustrate correlation}
\format{A data frame with 13 observations on the following 2 variables.
\describe{
\item{X}{a numeric vector}
\item{Y}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 2.23
}
\examples{
str(Correlat)
attach(Correlat)
plot(X,Y)
model <- lm(Y~X)
abline(model)
detach(Correlat)
}
\keyword{datasets}

