% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Correlat}
\alias{Correlat}
\title{Exercise to illustrate correlation}
\format{
A data frame/tibble with 13 observations on two variables
\describe{ 
\item{x}{a numeric vector} 
\item{y}{a numeric vector} 
}
}
\usage{
Correlat
}
\description{
Data for Exercise 2.23
}
\examples{

plot(y ~ x, data = Correlat)
model <- lm(y ~ x, data = Correlat)
abline(model)
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
