% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Counsel}
\alias{Counsel}
\title{Scores of 18 volunteers who participated in a counseling process}
\format{A data frame with 18 observations on the following variable.
\describe{
\item{score}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 6.96
}
\examples{
str(Counsel)
attach(Counsel)
EDA(score)
t.test(score,mu=70)
detach(Counsel)
}
\keyword{datasets}

