% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Cpi}
\alias{Cpi}
\title{Consumer price index from 1979 to 1998}
\format{A data frame with 20 observations on the following 2 variables.
\describe{
\item{Year}{a numeric vector}
\item{CPI}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 1.34
}
\examples{
str(Cpi)
attach(Cpi)
plot(Year,CPI,type="l",lty=2,lwd=2,col="red")
names(CPI) <- Year
barplot(CPI,col="pink",las=2,main="Problem 1.34")
detach(Cpi)
}
\keyword{datasets}

