% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Darwin}
\alias{Darwin}
\title{Charles Darwin's study of cross-fertilized and self-fertilized plants}
\format{A data frame with 30 observations on the following 7 variables.
\describe{ 
\item{pot}{a numeric vector} 
\item{cross}{a numeric vector} 
\item{self}{a numeric vector}
\item{height}{a numeric vector} 
\item{method}{a numeric vector} 
\item{differ}{a numeric vector} 
\item{sgnrnks}{a numeric vector} 
}}
\description{
Data for Exercise 7.62
}
\examples{

str(Darwin)
attach(Darwin)
qqnorm(differ)
qqline(differ)
shapiro.test(differ)
wilcox.test(cross,self,paired=TRUE)
detach(Darwin)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

