% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Dealers}
\alias{Dealers}
\title{Automobile dealers classified according to type dealership and service
rendered to customers}
\format{A data frame with 6 observations on the following 2 variables.
\describe{ 
\item{dealership}{a factor with levels \code{Honda}, \code{Toyota}, 
\code{Mazda}, \code{Ford}, \code{Dodge}, and \code{Saturn}} 
\item{service}{a factor with levels \code{Before Needed}, \code{Only When Recommended}} 
}}
\usage{
Dealers
}
\description{
Data for Example 2.22
}
\examples{

str(Dealers)
T1 <- xtabs(~dealership+service,data=Dealers)
T1
barplot(t(T1), beside = TRUE)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

