% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Dealers}
\alias{Dealers}
\title{Automobile dealers classified according to type dealership and service
rendered to customers}
\format{A data frame/tibble  with 122 observations on the following two variables.
\describe{ 
\item{type}{a factor with levels \code{Honda}, \code{Toyota}, \code{Mazda}, 
\code{Ford}, \code{Dodge}, and \code{Saturn}} 
\item{service}{a factor with levels \code{Replaces unnecessarily} and \code{Follows manufacturer guidelines}} 
}}
\usage{
Dealers
}
\description{
Data for Example 2.22
}
\examples{

xtabs(~type + service, data = Dealers)
T1 <- xtabs(~type + service, data = Dealers)
T1
addmargins(T1)
pt <- prop.table(T1, margin = 1)
pt
barplot(t(pt),  col = c("red", "skyblue"), legend = colnames(T1))
rm(T1, pt)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
