% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Dealers}
\alias{Dealers}
\title{Automobile dealers classified according to type dealership and service
rendered to customers}
\format{A data frame with 6 observations on the following 2 variables.
\describe{ 
\item{Replace}{a numeric vector} 
\item{Recomnd}{a numeric vector} 
}}
\description{
Data for Example 2.22
}
\examples{

str(Dealers)
attach(Dealers)
Deal <- as.matrix(Dealers)
rownames(Deal) <- c("Honda","Toyota","Mazda","Ford","Dodge","Saturn")
Dealers
barplot(t(Deal),beside=TRUE,legend=TRUE)
detach(Dealers)
remove(Deal)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

