% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Defectiv}
\alias{Defectiv}
\title{Number of defective items produced by 20 employees}
\format{A data frame with 20 observations on the following variable.
\describe{ 
\item{C1}{a numeric vector} 
}}
\usage{
Defectiv
}
\description{
Data for Exercise 1.27
}
\examples{

T1 <- table(Defectiv$C1)
barplot(T1, col = "green", ylab = "Frequency",xlab="Defective Items Produced by Employees",main="Problem 1.27")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

