% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Defectiv}
\alias{Defectiv}
\title{Number of defective items produced by 20 employees}
\format{A data frame with 20 observations on the following 4 variables.
\describe{ 
\item{C1}{a numeric vector} 
\item{number}{a numeric vector} 
\item{Count}{a numeric vector}
\item{Percent}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 1.27
}
\examples{

str(Defectiv)
attach(Defectiv)
table(C1)
barplot(table(C1),col="pink",ylab="Frequency",
xlab="Defective Items Produced by Employees",main="Problem 1.27")
detach(Defectiv)

}
\keyword{datasets}

