% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Defectiv}
\alias{Defectiv}
\title{Number of defective items produced by 20 employees}
\format{A data frame/tibble  with 20 observations on one variable.
\describe{ 
\item{number}{number of defective items produced by the employees in a small business firm} 
}}
\usage{
Defectiv
}
\description{
Data for Exercise 1.27
}
\examples{

T1 <- xtabs(~ number, data = Defectiv)
T1
barplot(T1, col = "pink", ylab = "Frequency",
xlab = "Defective Items Produced by Employees", main = "Problem 1.27")
rm(T1)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
