% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Degree}
\alias{Degree}
\title{Percent of bachelor's degrees awarded women in 1970 versus 1990}
\format{A data frame with 11 observations on the following 3 variables.
\describe{ 
\item{field}{a factor with levels 
\code{Business} \code{Education} \code{Engineering} \code{Fine Arts}
\code{Foreign Lng} \code{Health} \code{Life Sciences} \code{Physical Sci}
\code{Psychology} \code{Social Science}} 
\item{percent}{percent of bachelor's degrees awarded to women} 
\item{year}{a factor with 2 levels \code{1970} and \code{1990}} 
}}
\usage{
Degree
}
\description{
Data for Exercise 2.75
}
\examples{
str(Degree)
T1 <- xtabs(percent ~ field + year, data = Degree)
barplot(t(T1), beside = TRUE, legend = TRUE, cex.names = 0.5)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

