% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Degree}
\alias{Degree}
\title{Percent of bachelor's degrees awarded women in 1970 versus 1990}
\format{A data frame with 1064 observations on the following 2 variables.
\describe{ 
\item{Field}{a factor with levels \code{Health}
\code{Education} \code{Foreign Language} \code{Psychology} \code{Fine Arts}
\code{Life Sciences} \code{Business} \code{Social Science} \code{Physical Sciences}
\code{Engineering} \code{All Fields}} 
\item{Awarded}{a factor with levels \code{1970} and \code{1990}} 
}}
\usage{
Degree
}
\description{
Data for Exercise 2.75
}
\examples{

str(Degree)
T1 <- xtabs(~Field + Awarded, data = Degree)
T1
barplot(t(T1), beside = TRUE, col = c("red", "skyblue"), legend = colnames(T1))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

