% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Degree}
\alias{Degree}
\title{Percent of bachelor's degrees awarded women in 1970 versus 1990}
\format{A data frame with 11 observations on the following 3 variables.
\describe{ 
\item{Field}{a factor with levels 
\code{Business} \code{Education} \code{Engineering} \code{Fine Arts}
\code{Foreign Lng} \code{Health} \code{Life Sciences} \code{Physical Sci}
\code{Psychology} \code{Social Science}} 
\item{X1970}{a numeric vector} 
\item{X1990}{a numeric vector} 
}}
\usage{
Degree
}
\description{
Data for Exercise 2.75
}
\examples{

Dmat <- as.matrix(Degree[,2:3])
rownames(Dmat) <- Degree$Field
colnames(Dmat) <- c("1970","1990")
Dmat
barplot(t(Dmat),beside=TRUE,legend=TRUE,cex.names=.5)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

