% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Delay}
\alias{Delay}
\title{Delay times on 20 flights from four major air carriers}
\format{A data frame with 80 observations on the following 6 variables.
\describe{ 
\item{delay}{a numeric vector} 
\item{carrier}{a numeric vector} 
}}
\usage{
Delay
}
\description{
Data for Exercise 10.55
}
\examples{
boxplot(Delay$delay~Delay$carrier)
kruskal.test(Delay$delay~as.factor(Delay$carrier))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

