% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Delay}
\alias{Delay}
\title{Delay times on 20 flights from four major air carriers}
\format{A data frame with 80 observations on the following 6 variables.
\describe{
\item{CarrierA}{a numeric vector}
\item{CarrierB}{a numeric vector}
\item{CarrierC}{a numeric vector}
\item{CarrierD}{a numeric vector}
\item{delay}{a numeric vector}
\item{Carrier}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 10.55
}
\examples{
str(Delay)
attach(Delay)
boxplot(delay~Carrier)
kruskal.test(delay~as.factor(Carrier))
detach(Delay)
}
\keyword{datasets}

