% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Delay}
\alias{Delay}
\title{Delay times on 20 flights from four major air carriers}
\format{A data frame/tibble  with 80 observations on the following two variables.
\describe{ 
\item{delay}{the delay time (in minutes) for 80 randomly selected flights}
\item{carrier}{a factor with levels \code{A}, \code{B}, \code{C}, and \code{D}} 
}}
\usage{
Delay
}
\description{
Data for Exercise 10.55
}
\examples{

boxplot(delay ~ carrier, data = Delay)
kruskal.test(delay ~carrier, data = Delay)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
