% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Depend}
\alias{Depend}
\title{Number of dependent children for 50 families}
\format{A data frame with 50 observations on the following 4 variables.
\describe{
\item{C1}{a numeric vector}
\item{number}{a numeric vector}
\item{Count}{a numeric vector}
\item{Percent}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 1.26
}
\examples{
str(Depend)
attach(Depend)
table(C1)
barplot(table(C1),col="lightblue",main="Problem 1.26",
xlab="Number of Dependent Children",ylab="Frequency")
detach(Depend)
}
\keyword{datasets}

