% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Depend}
\alias{Depend}
\title{Number of dependent children for 50 families}
\format{A data frame/tibble  with 50 observations on one variable.
\describe{ 
\item{number}{number of dependent children in a family} 
}}
\usage{
Depend
}
\description{
Data for Exercise 1.26
}
\examples{

T1 <- xtabs(~ number, data = Depend)
T1
barplot(T1, col = "lightblue", main = "Problem 1.26",
xlab = "Number of Dependent Children", ylab = "Frequency")
rm(T1)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
