% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Depend}
\alias{Depend}
\title{Number of dependent children for 50 families}
\format{A data frame with 50 observations on 1 variable.
\describe{ 
\item{children}{The number of dependent children for 50 families.} 
}}
\usage{
Depend
}
\description{
Data for Exercise 1.26
}
\examples{

str(Depend)
T1 <- xtabs(~children, data = Depend)
T1
barplot(T1, col = "lightblue", main = "Number of Dependent \\nChildren for 50 Families")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

