% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Detroit}
\alias{Detroit}
\title{Educational levels of a sample of 40 auto workers in Detroit}
\format{A data frame with 40 observations on the following variable.
\describe{
\item{educ}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 5.21
}
\examples{
str(Detroit)
attach(Detroit)
EDA(educ)
detach(Detroit)
}
\keyword{datasets}

