% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Detroit}
\alias{Detroit}
\title{Educational levels of a sample of 40 auto workers in Detroit}
\format{A data frame/tibble with 40 observations on one variable
\describe{ 
\item{educ}{the educational level (in years) of a sample of 40 auto workers in a plant in Detroit} 
}}
\usage{
Detroit
}
\description{
Data for Exercise 5.21
}
\examples{

EDA(Detroit$educ)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
