% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Develop}
\alias{Develop}
\title{Demographic characteristics of developmental students at 2-year colleges and
4-year colleges}
\format{A data frame/tibble with 5656 observations on two variables
\describe{ 
\item{race}{a factor with levels \code{African American}, \code{American Indian},
\code{Asian}, \code{Latino}, and \code{White}} 
\item{college}{a factor with levels \code{Two-year} and \code{Four-year}}
}}
\source{
\emph{Research in Development Education} (1994), V. 11, 2.
}
\usage{
Develop
}
\description{
Data used for Exercise 8.50
}
\examples{

T1 <- xtabs(~race + college, data = Develop)
T1
chisq.test(T1)
rm(T1)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
