% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Develop}
\alias{Develop}
\title{Demographic characteristics of developmental students at 2-year colleges and
4-year colleges}
\format{A data frame with 5656 observations on the following two variables.
\describe{ 
\item{Race}{a factor with levels \code{African American}, \code{American Indian},
\code{Asian}, \code{Latino}, and \code{White}} 
\item{College}{a factor with levels \code{Two-year} and \code{Four-year}}
}}
\usage{
Develop
}
\description{
Data used for Exercise 8.50
}
\examples{

str(Develop)
T1 <- xtabs(~Race + College, data = Develop)
T1
chisq.test(T1)
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

