% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Devmath}
\alias{Devmath}
\title{Test scores for students who failed developmental mathematics in the fall
semester 1995}
\format{A data frame/tibble with 40 observations on one variable
\describe{ 
\item{score}{first exam score} 
}}
\source{
Data provided by Dr. Anita Kitchens.
}
\usage{
Devmath
}
\description{
Data for Exercise 6.47
}
\examples{

EDA(Devmath$score)
t.test(Devmath$score, mu = 80, alternative = "less")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
