% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Dice}
\alias{Dice}
\title{Outcomes and probabilities of the roll of a pair of fair dice}
\format{A data frame with 11 observations on the following 2 variables.
\describe{ 
\item{x}{a numeric vector} 
\item{P.x.}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 3.109
}
\examples{

str(Dice)
attach(Dice)
roll1 <- sample(1:6,2000,replace=TRUE)
roll2 <- sample(1:6,2000,replace=TRUE)
outcome <- roll1+roll2
table(outcome)/length(outcome)
detach(Dice)
remove(roll1,roll2,outcome)

}
\keyword{datasets}

