% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Dice}
\alias{Dice}
\title{Outcomes and probabilities of the roll of a pair of fair dice}
\format{A data frame/tibble  with 11 observations on the following two variables.
\describe{ 
\item{x}{possible outcomes for the sum of two dice} 
\item{px}{probability for outcome \code{x}} 
}}
\usage{
Dice
}
\description{
Data for Exercise 3.109
}
\examples{

roll1 <- sample(1:6, 20000, replace = TRUE)
roll2 <- sample(1:6, 20000, replace = TRUE)
outcome <- roll1 + roll2
T1 <- table(outcome)/length(outcome)
remove(roll1, roll2, outcome)
T1
round(t(Dice), 5)
rm(roll1, roll2, T1)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
