% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Diesel}
\alias{Diesel}
\title{Diesel fuel prices in 1999-2000 in nine regions of the country}
\format{A data frame with 65 observations on the following 11 variables.
\describe{ 
\item{Date}{a date field}
\item{NatAvg}{a numeric vector} 
\item{EstCst}{a numeric vector} 
\item{NE}{a numeric vector} 
\item{CltAtl}{a numeric vector} 
\item{LwrAtl}{a numeric vector} 
\item{Gulf}{a numeric vector} 
\item{Rocky}{a numeric vector}
\item{WstMt}{a numeric vector} 
\item{Coast}{a numeric vector} 
\item{Calif}{a numeric vector}
}}
\usage{
Diesel
}
\description{
Data for Exercise 2.8
}
\examples{

str(Diesel)
boxplot(Diesel[,-1])
boxplot(Diesel$NatAvg,Diesel$EstCst,Diesel$Gulf,Diesel$Rocky,Diesel$Calif,
names=c("National Average","East Coast","Gulf","Rocky","California"),col="pink")


}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

