% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Diplomat}
\alias{Diplomat}
\title{Parking tickets issued to diplomats}
\format{A data frame with 10 observations on the following 4 variables.
\describe{ 
\item{country}{a factor with levels \code{Brazil}
\code{Bulgaria} \code{Egypt} \code{Indonesia} \code{Israel} \code{Nigeria}
\code{Russia} \code{S. Korea} \code{Ukraine} \code{Venezuela}}
\item{number}{a numeric vector} 
\item{rate}{a numeric vector} 
\item{code}{a factor with levels \code{Br} \code{Bu}
\code{Eg} \code{In} \code{Is} \code{Ni} \code{Ru} \code{SK} \code{Uk}
\code{Ve}} 
}}
\usage{
Diplomat
}
\description{
Data for Exercises 1.14 and 1.47
}
\examples{
par(mfrow=c(1,2))
names(Diplomat$number) <- Diplomat$country
dotchart(Diplomat$number,main="Number of Tickets",col="blue",pch=1)
names(Diplomat$rate) <- Diplomat$country
dotchart(Diplomat$rate,main="Tickets/Vehicle/Month",col="red",pch=2)
barplot(Diplomat$rate,col="pink")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

