% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Domestic}
\alias{Domestic}
\title{Rates of domestic violence per 1,000 women by age groups}
\format{A data frame with 5 observations on the following 2 variables.
\describe{
\item{Age}{a factor with levels \code{12-19} \code{20-24}
\code{25-34} \code{35-49} \code{50-64}}
\item{Rate}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 1.20
}
\examples{
str(Domestic)
attach(Domestic)
names(Rate) <- Age
barplot(Rate,col="gold")
pie(Rate)
detach(Domestic)
}
\keyword{datasets}

