% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Dopamine}
\alias{Dopamine}
\title{Dopamine b-hydroxylase activity of schizophrenic patients treated with an
antipsychotic drug}
\format{A data frame with 25 observations on the following 4 variables.
\describe{ 
\item{DBH}{a numeric vector} 
\item{group}{a numeric vector} 
}}
\usage{
Dopamine
}
\description{
Data for Exercises 5.14 and 7.49
}
\examples{

str(Dopamine)
boxplot(Dopamine$DBH~Dopamine$group,names=c("Non Psychotic","Psychotic"))
t.test(Dopamine$DBH~Dopamine$group,var.equal=TRUE)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

