% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Dowjones}
\alias{Dowjones}
\title{Closing yearend Dow Jones Industrial averages from 1896 through 2000}
\format{A data frame with 105 observations on the following 3 variables.
\describe{ 
<<<<<<< HEAD
\item{year}{a numeric vector} 
\item{close}{a numeric vector} 
\item{X.change}{a numeric vector} 
=======
\item{year}{date} 
\item{close}{Dow Jones closing price} 
\item{change}{percent change from previous year} 
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}}
\usage{
Dowjones
}
\description{
Data for Exercise 1.35
}
\examples{

str(Dowjones)
<<<<<<< HEAD
attach(Dowjones)  
plot(year,close,type="l",lty=2,lwd=2,col="blue")     
barplot(close,col="blue",las=2,main="Problem 1.35",names.arg=FALSE)      
detach(Dowjones) 

=======
plot(close ~ year, data = Dowjones, type = "l", main = "Exercise 1.35")
\dontrun{
ggplot2::ggplot(data = Dowjones, aes(x = year, y = close)) +
geom_point(size = 0.5) + 
geom_line(color = "red") + 
theme_bw() + 
labs(y = "Dow Jones Closing Price")
}
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

