% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Drink}
\alias{Drink}
\title{Opinion on referendum by view on moral issue of selling alcoholic beverages}
\format{A data frame with 427 observations on the following 2 variables.
\describe{ 
\item{drink}{a factor with levels \code{OK}, \code{Tolerated}, and \code{Immoral}}
\item{view}{a factor with levels \code{For}, \code{Against}, and \code{Undecided}}
}}
\usage{
Drink
}
\description{
Data for Exercise 8.53
}
\examples{
head(Drink)
class(Drink)
CT <- xtabs(~drink + view, data = Drink)
CT
chisq.test(CT)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

