% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Drug}
\alias{Drug}
\title{Number of trials to master a task for a group of 28 subjects assigned to a
control and an experimental group}
\format{A data frame with 28 observations on the following 2 variables.
\describe{ 
\item{trials}{the number of trials required to master a given task} 
\item{group}{a factor with levels \code{experimental} and \code{control}} 
}}
\usage{
Drug
}
\description{
Data for Example 7.15
}
\examples{

str(Drug)
boxplot(trials ~ group, data = Drug)
wilcox.test(trials ~ group, data = Drug)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

