% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Drug}
\alias{Drug}
\title{Number of trials to master a task for a group of 28 subjects assigned to a
control and an experimental group}
\format{A data frame with 28 observations on the following 2 variables.
\describe{ 
\item{trials}{a numeric vector} 
\item{group}{a numeric vector} 
}}
\usage{
Drug
}
\description{
Data for Exercise 7.15
}
\examples{

str(Drug)
boxplot(Drug$trials~Drug$group)
wilcox.test(Drug$trials~Drug$group)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

