% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Dyslexia}
\alias{Dyslexia}
\title{Data on a group of college students diagnosed with dyslexia}
\format{A data frame with 8 observations on the following seven variables.
\describe{ 
\item{words}{number of words read per minute} 
\item{age}{age of participant} 
\item{gender}{a factor with levels \code{female} and 
\code{male}} 
\item{handed}{a factor with levels \code{left} and \code{right}}
\item{weight}{weight of participant (in pounds)} 
\item{height}{height of participant (in inches)} 
\item{children}{number of children in family} 
}}
\usage{
Dyslexia
}
\description{
Data for Exercise 2.90
}
\examples{

str(Dyslexia)
plot(height ~ weight, data = Dyslexia)
plot(words ~ handed, data = Dyslexia)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
