% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Earthqk}
\alias{Earthqk}
\title{One hundred year record of worldwide seismic activity(1770-1869)}
\format{A data frame with 100 observations on the following two variables.
\describe{ 
\item{year}{year seimic activity recorded} 
\item{severity}{annual incidence of sever earthquakes} 
}}
\usage{
Earthqk
}
\description{
Data for Exercise 6.97
}
\examples{

str(Earthqk)
EDA(Earthqk$severity)
t.test(Earthqk$severity, mu = 100, alternative = "greater")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

