% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Earthqk}
\alias{Earthqk}
\title{One hundred year record of worldwide seismic activity(1770-1869)}
\format{A data frame with 100 observations on the following 2 variables.
\describe{ 
\item{year}{a numeric vector} 
\item{severity}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 6.97
}
\examples{

str(Earthqk)
attach(Earthqk)
EDA(severity)
t.test(severity,mu=100,alternative="greater")
detach(Earthqk)

}
\keyword{datasets}

