% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Educat}
\alias{Educat}
\title{Crime rates versus the percent of the population without a high school
degree}
\format{A data frame with 51 observations on the following 3 variables.
\describe{
\item{State}{a factor with levels \code{Alabama}
\code{Alaska} \code{Arizona} \code{Arkansas} \code{California}
\code{Colorado} \code{Connecticut} \code{DC} \code{Delaware} \code{Florida}
\code{Georgia} \code{Hawaii} \code{Idaho} \code{Illinois} \code{Indiana}
\code{Iowa} \code{Kansas} \code{Kentucky} \code{Louisiana} \code{Maine}
\code{Maryland} \code{Massachusetts} \code{Michigan} \code{Minnesota}
\code{Mississippi} \code{Missour} \code{Montana} \code{Nebraska}
\code{Nevada} \code{New Hampshire} \code{New Jersey} \code{New Mexico}
\code{New York} \code{North Carolina} \code{North Dakota} \code{Ohio}
\code{Oklahoma} \code{Oregon} \code{Pennsylvania} \code{Rhode Island}
\code{South Carolina} \code{South Dakota} \code{Tennessee} \code{Texas}
\code{Utah} \code{Vermont} \code{Virginia} \code{Washington} \code{West
Virginia} \code{Wisconsin} \code{Wyoming}}
\item{nodegree}{a numeric vector}
\item{crime}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 2.41
}
\examples{
str(Educat)
attach(Educat)
plot(nodegree,crime,xlab="No Crime",ylab="Violent Crime Rate per 100,000")
detach(Educat)
}
\keyword{datasets}

