% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Eggs}
\alias{Eggs}
\title{Number of eggs versus amounts of feed supplement}
\format{A data frame with 12 observations on the following 7 variables.
\describe{ 
\item{feed}{a numeric vector} 
\item{eggs}{a numeric vector} 
\item{SRES1}{a numeric vector}
\item{FITS1}{a numeric vector} 
\item{c1sq}{a numeric vector}
\item{SRES2}{a numeric vector} 
\item{FITS2}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 9.22
}
\examples{

str(Eggs)
attach(Eggs)
plot(feed,eggs)
model <- lm(eggs~feed)
abline(model)
summary(model)
detach(Eggs)
remove(model)

}
\keyword{datasets}

