% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Eggs}
\alias{Eggs}
\title{Number of eggs versus amounts of feed supplement}
\format{A data frame/tibble  with 12 observations on the following two variables.
\describe{ 
\item{feed}{amount of feed supplement} 
\item{eggs}{number of eggs per day for 100 chickens} 
}}
\usage{
Eggs
}
\description{
Data for Exercise 9.22
}
\examples{

plot(eggs ~ feed, data = Eggs)
model <- lm(eggs ~ feed, data = Eggs)
abline(model)
summary(model)
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
