% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Energy}
\alias{Energy}
\title{Amount of energy consumed by homes versus their sizes}
\format{A data frame with 12 observations on the following 5 variables.
\describe{ 
\item{Size}{a numeric vector} 
\item{kilowatt}{a numeric vector} 
\item{SRES1}{a numeric vector}
\item{FITS1}{a numeric vector} 
\item{Residuals}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercises 2.5, 2.24, and 2.55
}
\examples{

str(Energy)
attach(Energy)
plot(Size,kilowatt)
cor(Size,kilowatt)
model <- lm(kilowatt~Size)
plot(Size,resid(model))
detach(Energy)

}
\keyword{datasets}

