% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Engineer}
\alias{Engineer}
\title{Salaries after 10 years for graduates of three different universities}
\format{A data frame with 51 observations on the following 6 variables.
\describe{ 
\item{UnivA}{a numeric vector} 
\item{UnivB}{a numeric vector} 
\item{UnivC}{a numeric vector}
\item{salary}{a numeric vector} 
\item{university}{a numeric vector} 
\item{ranks}{a numeric vector} 
}}
\description{
Data for Example 10.7
}
\examples{

str(Engineer)
attach(Engineer)
boxplot(salary~university)
kruskal.test(salary~as.factor(university))
detach(Engineer)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

