% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Entrance}
\alias{Entrance}
\title{College entrance exam scores for 24 high school seniors}
\format{A data frame with 24 observations on the following variable.
\describe{ 
\item{score}{a numeric vector} 
}}
\description{
Data for Example 1.8
}
\examples{

str(Entrance)
attach(Entrance)
stem(score)
detach(Entrance)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

