% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Entrance}
\alias{Entrance}
\title{College entrance exam scores for 24 high school seniors}
\format{
A data frame/tibble with 24 observations on one variable
\describe{ 
\item{score}{college entrance exam score} 
}
}
\usage{
Entrance
}
\description{
Data for Example 1.8
}
\examples{

stem(Entrance$score)
stem(Entrance$score, scale = 2)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
