% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Executiv}
\alias{Executiv}
\title{Ages of 25 executives}
\format{A data frame with 25 observations on the following variable.
\describe{ 
\item{Age}{a numeric vector} 
}}
\description{
Data for Exercise 1.104
}
\examples{

str(Executiv)
attach(Executiv)
EDA(Age)
detach(Executiv)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
