% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Executiv}
\alias{Executiv}
\title{Ages of 25 executives}
\format{A data frame/tibble with 25 observations on one variable
\describe{ 
\item{age}{a numeric vector} 
}}
\usage{
Executiv
}
\description{
Data for Exercise 1.104
}
\examples{

hist(Executiv$age, xlab = "Age of banking executives", 
breaks = 5, main = "", col = "gray")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
