% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Exercise}
\alias{Exercise}
\title{Weight loss for 30 members of an exercise program}
\format{A data frame with 30 observations on the following variable.
\describe{ 
\item{loss}{a numeric vector} 
}}
\description{
Data for Exercise 1.44
}
\examples{

attach(Exercise)
stem(loss)
detach(Exercise)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
