% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Fabric}
\alias{Fabric}
\title{Measures of softness of 10 different clothing garments washed with and
without a softener}
\format{A data frame/tibble with 20 observations on the following 3 variables.
\describe{ 
\item{garment}{a numeric vector} 
\item{softner}{a character variable with values \code{with} and \code{without}} 
\item{softness}{a numeric vector} 
}}
\usage{
Fabric
}
\description{
Data for Example 7.21
}
\examples{

wilcox.test(softness ~ softner, data = Fabric, 
            paired = TRUE, alternative = "greater")

\dontrun{
T7 <- tidyr::spread(Fabric, softner, softness) \%>\% 
mutate(di = with - without, adi = abs(di), rk = rank(adi), 
       srk = sign(di)*rk)
T7
t.test(T7$srk, alternative = "greater")
}
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
