% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Fabric}
\alias{Fabric}
\title{Measures of softness of 10 different clothing garments washed with and
without a softener}
\format{A data frame with 10 observations on the following 3 variables.
\describe{ 
\item{Type}{a numeric vector} 
\item{With}{a numeric vector} 
\item{Without}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.21
}
\examples{

str(Fabric)
attach(Fabric)
DIF <- With - Without
qqnorm(DIF)
qqline(DIF)
shapiro.test(DIF)
wilcox.test(With,Without,paired=TRUE,alternative="greater")
detach(Fabric)
remove(DIF)

}
\keyword{datasets}

