% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Faithful}
\alias{Faithful}
\title{Waiting times between successive eruptions of the Old Faithful geyser}
\format{A data frame with 299 observations on the following 2 variables.
\describe{ 
\item{Time}{a numeric vector} 
\item{Eruption}{a numeric vector} 
}}
\description{
Data for Exercise 5.12 and 5.111
}
\examples{

str(Faithful)
attach(Faithful)
hist(Time,prob=TRUE,xlab="Waiting time between eruptions",col="tomato")
lines(density(Time),col="red",lwd=3)
t.test(Time)$conf
detach(Faithful)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

