% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Family}
\alias{Family}
\title{Size of family versus cost per person per week for groceries}
\format{A data frame with 20 observations on the following 2 variables.
\describe{ 
\item{Number}{a numeric vector} 
\item{Cost}{a numeric vector} 
}}
\description{
Data for Exercise 2.89
}
\examples{

str(Family)
attach(Family)
plot(Number,Cost)
cor(Number,Cost)
lm(Cost~Number)
detach(Family)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

