% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Ferraro1}
\alias{Ferraro1}
\title{Choice of presidental ticket in 1984 by gender}
\format{A data frame with 2 observations on the following 4 variables.
\describe{ 
\item{gender}{a factor with levels \code{Men}
\code{Women}} 
\item{Reag.Bs}{a numeric vector}
\item{Mond.Fer}{a numeric vector} 
\item{undecide}{a numeric vector} 
}}
\description{
Data for Exercise 8.23
}
\examples{

str(Ferraro1)
attach(Ferraro1)
Ferraro1
chisq.test(Ferraro1[,2:4])
detach(Ferraro1)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
