% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Ferraro2}
\alias{Ferraro2}
\title{Choice of vice presidental candidate in 1984 by gender}
\format{A data frame with 2 observations on the following 4 variables.
\describe{ 
\item{gender}{a factor with levels \code{Men}
\code{Women}} 
\item{Bush}{a numeric vector} 
\item{Ferraro}{a numeric vector} 
\item{undecide}{a numeric vector} 
}}
\description{
Data for Exercise 8.23
}
\examples{

str(Ferraro2)
attach(Ferraro2)
Ferraro2
chisq.test(Ferraro2[,2:4])
detach(Ferraro2)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

