% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Ferraro2}
\alias{Ferraro2}
\title{Choice of vice presidental candidate in 1984 by gender}
\format{A data frame/tibble with 1000 observations on two variables
\describe{ 
\item{gender}{a factor with levels \code{Men} and
\code{Women}} 
\item{candidate}{a character vector of 1984 president and vice-president candidates} 
}}
\usage{
Ferraro2
}
\description{
Data for Exercise 8.23
}
\examples{

T1 <- xtabs(~gender + candidate, data = Ferraro2)
T1
chisq.test(T1)  
rm(T1)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
