% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Firstchi}
\alias{Firstchi}
\title{Ages of women at the birth of their first child}
\format{A data frame with 87 observations on the following variable.
\describe{
\item{age}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 5.11
}
\examples{
str(Firstchi)
attach(Firstchi)
EDA(age)
detach(Firstchi)
}
\keyword{datasets}

