% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Firstchi}
\alias{Firstchi}
\title{Ages of women at the birth of their first child}
\format{
A data frame/tibble with 87 observations on one variable
\describe{ 
\item{age}{age of woman at birth of her first child} 
}
}
\usage{
Firstchi
}
\description{
Data for Exercise 5.11
}
\examples{

EDA(Firstchi$age)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
