% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Fitness}
\alias{Fitness}
\title{Number of sit-ups before and after a physical fitness course}
\format{A data frame with 9 observations on the following 2 variables.
\describe{ 
\item{Before}{a numeric vector} 
\item{After}{a numeric vector} 
}}
\description{
Data for Exercise 7.71
}
\examples{

str(Fitness)
attach(Fitness)
DIF <- After - Before
qqnorm(DIF)
qqline(DIF)
shapiro.test(DIF)
t.test(After,Before,paired=TRUE,alternative="greater")
detach(Fitness)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

