% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Florida2000}
\alias{Florida2000}
\title{Florida voter results in the 2000 presidential election}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 67 rows and 12 columns.}
\usage{
Florida2000
}
\description{
Data for Statistical Insight Chapter 2
}
\examples{

plot(buchanan ~ total, data = Florida2000, 
xlab = "Total votes cast (in thousands)", 
ylab = "Votes for Buchanan")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
