% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Fluid}
\alias{Fluid}
\title{Breakdown times of an insulating fluid under various levels of voltage
stress}
\format{A data frame with 76 observations on the following 10 variables.
\describe{ 
\item{X26kV}{a numeric vector} 
\item{X28kV}{a numeric vector} 
\item{X30kV}{a numeric vector}
\item{X32kV}{a numeric vector} 
\item{X34kV}{a numeric vector} 
\item{X36kV}{a numeric vector} 
\item{X38kV}{a numeric vector} 
\item{response}{a numeric vector}
\item{group}{a numeric vector} 
\item{ln.resp.}{a numeric vector} 
}}
\description{
Data for Exercise 5.76
}
\examples{

str(Fluid)
attach(Fluid)
stem(X34kV)
SIGN.test(X34kV)
detach(Fluid)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

