% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Food}
\alias{Food}
\title{Annual food expenditures for 40 single households in Ohio}
\format{A data frame/tibble with 40 observations on one variable
\describe{ 
\item{expenditure}{a numeric vector recording annual food expenditure (in dollars) in the state of Ohio.} 
}}
\source{
Bureau of Labor Statistics.
}
\usage{
Food
}
\description{
Data for Exercise 5.106
}
\examples{

EDA(Food$expenditure)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
