% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Food}
\alias{Food}
\title{Annual food expenditures for 40 single households in Ohio}
\format{A data frame with 40 observations on the following variable.
\describe{ 
\item{food}{a numeric vector} 
}}
\description{
Data for Exercise 5.106
}
\examples{

str(Food)
attach(Food)
EDA(food)
detach(Food)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

