% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Framingh}
\alias{Framingh}
\title{Cholesterol values of 62 subjects in the Framingham Heart Study}
\format{A data frame with 62 observations on the following variable.
\describe{ 
\item{cholest}{a numeric vector} 
}}
\description{
Data for Exercises 1.55, 1.75, 3.69, and 5.60
}
\examples{

str(Framingh)
attach(Framingh)
stem(cholest)
hist(cholest,prob=TRUE,ylim=c(0,.012))
lines(density(cholest))
boxplot(cholest,col="brown")
sum(cholest>200&cholest<240)/length(cholest)
detach(Framingh)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
