% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Freshman}
\alias{Freshman}
\title{Ages of a random sample of 30 college freshmen}
\format{A data frame with 30 observations on the following variable.
\describe{ 
\item{age}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 6.53
}
\examples{

str(Freshman)
attach(Freshman)
SIGN.test(age,md=19)
detach(Freshman)

}
\keyword{datasets}

