% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Galaxie}
\alias{Galaxie}
\title{Velocities of 82 galaxies in the Corona Borealis region}
\format{A data frame with 82 observations on the following variable.
\describe{ 
\item{velocity}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Example 5.2
}
\examples{

str(Galaxie)
attach(Galaxie)
EDA(velocity)
detach(Galaxie)

}
\keyword{datasets}

