% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Galaxie}
\alias{Galaxie}
\title{Velocities of 82 galaxies in the Corona Borealis region}
\format{A data frame/tibble with 82 observations on one variable
\describe{ 
\item{velocity}{velocity measured in kilometers per second} 
}}
\source{
K. Roeder, "Density Estimation with Confidence Sets Explained by Superclusters
and Voids in the Galaxies," \emph{Journal of the American Statistical Association}, 85
(1990), 617-624
}
\usage{
Galaxie
}
\description{
Data for Example 5.2
}
\examples{

EDA(Galaxie$velocity)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
