% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Galaxie}
\alias{Galaxie}
\title{Velocities of 82 galaxies in the Corona Borealis region}
\format{A data frame/tibble with 82 observations on the following variable.
\describe{ 
\item{velocity}{velocity measured in kilometers per second} 
}}
\usage{
Galaxie
}
\description{
Data for Example 5.2
}
\examples{

EDA(Galaxie$velocity)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
