% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Galaxie}
\alias{Galaxie}
\title{Velocities of 82 galaxies in the Corona Borealis region}
\format{A data frame with 82 observations on the following variable.
\describe{ 
\item{velocity}{a numeric vector} 
}}
\description{
Data for Example 5.2
}
\examples{

str(Galaxie)
attach(Galaxie)
EDA(velocity)
detach(Galaxie)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

