% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Gasoline}
\alias{Gasoline}
\title{Price of regular unleaded gasoline obtained from 25 service stations}
\format{A data frame with 25 observations on the following variable.
\describe{ 
\item{price}{a numeric vector} 
}}
\description{
Data for Exercise 1.45
}
\examples{

str(Gasoline)
attach(Gasoline)
stem(price)
detach(Gasoline)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
