% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{German}
\alias{German}
\title{Number of errors in copying a German passage before and after an
experimental course in German}
\format{A data frame with 10 observations on the following 4 variables.
\describe{ 
\item{Before}{a numeric vector} 
\item{After}{a numeric vector} 
\item{differ}{a numeric vector}
\item{sgnrnks}{a numeric vector} 
}}
\description{
Data for Exercise 7.60
}
\examples{

str(German)
attach(German)
qqnorm(differ)
qqline(differ)
shapiro.test(differ)
wilcox.test(Before,After,paired=TRUE)
detach(German)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

