% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{German}
\alias{German}
\title{Number of errors in copying a German passage before and after an
experimental course in German}
\format{A data frame/tibble with 10 observations on the following 3 variables.
\describe{ 
\item{student}{a character variable indicating student number} 
\item{when}{a character variable with either code\{Before} or \code{After} 
to indicate when the student received experimental instruction in German
\item{errors}{A vector containing the number of errors}
}}
\usage{
German
}
\description{
Data for Exercise 7.60
}
\examples{

t.test(errors ~ when, data = German, paired = TRUE)
wilcox.test(errors ~ when, data = German)

\dontrun{
T8 <- tidyr::spread(German, when, errors) \%>\%
mutate(di = After - Before, adi = abs(di), rk = rank(adi), srk = sign(di)*rk)
T8
qqnorm(T8$di)
qqline(T8$di)
t.test(T8$srk)
}

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
