% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Golf}
\alias{Golf}
\title{Distances a golf ball can be driven by 20 professional golfers}
\format{A data frame with 20 observations on the following variable.
\describe{ 
\item{yards}{a numeric vector} 
}}
\description{
Data for Exercise 5.24
}
\examples{

str(Golf)
attach(Golf)
stem(yards)
EDA(yards)
detach(Golf)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

