% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Gpa}
\alias{Gpa}
\title{High school GPA versus college GPA}
\format{A data frame with 10 observations on the following 2 variables.
\describe{
\item{HSGPA}{a numeric vector}
\item{CollGPA}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 2.13
}
\examples{
str(Gpa)
attach(Gpa)
plot(HSGPA,CollGPA)
model <- lm(CollGPA~HSGPA)
abline(model)
model
r <- resid(model)
yhat <- fitted(model)
Table2.1 <- cbind(HSGPA,CollGPA,yhat,r)
Table2.1
remove(r,yhat,model,Table2.1)
detach(Gpa)
}
\keyword{datasets}

