% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Grades}
\alias{Grades}
\title{Test grades in a beginning statistics class}
\format{A data frame with 29 observations on the following variable.
\describe{ 
\item{grades}{a numeric vector} 
}}
\description{
Data for Exercise 1.120
}
\examples{

str(Grades)
attach(Grades)
EDA(grades)
detach(Grades)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

