% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Grades}
\alias{Grades}
\title{Test grades in a beginning statistics class}
\format{A data frame with 29 observations on the following variable.
\describe{ 
\item{grades}{a numeric vector} 
}}
\usage{
Grades
}
\description{
Data for Exercise 1.120
}
\examples{

hist(Grades$grades, main = "", xlab = "Test grades", right = FALSE)

\dontrun{
ggplot2::ggplot(data = Grades, aes(x = grades, y = ..density..)) + 
    geom_histogram(fill = "pink", binwidth = 5, color = "black") + 
    geom_density(lwd = 2, color = "red") + 
    theme_bw() 
}

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
