% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Graduate}
\alias{Graduate}
\title{Graduation rates for student athletes in the Southeastern Conf.}
\format{A data frame with 12 observations on the following 3 variables.
\describe{ 
\item{School}{a factor with levels \code{Alabama}
\code{Arkansas} \code{Auburn} \code{Florida} \code{Georgia} \code{Kentucky}
\code{Louisiana St} \code{Mississippi} \code{Mississippi St} \code{South
Carolina} \code{Tennessee} \code{Vanderbilt}} 
\item{Code}{a factor with levels \code{Al} \code{Ar} \code{Au} \code{Fl} \code{Ge} \code{Ke}
\code{LSt} \code{Mi} \code{MSt} \code{SC} \code{Te} \code{Va}}
\item{Percent}{a numeric vector} 
}}
\description{
Data for Exercise 1.118
}
\examples{

str(Graduate)
attach(Graduate)
names(Percent) <- School
barplot(Percent,las=2,cex.names=.65,col="tomato")
detach(Graduate)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

