% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Graduate}
\alias{Graduate}
\title{Graduation rates for student athletes in the Southeastern Conf.}
\format{
A data frame/tibble with 12 observations on three variables
\describe{ 
\item{school}{a character variable with values \code{Alabama},
\code{Arkansas}, \code{Auburn}, \code{Florida}, \code{Georgia}, \code{Kentucky},
\code{Louisiana St}, \code{Mississippi}, \code{Mississippi St}, \code{South
Carolina,} \code{Tennessee}, and \code{Vanderbilt}} 
\item{code}{a character variable with values \code{Al}, \code{Ar}, \code{Au}
 \code{Fl}, \code{Ge}, \code{Ke}, \code{LSt}, \code{Mi}, \code{MSt}, \code{SC}, 
 \code{Te}, and \code{Va}}
\item{percent}{graduation rate} 
}
}
\usage{
Graduate
}
\description{
Data for Exercise 1.118
}
\examples{

barplot(Graduate$percent, names.arg = Graduate$school, 
        las = 2, cex.names = 0.7, col = "tomato")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
