% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Greenriv}
\alias{Greenriv}
\title{Varve thickness from a sequence through an Eocene lake deposit in the Rocky
Mountains}
\format{
A data frame/tibble with 37 observations on one variable
\describe{ 
\item{thick}{varve thickness in millimeters} 
}
}
\usage{
Greenriv
}
\description{
Data for Exercise 6.57
}
\examples{

stem(Greenriv$thick)
SIGN.test(Greenriv$thick, md = 7.3, alternative = "greater")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
