% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Grnriv2}
\alias{Grnriv2}
\title{Thickness of a varved section of the Green river oil shale deposit near a
major lake in the Rocky Mountains}
\format{
A data frame/tibble with 101 observations on one variable
\describe{ 
\item{thick}{varve thickness (in millimeters)} 
}
}
\source{
J. Davis, \emph{Statistics and Data Analysis in Geology}, 2nd Ed., Jon Wiley and Sons, New York.
}
\usage{
Grnriv2
}
\description{
Data for Exercises 6.45 and 6.98
}
\examples{

stem(Grnriv2$thick)
t.test(Grnriv2$thick, mu = 8, alternative = "less")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
