% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Groupabc}
\alias{Groupabc}
\title{Group data to illustrate analysis of variance}
\format{A data frame/tibble with 45 observations on the following 2 variables.
\describe{ 
\item{group}{a factor with levels \code{A}, \code{B}, and \code{C}} 
\item{response}{a numeric vector} 
}}
\usage{
Groupabc
}
\description{
Data for Exercise 10.42
}
\examples{

boxplot(response ~ group, data = Groupabc, 
        col = c("red", "blue", "green"))
        anova(lm(response ~ group, data = Groupabc))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
