% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Groups}
\alias{Groups}
\title{An illustration of analysis of variance}
\format{A data frame with 26 observations on the following 3 variables.
\describe{ 
\item{GroupA}{a numeric vector} 
\item{GroupB}{a numeric vector} 
\item{GroupC}{a numeric vector} 
}}
\description{
Data for Exercise 10.4
}
\examples{

str(Groups)
attach(Groups)
STACKED <-stack(Groups)
STACKED[1:5,]
boxplot(values~ind,col=c("red","blue","green"),data=STACKED)
anova(lm(values~ind,data=STACKED))
remove(STACKED)
detach(Groups)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

