% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Groups}
\alias{Groups}
\title{An illustration of analysis of variance}
\format{A data frame/tibble with 78 observations on the following 2 variables.
\describe{ 
\item{group}{a factor with levels \code{A}, \code{B}, and \code{C}} 
\item{response}{a numeric vector} 
}}
\usage{
Groups
}
\description{
Data for Exercise 10.4
}
\examples{

boxplot(response ~ group, data = Groups, col = c("red", "blue", "green"))
anova(lm(response ~ group, data = Groups))


}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
