% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Gym}
\alias{Gym}
\title{Children's age versus number of completed gymnastic activities}
\format{A data frame/tibble with eight observations on three variables
\describe{
\item{age}{age of child} 
\item{number}{number of gymnastic activities successfully completed} 
}}
\usage{
Gym
}
\description{
Data for Exercises 2.21 and 9.14
}
\examples{

plot(number ~ age, data = Gym)
model <- lm(number ~ age, data = Gym)
abline(model, col = "red")
summary(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
