% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Gym}
\alias{Gym}
\title{Children's age versus number of completed gymnastic activities}
\format{A data frame with 8 observations on the following 3 variables.
\describe{
\item{age}{a numeric vector} 
\item{number}{a numeric vector} 
\item{x.}{a numeric vector} 
}}
\description{
Data for Exercises 2.21, 9.14, and 9.32
}
\examples{

str(Gym)
attach(Gym)
plot(age,number)
model <- lm(number~age)
abline(model)
cor(age,number)
detach(Gym)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
