% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Habits}
\alias{Habits}
\title{Study habits of students in two matched school districts}
\format{A data frame with 11 observations on the following 4 variables.
\describe{ 
\item{A}{a numeric vector} 
\item{B}{a numeric vector} 
\item{differ}{a numeric vector} 
\item{signrks}{a numeric vector} 
}}
\description{
Data for Exercise 7.57
}
\examples{

str(Habits)
attach(Habits)
qqnorm(differ)
qqline(differ)
shapiro.test(differ)
t.test(B,A,paired=TRUE,alternative="less")
wilcox.test(B,A,paired=TRUE,alternative="less")
detach(Habits)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

