% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Haptologo}
\alias{Haptologo}
\title{Haptoglobin concentration in blood serum of 8 healthy adults}
\format{A data frame with 8 observations on the following variable.
\describe{ 
\item{concent}{a numeric vector} 
}}
\description{
Data for Example 6.9
}
\examples{

str(Haptologo)
attach(Haptologo)
qqnorm(concent,col="blue")
qqline(concent,col="red")
shapiro.test(concent)
t.test(concent,mu=2,alternative="less")
detach(Haptologo)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
