% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Hardware}
\alias{Hardware}
\title{Daily receipts for a small hardware store for 31 working days}
\format{A data frame with 31 observations on one variable
\describe{ 
\item{receipt}{a numeric vector of daily receipts (in dollars)} 
}}
\source{
J.C. Miller and J.N. Miller, (1988), \emph{Statistics for Analytical Chemistry}, 2nd Ed. 
(New York: Halsted Press).
}
\usage{
Hardware
}
\description{
Daily receipts for a small hardware store for 31 working days
}
\examples{

stem(Hardware$receipt)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
