% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Hardware}
\alias{Hardware}
\title{Daily receipts for a small hardware store for 31 working days}
\format{A data frame with 31 observations on the following variable.
\describe{
\item{receipt}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Example 2.18
}
\examples{
data(Hardware)
}
\keyword{datasets}

