% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Heat}
\alias{Heat}
\title{Primary heating sources of homes on indian reservations versus all
households}
\format{A data frame with 6 observations on the following 4 variables.
\describe{ 
\item{Fuel}{a factor with levels \code{Electricity}
\code{Fuel Oil} \code{LP bottled gas} \code{Other} \code{Utility gas}
\code{Wood}} 
\item{Reserv}{a numeric vector} 
\item{All.US}{a numeric vector} 
\item{Not.Rese}{a numeric vector} 
}}
\description{
Data for Exercise 1.29
}
\examples{

str(Heat)
attach(Heat)
MAT <- cbind(Reserv, All.US, Not.Rese) 
row.names(MAT) <- c("Utility Gas","LP bottled Gas","Electricity",
"Fuel Oil","Wood","Other Fuel")
MAT
barplot(t(MAT),beside=TRUE,legend=TRUE,main="Heating of American Indian Homes")
sum(Reserv)
sum(All.US)
sum(Not.Rese)
detach(Heat)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

