% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Heating}
\alias{Heating}
\title{Fuel efficiency ratings for three types of oil heaters}
\format{A data frame with 30 observations on the following 6 variables.
\describe{ 
\item{TypeA}{a numeric vector} 
\item{TypeB}{a numeric vector} 
\item{TypeC}{a numeric vector}
\item{Rating}{a numeric vector} 
\item{Type}{a numeric vector} 
\item{Ranks}{a numeric vector} 
}}
\description{
Data for Exercise 10.32
}
\examples{

str(Heating)
attach(Heating)
boxplot(Rating~Type)
kruskal.test(Rating~as.factor(Type))
detach(Heating)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
