% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Heating}
\alias{Heating}
\title{Fuel efficiency ratings for three types of oil heaters}
\format{A data frame/tibble with 90 observations on the two variables
\describe{ 
\item{type}{a factor with levels \code{A}, \code{B}, and \code{C} denoting the type of oil heater} 
\item{efficiency}{heater efficiency rating} 
}}
\usage{
Heating
}
\description{
Data for Exercise 10.32
}
\examples{

boxplot(efficiency ~ type, data = Heating, 
        col = c("red", "blue", "green"))
kruskal.test(efficiency ~ type, data = Heating)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
