% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Hodgkin}
\alias{Hodgkin}
\title{Results of treatments for Hodgkin's disease}
\format{A data frame with 4 observations on the following 4 variables.
\describe{
\item{Histological}{a factor with levels \code{LD}
\code{LP} \code{MC} \code{NS}}
\item{Positive}{a numeric vector}
\item{Partial}{a numeric vector}
\item{None}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 2.77
}
\examples{
str(Hodgkin)
attach(Hodgkin)
HOD <- as.matrix(Hodgkin[,2:4])
rownames(HOD) <- Histological
HOD
barplot(t(HOD),legend=TRUE,beside=TRUE)
detach(Hodgkin)
remove(HOD)
}
\keyword{datasets}

