% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Homework}
\alias{Homework}
\title{Number of hours per week spent on homework for private and public high
school students}
\format{A data frame with 30 observations on the following 2 variables.
\describe{ 
\item{school}{type of school either \code{private} or \code{public}} 
\item{time}{number of hours per week spent on homework} 
}}
\usage{
Homework
}
\description{
Data for Exercise 7.78
}
\examples{

boxplot(time ~ school, data = Homework, 
        ylab = "Hours per week spent on homework")
#
t.test(time ~ school, data = Homework)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
