% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Homework}
\alias{Homework}
\title{Number of hours per week spent on homework for private and public high
school students}
\format{A data frame with 15 observations on the following 2 variables.
\describe{ 
\item{Private}{a numeric vector} 
\item{Public}{a numeric vector} 
}}
\description{
Data for Exercise 7.78
}
\examples{

str(Homework)
attach(Homework)
boxplot(Private,Public)
t.test(Private,Public,conf.level=.98)
detach(Homework)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

