% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Homework}
\alias{Homework}
\title{Number of hours per week spent on homework for private and public high
school students}
\format{A data frame with 15 observations on the following 2 variables.
\describe{
\item{Private}{a numeric vector}
\item{Public}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.78
}
\examples{
str(Homework)
attach(Homework)
boxplot(Private,Public)
t.test(Private,Public,conf.level=.98)
detach(Homework)
}
\keyword{datasets}

