% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Honda}
\alias{Honda}
\title{Miles per gallon for a Honda Civic on 35 different occasions}
\format{A data frame/tibble with 35 observations on one variable
\describe{ 
\item{mileage}{miles per gallon for a Honda Civic} 
}}
\usage{
Honda
}
\description{
Data for Statistical Insight Chapter 6
}
\examples{

t.test(Honda$mileage, mu = 40, alternative = "less")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
