% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Hostile}
\alias{Hostile}
\title{Hostility levels of high school students from rural, suburban, and urban
areas}
\format{A data frame with 45 observations on the following 6 variables.
\describe{ 
\item{Rural}{a numeric vector} 
\item{Suburban}{a numeric vector} 
\item{Urban}{a numeric vector} 
\item{HLT}{a numeric vector} 
\item{Type}{a numeric vector} 
\item{Ranks}{a numeric vector} 
}}
\description{
Data for Example 10.6
}
\examples{

str(Hostile)
attach(Hostile)
boxplot(HLT~Type)
kruskal.test(HLT~as.factor(Type))
detach(Hostile)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

