% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Hostile}
\alias{Hostile}
\title{Hostility levels of high school students from rural, suburban, and urban
areas}
\format{A data frame/tibble with 135 observations on two variables
\describe{ 
\item{location}{a factor with the location of the high school student (\code{Rural}, \code{Suburban}, or \code{Urban})} 
\item{hostility}{the score from the Hostility Level Test} 
}}
\usage{
Hostile
}
\description{
Data for Example 10.6
}
\examples{

boxplot(hostility ~ location, data = Hostile, 
        col = c("red", "blue", "green"))
kruskal.test(hostility ~ location, data = Hostile)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
