% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Iceberg}
\alias{Iceberg}
\title{Number of icebergs sighted each month south of Newfoundland and south of the
Grand Banks in 1920}
\format{A data frame with 12 observations on the following 3 variables.
\describe{
 \item{Month}{a factor with levels \code{Apr} \code{Aug}
\code{Dec} \code{Feb} \code{Jan} \code{Jul} \code{Jun} \code{Mar} \code{May}
\code{Nov} \code{Oct} \code{Sep}} 
\item{Newfound}{a numeric vector}
\item{GrandBk}{a numeric vector} 
}}
\description{
Data for Exercise 2.46 and 2.60
}
\examples{

str(Iceberg)
attach(Iceberg)
plot(GrandBk,Newfound)
abline(lm(Newfound~GrandBk))
detach(Iceberg)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

