% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Iceberg}
\alias{Iceberg}
\title{Number of icebergs sighted each month south of Newfoundland and south of the
Grand Banks in 1920}
\format{A data frame with 12 observations on the following 3 variables.
\describe{
 \item{month}{a character variable with abbreviated months of the year} 
\item{Newfoundland}{number of icebergs sighted south of Newfoundland}
\item{Grand Banks}{number of icebergs sighted south of Grand Banks} 
}}
\usage{
Iceberg
}
\description{
Data for Exercise 2.46 and 2.60
}
\examples{

plot(Newfoundland ~ `Grand Banks`, data = Iceberg)
abline(lm(Newfoundland ~ `Grand Banks`, data = Iceberg))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
