% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Iceberg}
\alias{Iceberg}
\title{Number of icebergs sighted each month south of Newfoundland and south of the
Grand Banks in 1920}
\format{A data frame with 12 observations on the following 3 variables.
\describe{
 \item{month}{a character variable with abbreviated months of the year} 
\item{Newfoundland}{number of icebergs sighted south of Newfoundland}
\item{Grand Banks}{number of icebergs sighted south of Grand Banks} 
}}
\source{
N. Shaw, \emph{Manual of Meteorology}, Vol. 2 (London: Cambridge University Press 1942),
7; and F. Mosteller and J. Tukey, \emph{Data Analysis and Regression} (Reading, MA: Addison - Wesley, 1977)
}
\usage{
Iceberg
}
\description{
Data for Exercise 2.46 and 2.60
}
\examples{

plot(Newfoundland ~ `Grand Banks`, data = Iceberg)
abline(lm(Newfoundland ~ `Grand Banks`, data = Iceberg))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
