% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Income}
\alias{Income}
\title{Percent change in personal income from 1st to 2nd quarter in 2000}
\format{A data frame with 51 observations on the following 6 variables.
\describe{ 
\item{State}{a factor with levels \code{Alabama}
\code{Alaska} \code{Arizona} \code{Arkansas} \code{California}
\code{Colorado} \code{Connecticut} \code{Delaware} \code{District of
Colunbia} \code{Florida} \code{Georgia} \code{Hawaii} \code{Idaho}
\code{Illinois} \code{Indiana} \code{Iowa} \code{Kansas} \code{Kentucky}
\code{Louisiana} \code{Maine} \code{Maryland} \code{Massachusetts}
\code{Michigan} \code{Minnesota} \code{Mississippi} \code{Missour}
\code{Montana} \code{Nebraska} \code{Nevada} \code{New Hampshire} \code{New
Jersey} \code{New Mexico} \code{New York} \code{North Carolina} \code{North
Dakota} \code{Ohio} \code{Oklahoma} \code{Oregon} \code{Pennsylvania}
\code{Rhode Island} \code{South Carolina} \code{South Dakota}
\code{Tennessee} \code{Texas} \code{Utah} \code{Vermont} \code{Virginia}
\code{Washington} \code{West Virginia} \code{Wisconsin} \code{Wyoming}}
\item{income}{a numeric vector} 
\item{C3}{a numeric vector}
\item{Class}{a numeric vector} 
\item{freq}{a numeric vector}
\item{percent}{a numeric vector} 
}}
\description{
Data for Exercise 1.33
}
\examples{

str(Income)
attach(Income)
CATS <-factor(cut(income,breaks=c(0.5,1.0,1.5,2,max(income)) ))    
table(CATS)
table(CATS)/length(income)
barplot(table(CATS),col="lightblue",main="Problem 1.33")
detach(Income)
remove(CATS)      

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

