% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Independent}
\alias{Independent}
\title{Illustrates a comparison problem for long-tailed distributions}
\format{A data frame with 46 observations on the following 3 variables.
\describe{ 
\item{score}{a numeric vector} 
\item{group}{a numeric vector} 
\item{ranks}{a numeric vector} 
}}
\description{
Data for Exercise 7.41
}
\examples{

str(Independent)
attach(Independent)
boxplot(score~group)
wilcox.test(score~group)
detach(Independent)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

