% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Independent}
\alias{Independent}
\title{Illustrates a comparison problem for long-tailed distributions}
\format{A data frame/tibble with 46 observations on two variables
\describe{ 
\item{score}{a numeric vector} 
\item{group}{a factor with levels \code{A} and \code{B}} 
}}
\usage{
Independent
}
\description{
Data for Exercise 7.41
}
\examples{

qqnorm(Independent$score[Independent$group=="A"])
qqline(Independent$score[Independent$group=="A"])
qqnorm(Independent$score[Independent$group=="B"])
qqline(Independent$score[Independent$group=="B"])
boxplot(score ~ group, data = Independent)
wilcox.test(score ~ group, data = Independent)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
