% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Indian}
\alias{Indian}
\title{Educational attainment versus per capita income and poverty rate for
American indians living on reservations}
\format{A data frame with 10 observations on the following 4 variables.
\describe{ 
\item{Reserv}{a factor with levels \code{Blackfeet}
\code{Fort Apache} \code{Gila River} \code{Hopi} \code{Navajo} \code{Papago}
\code{Pine Ridge} \code{Rosebud} \code{San Carlos} \code{Zuni Pueblo}}
\item{highsch}{a numeric vector} 
\item{income}{a numeric vector} 
\item{poverty}{a numeric vector} 
}}
\description{
Data for Exercise 2.95
}
\examples{

str(Indian)
attach(Indian)
par(mfrow=c(1,2))
plot(highsch,income,xlab="Percent High School Graduates", ylab="Per capita income")
plot(highsch,poverty,xlab="Percent High School Graduates", ylab="Poverty rate")
par(mfrow=c(1,1))
cor(cbind(highsch,income,poverty))
detach(Indian)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

