% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Indiapol}
\alias{Indiapol}
\title{Average miles per hour for the winners of the Indianapolis 500 race}
\format{A data frame with 39 observations on the following 3 variables.
\describe{ 
\item{year}{a numeric vector} 
\item{speed}{a numeric vector} 
\item{yr.1960}{a numeric vector} 
}}
\description{
Data for Exercise 1.128
}
\examples{

str(Indiapol)
attach(Indiapol)
plot(year,speed,type="l")
detach(Indiapol)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

