% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Indiapol}
\alias{Indiapol}
\title{Average miles per hour for the winners of the Indianapolis 500 race}
\format{
A data frame/tibble with 39 observations on two variables
\describe{ 
\item{year}{the year of the race} 
\item{speed}{the winners average speed (in mph)} 
}
}
\source{
The World Almanac and Book of Facts, 2000, p. 1004.
}
\usage{
Indiapol
}
\description{
Data for Exercise 1.128
}
\examples{

plot(speed ~ year, data = Indiapol, type = "b")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
