% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Inflatio}
\alias{Inflatio}
\title{Private pay increase of salaried employees versus inflation rate}
\format{A data frame with 24 observations on the following 5 variables.
\describe{
\item{year}{a numeric vector} 
\item{pay}{a numeric vector} 
\item{increase}{a numeric vector}
\item{inflation}{a numeric vector} 
\item{C6.T}{a factor with levels \code{alow} \code{bmiddle} \code{high}} 
}}
\description{
Data for Exercises 2.12 and 2.29
}
\examples{

str(Inflatio)
attach(Inflatio)
plot(inflation,increase)
cor(inflation,increase,use="complete.obs")
detach(Inflatio)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

