% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Inletoil}
\alias{Inletoil}
\title{Inlet oil temperature through a valve}
\format{A data frame/tibble with 12 observations on the following variable.
\describe{ 
\item{temp}{inlet oil temperature (Fahrenheit)} 
}}
\usage{
Inletoil
}
\description{
Data for Exercises 5.91 and 6.48
}
\examples{

hist(Inletoil$temp, breaks = 3)
qqnorm(Inletoil$temp)
qqline(Inletoil$temp)
t.test(Inletoil$temp)
t.test(Inletoil$temp, mu = 98, alternative = "less")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
