% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Inletoil}
\alias{Inletoil}
\title{Inlet oil temperature through a valve}
\format{A data frame with 12 observations on the following variable.
\describe{ 
\item{temp}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercises 5.91 and 6.48
}
\examples{

str(Inletoil)
attach(Inletoil)
t.test(temp)$conf
t.test(temp,mu=98,alternative="less")
detach(Inletoil)

}
\keyword{datasets}

