% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Inmate}
\alias{Inmate}
\title{Type of drug offense by race}
\format{A data frame/tibble with 28,037 observations on the following 2 variables.
\describe{ 
\item{race}{a factor with levels \code{white},
\code{black}, and \code{hispanic}} 
\item{drug}{a factor with levels \code{heroin}, \code{crack}, \code{cocaine}, and \code{marijuana}}
}}
\usage{
Inmate
}
\description{
Data for Statistical Insight Chapter 8
}
\examples{

T1 <- xtabs(~race + drug, data = Inmate)
T1
chisq.test(T1)
rm(T1)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
