% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Inmate}
\alias{Inmate}
\title{Type of drug offense by race}
\format{A data frame with 3 observations on the following 5 variables.
\describe{ 
\item{Race}{a factor with levels \code{black}
\code{hispanic} \code{white}} 
\item{heroin}{a numeric vector}
\item{crack}{a numeric vector} 
\item{cocaine}{a numeric vector} 
\item{marijuan}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Statistical Insight Chapter 8
}
\examples{

str(Inmate)
attach(Inmate)
Inmate
chisq.test(Inmate[,2:5])
detach(Inmate)

}
\keyword{datasets}

