% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Inspect}
\alias{Inspect}
\title{Percent of vehicles passing inspection by type inspection station}
\format{A data frame with 6 observations on the following 4 variables.
\describe{ 
\item{Type}{a factor with levels \code{auto inspection}
\code{auto repair} \code{car care center} \code{gas station} \code{new car
dealer} \code{tire store}} 
\item{less70}{a numeric vector}
\item{X70.85}{a numeric vector} 
\item{great85}{a numeric vector} 
}}
\description{
Data for Exercise 5.89
}
\examples{

str(Inspect)
attach(Inspect)
Inspect
chisq.test(Inspect[,2:4])
detach(Inspect)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
