% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Insulate}
\alias{Insulate}
\title{Heat loss through a new insulating medium}
\format{A data frame with 10 observations on the following 2 variables.
\describe{ 
\item{temp}{a numeric vector} 
\item{loss}{a numeric vector} 
}}
\description{
Data for Exercise 9.50
}
\examples{

str(Insulate)
attach(Insulate)
summary(lm(loss~temp))
detach(Insulate)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

