% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Iqgpa}
\alias{Iqgpa}
\title{GPA versus IQ for 12 individuals}
\format{A data frame with 12 observations on the following 2 variables.
\describe{ 
\item{IQ}{a numeric vector} 
\item{GPA}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercises 9.51 and 9.52
}
\examples{

str(Iqgpa)
attach(Iqgpa)
plot(IQ,GPA)
model <- lm(GPA~IQ)
abline(model)
summary(model)
detach(Iqgpa)
remove(model)

}
\keyword{datasets}

