% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Iqgpa}
\alias{Iqgpa}
\title{GPA versus IQ for 12 individuals}
\format{A data frame/tibble with 12 observations on the following 2 variables.
\describe{ 
\item{iq}{IQ scores} 
\item{gpa}{Grade point average} 
}}
\usage{
Iqgpa
}
\description{
Data for Exercises 9.51 and 9.52
}
\examples{

plot(gpa ~ iq, data = Iqgpa, col = "blue", pch = 19)
model <- lm(gpa ~ iq, data = Iqgpa)
summary(model)
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
