% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Irises}
\alias{Irises}
\title{R.A. Fishers famous data on sepal length of a species of Iris Setosa}
\format{A data frame with 150 observations on the following 14 variables.
\describe{ 
\item{sepalL1}{a numeric vector} 
\item{sepalW1}{a numeric vector} 
\item{petalL1}{a numeric vector}
\item{petalW1}{a numeric vector} 
\item{sepalL2}{a numeric vector} 
\item{sepalW2}{a numeric vector} 
\item{petalL2}{a numeric vector} 
\item{peatalW2}{a numeric vector}
\item{sepalL3}{a numeric vector} 
\item{sepalW3}{a numeric vector} 
\item{petalL3}{a numeric vector} 
\item{petalW3}{a numeric vector} 
\item{sepalL}{a numeric vector}
\item{sample}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercises 1.15 and 5.19
}
\examples{

str(Irises)
attach(Irises)
EDA(sepalL1)
t.test(sepalL1,conf.level=.99)$conf
detach(Irises)

}
\keyword{datasets}

