% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Jdpower}
\alias{Jdpower}
\title{Number of problems reported per 100 cars in 1994 versus 1995s}
\format{A data frame/tibble with 29 observations on three variables
\describe{ 
\item{car}{a factor with levels \code{Acura}, \code{BMW},
\code{Buick}, \code{Cadillac}, \code{Chevrolet}, \code{Dodge} \code{Eagle},
\code{Ford}, \code{Geo}, \code{Honda}, \code{Hyundai}, \code{Infiniti},
\code{Jaguar}, \code{Lexus}, \code{Lincoln}, \code{Mazda}, \code{Mercedes-Benz},
\code{Mercury}, \code{Mitsubishi}, \code{Nissan}, \code{Oldsmobile},
\code{Plymouth}, \code{Pontiac}, \code{Saab}, \code{Saturn}, and \code{Subaru},
\code{Toyota} \code{Volkswagen}, \code{Volvo}} 
\item{1994}{number of problems per 100 cars in 1994} 
\item{1995}{number of problems per 100 cars in 1995} 
}}
\source{
\emph{USA Today}, May 25, 1995.
}
\usage{
Jdpower
}
\description{
Data for Exercise 2.14, 2.17, 2.31, 2.33, and 2.40
}
\examples{

model <- lm(`1995` ~ `1994`, data = Jdpower)
summary(model)
plot(`1995` ~ `1994`, data = Jdpower)
abline(model, col = "red")
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
