% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Jdpower}
\alias{Jdpower}
\title{Number of problems reported per 100 cars in 1994 versus 1995s}
\format{A data frame with 29 observations on the following 3 variables.
\describe{ 
\item{Car}{a factor with levels \code{Acura} \code{BMW}
\code{Buick} \code{Cadillac} \code{Chevrolet} \code{Dodge} \code{Eagle}
\code{Ford} \code{Geo} \code{Honda} \code{Hyundai} \code{Infiniti}
\code{Jaguar} \code{Lexus} \code{Lincoln} \code{Mazda} \code{Mercedes-Benz}
\code{Mercury} \code{Mitsubishi} \code{Nissan} \code{Oldsmobile}
\code{Plymouth} \code{Pontiac} \code{Saab} \code{Saturn} \code{Subaru}
\code{Toyota} \code{Volkswagen} \code{Volvo}} 
\item{X1994}{a numeric vector} 
\item{X1995}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 2.14, 2.17, 2.31, 2.33, and 2.40
}
\examples{

str(Jdpower)
attach(Jdpower)
plot(X1994,X1995)
model <- lm(X1995~X1994)
abline(model)
model
cor(X1995,X1994)
detach(Jdpower)

}
\keyword{datasets}

