% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Jobsat}
\alias{Jobsat}
\title{Job satisfaction and stress level for 9 school teachers}
\format{A data frame/tibble with nine observations on two variables
\describe{ 
\item{wspt}{Wilson Stress Profile score for teachers} 
\item{satisfaction}{job satisfaction score} 
}}
\usage{
Jobsat
}
\description{
Data for Exercise 9.60
}
\examples{

plot(satisfaction ~ wspt, data = Jobsat)
model <- lm(satisfaction ~ wspt, data = Jobsat)
abline(model)
summary(model)
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
