% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Kidsmoke}
\alias{Kidsmoke}
\title{Smoking habits of boys and girls ages 12 to 18}
\format{A data frame/tibble with 1000 observations on two variables
\describe{ 
\item{gender}{character vector with values \code{female} and \code{male}} 
\item{smoke}{a character vector with values \code{no} and \code{yes}} 
}}
\usage{
Kidsmoke
}
\description{
Data for Exercise 4.85
}
\examples{

T1 <- xtabs(~smoke + gender, data = Kidsmoke)
T1
prop.table(T1)
prop.table(T1, 1)
prop.table(T1, 2)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
