% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Kidsmoke}
\alias{Kidsmoke}
\title{Smoking habits of boys and girls ages 12 to 18}
\format{A data frame with 1000 observations on the following 2 variables.
\describe{ 
\item{gender}{a numeric vector} 
\item{smoke}{a numeric vector} 
}}
\description{
Data for Exercise 4.85
}
\examples{

str(Kidsmoke)
attach(Kidsmoke)
table(gender,smoke)
addmargins(table(gender,smoke))
addmargins(table(gender,smoke)/1000)
detach(Kidsmoke)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
