% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Kilowatt}
\alias{Kilowatt}
\title{Rates per kilowatt-hour for each of the 50 states and DC}
\format{A data frame/tibble with 51 observations on the following 2 variables.
\describe{ 
\item{state}{a factor with levels \code{Alabama}
\code{Alaska}, \code{Arizona}, \code{Arkansas} \code{California},
\code{Colorado}, \code{Connecticut}, \code{Delaware}, \code{District of
Columbia}, \code{Florida},\code{Georgia}, \code{Hawaii}, \code{Idaho},
\code{Illinois}, \code{Indiana}, \code{Iowa} \code{Kansas} \code{Kentucky},
\code{Louisiana}, \code{Maine}, \code{Maryland}, \code{Massachusetts},
\code{Michigan}, \code{Minnesota}, \code{Mississippi}, \code{Missour},
\code{Montana} \code{Nebraska}, \code{Nevada}, \code{New Hampshire}, \code{New
Jersey}, \code{New Mexico}, \code{New York}, \code{North Carolina}, \code{North
Dakota}, \code{Ohio}, \code{Oklahoma}, \code{Oregon}, \code{Pennsylvania},
\code{Rhode Island}, \code{South Carolina}, \code{South Dakota},
\code{Tennessee}, \code{Texas}, \code{Utah}, \code{Vermont}, \code{Virginia}
\code{Washington}, \code{West Virginia}, \code{Wisconsin}, and \code{Wyoming}}
\item{rate}{a numeric vector indicating rates for kilowatt per hour} 
}}
\usage{
Kilowatt
}
\description{
Data for Example 5.9
}
\examples{

BSDA::EDA(Kilowatt$rate)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
