% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Kinder}
\alias{Kinder}
\title{Reading scores for first grade children who attended kindergarten versus
those who did not}
\format{A data frame with 8 observations on the following 3 variables.
\describe{
\item{Pair}{a numeric vector}
\item{Kinder}{a numeric vector}
\item{NoKinder}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.68
}
\examples{
str(Kinder)
attach(Kinder)
DIF <- Kinder - NoKinder
qqnorm(DIF)
qqline(DIF)
shapiro.test(DIF)
t.test(Kinder, NoKinder,paired=TRUE,alternative="greater")
detach(Kinder)
remove(DIF)
}
\keyword{datasets}

