% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Laminect}
\alias{Laminect}
\title{Median costs of laminectomies at hospitals across North Carolina in 1992}
\format{A data frame with 46 observations on the following 5 variables.
\describe{
 \item{cost}{a numeric vector}
 \item{class}{a numeric vector}
 \item{Rural}{a numeric vector}
 \item{Regional}{a numeric vector}
 \item{Metropol}{a numeric vector}
 }}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 10.18
}
\examples{
str(Laminect)
attach(Laminect)
boxplot(cost~class)
anova(lm(cost~as.factor(class)))
detach(Laminect)
}
\keyword{datasets}

