% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Lead}
\alias{Lead}
\title{Lead levels in children's blood whose parents worked in a battery factory}
\format{A data frame with 33 observations on the following 3 variables.
\describe{ 
\item{Pair}{a numeric vector} 
\item{exposed}{a numeric vector} 
\item{control}{a numeric vector} 
}}
\description{
Data for Example 1.17
}
\examples{

str(Lead)
attach(Lead)
boxplot(exposed,control, names=c("Exposed","Control"),col=c("red","blue"))
detach(Lead)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

