% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Lead}
\alias{Lead}
\title{Lead levels in children's blood whose parents worked in a battery factory}
\format{A data frame/tibble with 66 observations on the following 2 variables.
\describe{ 
\item{group}{a character vector with values \code{exposed} and \code{control}} 
\item{lead}{a numeric vector indicating the level of lead in children's blood in micrograms/dl} 
}}
\source{
Morton, D. et al. (1982), "Lead Absorption in Children of Employees in a Lead-Related
Industry," \emph{American Journal of Epidemiology, 155,} 549-555.
}
\usage{
Lead
}
\description{
Data for Example 1.17
}
\examples{

boxplot(lead ~ group, data = Lead)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
