% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Leader}
\alias{Leader}
\title{Leadership exam scores by age for employees on an industrial plant}
\format{A data frame/tibble with 34 observations on two variables
\describe{ 
\item{age}{a character vector indicating age with values \code{under35} and \code{over35}} 
\item{score}{score on a leadership exam} 
}}
\usage{
Leader
}
\description{
Data for Exercise 7.31
}
\examples{

boxplot(score ~ age, data = Leader, col = c("gray", "green"))
t.test(score ~ age, data = Leader)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
