% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Leader}
\alias{Leader}
\title{Leadership exam scores by age for employees on an industrial plant}
\format{A data frame with 34 observations on the following 2 variables.
\describe{ 
\item{under35}{a numeric vector} 
\item{over35}{a numeric vector} 
}}
\description{
Data for Exercise 7.31
}
\examples{

str(Leader)
attach(Leader)
boxplot(under35,over35,names=c("Under 35","Over 35"),col=c("green","brown"))
t.test(under35,over35)
detach(Leader)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
