% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Lethal}
\alias{Lethal}
\title{Survival time of mice injected with an experimental lethal drug}
\format{A data frame with 30 observations on the following variable.
\describe{ 
\item{survival}{a numeric vector} 
}}
\description{
Data for Example 6.12
}
\examples{

str(Lethal)
attach(Lethal)
SIGN.test(survival,md=45,alternative="less")
detach(Lethal)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

