% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Lethal}
\alias{Lethal}
\title{Survival time of mice injected with an experimental lethal drug}
\format{A data frame with 30 observations on the following variable.
\describe{
\item{survival}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Example 6.12
}
\examples{
str(Lethal)
attach(Lethal)
SIGN.test(survival,md=45,alternative="less")
detach(Lethal)
}
\keyword{datasets}

