% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Lethal}
\alias{Lethal}
\title{Survival time of mice injected with an experimental lethal drug}
\format{
A data frame/tibble with 30 observations on one variable
\describe{ 
\item{survival}{a numeric vector indicating time surivived 
after injection (in seconds)} 
}
}
\usage{
Lethal
}
\description{
Data for Example 6.12
}
\examples{

SIGN.test(Lethal$survival, md = 45, alternative = "less")


}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
