% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Life}
\alias{Life}
\title{Life expectancy of men and women in U.S.}
\format{A data frame with 8 observations on the following 3 variables.
\describe{ 
\item{year}{a numeric vector} 
\item{Men}{a numeric vector} 
\item{Women}{a numeric vector} 
}}
\description{
Data for Exercise 1.31
}
\examples{

str(Life)
attach(Life)
plot(year,Men,type="l",ylim=c(min(Men,Women),max(Men,Women)),col="blue",
main="Life Expectancy versus Year",ylab="Age",xlab="Year")
lines(year,Women,col="red")
text(1955,65,"Men",col="blue")
text(1955,70,"Women",col="red")
detach(Life)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

