% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Lifespan}
\alias{Lifespan}
\title{Life span of electronic components used in a spacecraft versus heat}
\format{
A data frame/tibble with six observations two variables
\describe{ 
\item{heat}{temperature (in Celcius)} 
\item{life}{lifespan of component (in hours)} 
}
}
\usage{
Lifespan
}
\description{
Data for Exercise 2.4, 2.37, and 2.49
}
\examples{

plot(life ~ heat, data = Lifespan)
model <- lm(life ~ heat, data = Lifespan)
abline(model, col = "red")
resid(model)
sum((resid(model))^2)
anova(model)
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
