% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Lifespan}
\alias{Lifespan}
\title{Life span of electronic components used in a spacecraft versus heat}
\format{A data frame/tibble with 6 observations on the following 2 variables.
\describe{ 
\item{heat}{temperature in degrees C} 
\item{life}{lifespan in hours} 
}}
\usage{
Lifespan
}
\description{
Data for Exercise 2.4, 2.37, and 2.49
}
\examples{

plot(life ~ heat, data = Lifespan)
model <- lm(life ~ heat, data = Lifespan)
abline(model)
resid(model)
sum((resid(model))^2)
anova(model)
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
