% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Lifespan}
\alias{Lifespan}
\title{Life span of electronic components used in a spacecraft versus heat}
\format{A data frame with 6 observations on the following 4 variables.
\describe{ 
\item{heat}{a numeric vector} 
\item{life}{a numeric vector} 
\item{RESI1}{a numeric vector}
\item{FITS1}{a numeric vector} 
}}
\description{
Data for Exercise 2.4, 2.37, and 2.49
}
\examples{

str(Lifespan)
attach(Lifespan)
plot(heat,life)
model <- lm(life~heat)
model
resid(model)
sum((resid(model))^2)
anova(model)
# plot(model)  # Used for diagnostic purposes
detach(Lifespan)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

