% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Lodge}
\alias{Lodge}
\title{Measured traffic at three prospective locations for a motor lodge}
\format{A data frame/tibble with 45 observations on six variables
\describe{ 
\item{traffic}{a numeric vector indicating the amount of vehicles that passed a site in 1 hour} 
\item{site}{a numeric vector with values \code{1}, \code{2}, and \code{3}} 
\item{ranks}{ranks for variable \code{traffic}}
}}
\usage{
Lodge
}
\description{
Data for Exercise 10.33
}
\examples{

boxplot(traffic ~ site, data = Lodge, col = cm.colors(3))
anova(lm(traffic ~ factor(site), data = Lodge))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
