% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Lodge}
\alias{Lodge}
\title{Measured traffic at three prospective locations for a motor lodge}
\format{A data frame with 45 observations on the following 6 variables.
\describe{ 
\item{SiteA}{a numeric vector} 
\item{SiteB}{a numeric vector} 
\item{SiteC}{a numeric vector}
\item{Traffic}{a numeric vector} 
\item{Site}{a numeric vector} 
\item{Ranks}{a numeric vector} 
}}
\description{
Data for Exercise 10.33
}
\examples{

str(Lodge)
attach(Lodge)
boxplot(Traffic~Site)
anova(lm(Traffic~as.factor(Site)))
detach(Lodge)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
