% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Longtail}
\alias{Longtail}
\title{Long-tailed distributions to illustrate Kruskal Wallis test}
\format{
A data frame/tibble with 60 observations on three variables
\describe{ 
\item{score}{a numeric vector} 
\item{group}{a numeric vector with values \code{1}, \code{2}, and \code{3}} 
\item{ranks}{ranks for variable \code{score}} 
}
}
\usage{
Longtail
}
\description{
Data for Exercise 10.45
}
\examples{

boxplot(score ~ group, data = Longtail, col = heat.colors(3))
kruskal.test(score ~ factor(group), data = Longtail)
anova(lm(score ~ factor(group), data = Longtail))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
