% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Longtail}
\alias{Longtail}
\title{Long-tailed distributions to illustrate Kruskal Wallis test}
\format{A data frame with 60 observations on the following 6 variables.
\describe{
\item{GroupA}{a numeric vector}
\item{GroupB}{a numeric vector}
\item{GroupC}{a numeric vector}
\item{score}{a numeric vector}
\item{Group}{a numeric vector}
\item{Ranks}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 10.45
}
\examples{
str(Longtail)
attach(Longtail)
boxplot(score~Group)
kruskal.test(score~as.factor(Group))
anova(lm(score~as.factor(Group)))
detach(Longtail)
}
\keyword{datasets}

