% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Lowabil}
\alias{Lowabil}
\title{Reading skills of 24 matched low ability students}
\format{A data frame with 12 observations on the following 3 variables.
\describe{ 
\item{Pair}{a numeric vector} 
\item{Experimt}{a numeric vector} 
\item{Control}{a numeric vector} 
}}
\description{
Data for Example 7.18
}
\examples{

str(Lowabil)
attach(Lowabil)
DIF <- Experimt - Control
qqnorm(DIF)
qqline(DIF)
shapiro.test(DIF)
t.test(Experimt,Control,paired=TRUE)
detach(Lowabil)
remove(DIF)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
