% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Lowabil}
\alias{Lowabil}
\title{Reading skills of 24 matched low ability students}
\format{A data frame/tibble with 12 observations on three variables
\describe{ 
\item{pair}{a numeric indicator of pair} 
\item{experiment}{score of the  child with the experimental method} 
\item{control}{score of the child with the standard method} 
}}
\usage{
Lowabil
}
\description{
Data for Example 7.18
}
\examples{

diff = Lowabil$experiment - Lowabil$control
qqnorm(diff)
qqline(diff)
shapiro.test(diff)
t.test(Lowabil$experiment, Lowabil$control, paired = TRUE)
# OR
t.test(diff)
rm(diff)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
