% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Magnesiu}
\alias{Magnesiu}
\title{Magnesium concentration and distances between samples}
\format{A data frame with 20 observations on the following 4 variables.
\describe{
\item{distance}{a numeric vector}
\item{magnesiu}{a numeric vector}
\item{SRES1}{a numeric vector}
\item{FITS1}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 9.9
}
\examples{
str(Magnesiu)
attach(Magnesiu)
model <- lm(magnesiu~distance)
plot(distance,magnesiu)
abline(model)
summary(model)
detach(Magnesiu)
remove(model)
}
\keyword{datasets}

