% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Malpract}
\alias{Malpract}
\title{Amounts awarded in 17 malpractice cases}
\format{A data frame with 17 observations on the following variable.
\describe{ 
\item{award}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 5.73
}
\examples{

str(Malpract)
attach(Malpract)
SIGN.test(award,conf.level=.90)
detach(Malpract)

}
\keyword{datasets}

