% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Malpract}
\alias{Malpract}
\title{Amounts awarded in 17 malpractice cases}
\format{A data frame/tibble with 17 observations on one variable
\describe{ 
\item{award}{malpractice reward (in $1000)} 
}}
\usage{
Malpract
}
\description{
Data for Exercise 5.73
}
\examples{

SIGN.test(Malpract$award, conf.level = 0.90)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
