% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Manager}
\alias{Manager}
\title{Advertised salaries offered general managers of major corporations in 1995}
\format{A data frame/tibble with 26 observations on the following variable.
\describe{ 
\item{salary}{random sample of advertised salaries of top executives} 
}}
\usage{
Manager
}
\description{
Data for Exercise 5.81
}
\examples{

stem(Manager$salary)
BSDA::SIGN.test(Manager$salary)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
