% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Manager}
\alias{Manager}
\title{Advertised salaries offered general managers of major corporations in 1995}
\format{A data frame with 26 observations on the following variable.
\describe{
\item{salary}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 5.81
}
\examples{
str(Manager)
attach(Manager)
stem(salary)
SIGN.test(salary)
detach(Manager)
}
\keyword{datasets}

