% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Marked}
\alias{Marked}
\title{Percent of marked cars in 65 police departments in Florida}
\format{A data frame with 65 observations on the following variable.
\describe{
\item{percent}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 6.100
}
\examples{
str(Marked)
attach(Marked)
EDA(percent)
t.test(percent,mu=60,alternative="greater")
SIGN.test(percent,md=60,alternative="greater")
detach(Marked)
}
\keyword{datasets}

