% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Marked}
\alias{Marked}
\title{Percent of marked cars in 65 police departments in Florida}
\format{
A data frame/tibble with 65 observations on one variable
\describe{ 
\item{percent}{percentage of marked cars in 65 Florida police departments}
}
}
\source{
\emph{Law Enforcement Management and Administrative Statistics, 1993}, Bureau of 
Justice Statistics, NCJ-148825, September 1995, p. 147-148.
}
\usage{
Marked
}
\description{
Data for Exercise 6.100
}
\examples{

EDA(Marked$percent)
SIGN.test(Marked$percent, md = 60, alternative = "greater")
t.test(Marked$percent, mu = 60, alternative = "greater")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
