% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{MATH}
\alias{MATH}
\title{Standardized math test scores for 30 students}
\format{A data frame with 30 observations on the following variable.
\describe{ 
\item{math}{a numeric vector} 
}}
\description{
Data for Exercise 1.69
}
\examples{

str(MATH)
attach(MATH)
hist(math,col="pink")
CharlieZ <- (62-mean(math))/sd(math)
CharlieZ
detach(MATH)
remove(CharlieZ)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

