% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Math}
\alias{Math}
\title{Standardized math test scores for 30 students}
\format{A data frame/tibble with 30 observations on the following variable.
\describe{ 
\item{score}{scores on a standardized test for 30 tenth graders} 
}}
\usage{
Math
}
\description{
Data for Exercise 1.69
}
\examples{

stem(Math$score)
hist(Math$score, main = "Math Scores", xlab = "score", freq = FALSE)
lines(density(Math$score), col = "red")
CharlieZ <- (62 - mean(Math$score))/sd(Math$score)
CharlieZ
scale(Math$score)[which(Math$score == 62)]

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
