% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Mathcomp}
\alias{Mathcomp}
\title{Standardized math competency for a group of entering freshmen at a small
community college}
\format{A data frame with 31 observations on the following variable.
\describe{ 
\item{score}{a numeric vector} 
}}
\description{
Data for Exercise 5.26
}
\examples{

str(Mathcomp)
attach(Mathcomp)
stem(score)
EDA(score)
detach(Mathcomp)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

