% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Mathcomp}
\alias{Mathcomp}
\title{Standardized math competency for a group of entering freshmen at a small
community college}
\format{A data frame/tibble with 31 observations one variable
\describe{ 
\item{score}{scores of 31 entering freshmen at a community college on a national standardized test} 
}}
\usage{
Mathcomp
}
\description{
Data for Exercise 5.26
}
\examples{

stem(Mathcomp$score)
BSDA::EDA(Mathcomp$score)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
