% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Maze}
\alias{Maze}
\title{Error scores for four groups of experimental animals running a maze}
\format{A data frame with 32 observations on the following 6 variables.
\describe{
\item{CondA}{a numeric vector}
\item{CondB}{a numeric vector}
\item{CondC}{a numeric vector}
\item{CondD}{a numeric vector}
\item{score}{a numeric vector}
\item{condition}{a factor with levels \code{CondA}
\code{CondB} \code{CondC} \code{CondD}}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 10.13
}
\examples{
str(Maze)
attach(Maze)
boxplot(score~condition)
anova(lm(score~condition))
detach(Maze)
}
\keyword{datasets}

