% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Maze}
\alias{Maze}
\title{Error scores for four groups of experimental animals running a maze}
\format{A data frame/tibble with 32 observations on the following 2 variables.
\describe{ 
\item{score}{error scores for animals running through a maze under different conditions} 
\item{condition}{a factor with levels \code{CondA},
\code{CondB,} \code{CondC}, and \code{CondD}} 
}}
\usage{
Maze
}
\description{
Data for Exercise 10.13
}
\examples{

boxplot(score ~ condition, data = Maze, col = rainbow(4))
anova(lm(score ~ condition, data = Maze))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
