% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Median}
\alias{Median}
\title{Illustrates test of equality of medians with the Kruskal Wallis test}
\format{A data frame with 15 observations on the following 3 variables.
\describe{ 
\item{Sample1}{a numeric vector} 
\item{Sample2}{a numeric vector} 
\item{Sample3}{a numeric vector} 
}}
\description{
Data for Exercise 10.52
}
\examples{

str(Median)
attach(Median)
STACKED <-stack(Median)
STACKED[1:5,]
boxplot(values~ind,col=c("red","blue","green"),data=STACKED)
anova(lm(values~ind,data=STACKED))
kruskal.test(values~ind,data=STACKED)
remove(STACKED)
detach(Median)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

