% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Mental}
\alias{Mental}
\title{Median mental ages of 16 girls}
\format{A data frame with 16 observations on the following variable.
\describe{ 
\item{age}{a numeric vector} 
}}
\description{
Data for Exercise 6.52
}
\examples{

str(Mental)
attach(Mental)
SIGN.test(age,md=100)
detach(Mental)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

