% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Mental}
\alias{Mental}
\title{Median mental ages of 16 girls}
\format{A data frame/tibble with 16 observations on one variable
\describe{ 
\item{age}{mental age of 16 girls} 
}}
\usage{
Mental
}
\description{
Data for Exercise 6.52
}
\examples{

BSDA::SIGN.test(Mental$age, md = 100)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
