% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Mercury}
\alias{Mercury}
\title{Concentration of mercury in 25 lake trout}
\format{A data frame/tibble with 25 observations on one variable
\describe{ 
\item{mercury}{a numeric vector measuring mercury (in parts per million)} 
}}
\usage{
Mercury
}
\description{
Data for Example 1.9
}
\examples{

stem(Mercury$mercury)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
