% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Mercury}
\alias{Mercury}
\title{Concentration of mercury in 25 lake trout}
\format{A data frame with 25 observations on the following variable.
\describe{ 
\item{mercury}{a numeric vector} 
}}
\description{
Data for Example 1.9
}
\examples{

str(Mercury)
attach(Mercury)
stem(mercury)
detach(Mercury)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
