% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Metrent}
\alias{Metrent}
\title{Monthly rental costs in metro areas with 1 million or more persons}
\format{A data frame/tibble with 46 observations on one variable
\describe{ 
\item{rent}{monthly rent in dollars} 
}}
\source{
U.S. Bureau of the Census, \emph{Housing in the Metropolitan Areas, Statistical Brief} SB/94/19,
September 1994
}
\usage{
Metrent
}
\description{
Data for Exercise 5.117
}
\examples{

boxplot(Metrent$rent)
t.test(Metrent$rent, conf.level = 0.99)$conf

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
