% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Miller}
\alias{Miller}
\title{Miller personality test scores for a group of college students applying for
graduate school}
\format{A data frame with 25 observations on the following variable.
\describe{ 
\item{miller}{a numeric vector} 
}}
\description{
Data for Example 5.7
}
\examples{

str(Miller)
attach(Miller)
stem(miller)
fivenum(miller)
boxplot(miller)
qqnorm(miller,col="blue")
qqline(miller,col="red")
detach(Miller)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

