% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Miller}
\alias{Miller}
\title{Miller personality test scores for a group of college students applying for
graduate school}
\format{A data frame/tibble with 25 observations on one variable
\describe{ 
\item{miller}{scores on the Miller Personality test} 
}}
\usage{
Miller
}
\description{
Data for Example 5.7
}
\examples{

stem(Miller$miller)
fivenum(Miller$miller)
boxplot(Miller$miller)
qqnorm(Miller$miller,col = "blue")
qqline(Miller$miller, col = "red")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
