% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Miller1}
\alias{Miller1}
\title{Twenty scores on the Miller personality test}
\format{A data frame/tibble with 20 observations on one variable
\describe{ 
\item{miller}{scores on the Miller personality test} 
}}
\usage{
Miller1
}
\description{
Data for Exercise 1.41
}
\examples{

stem(Miller1$miller)
stem(Miller1$miller, scale = 2)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
