% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Miller1}
\alias{Miller1}
\title{Twenty scores on the Miller personality test}
\format{A data frame with 20 observations on the following variable.
\describe{ 
\item{miller}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 1.41
}
\examples{

str(Miller1)
attach(Miller1)
stem(miller)
stem(miller,scale=2)
detach(Miller1)

}
\keyword{datasets}

