% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Moisture}
\alias{Moisture}
\title{Moisture content and depth of core sample for marine muds in eastern
Louisiana}
\format{A data frame with 16 observations on the following 4 variables.
\describe{ 
\item{depth}{a numeric vector} 
\item{moisture}{a numeric vector} 
\item{lnmoist}{a numeric vector}
\item{depthsq}{a numeric vector} 
}}
\description{
Data for Exercise 9.37
}
\examples{

str(Moisture)
attach(Moisture)
model <- lm(moisture~depth)
plot(depth,resid(model))
detach(Moisture)
remove(model)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

