% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Monoxide}
\alias{Monoxide}
\title{Carbon monoxide emitted by smoke stacks of a manufacturer and a competitor}
\format{A data frame with 10 observations on the following 2 variables.
\describe{ 
\item{manufac}{a numeric vector} 
\item{compet}{a numeric vector} 
}}
\description{
Data for Exercise 7.45
}
\examples{

str(Monoxide)
attach(Monoxide)
t.test(manufac,compet)
wilcox.test(manufac,compet)
detach(Monoxide)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

