% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Movie}
\alias{Movie}
\title{Moral attitude scale on 15 subjects before and after viewing a movie}
\format{A data frame/tibble with 12 observations on the following 3 variables.
\describe{ 
\item{before}{moral aptitude before viewing the movie} 
\item{after}{moral aptitude after viewing the movie} 
\item{differ}{a numeric vector} 
}}
\usage{
Movie
}
\description{
Data for Exercise 7.53
}
\examples{

qqnorm(Movie$differ)
qqline(Movie$differ)
shapiro.test(Movie$differ)
t.test(Movie$after, Movie$before, paired = TRUE, conf.level = 0.99)
wilcox.test(Movie$after, Movie$before, paired = TRUE)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
