% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Music}
\alias{Music}
\title{Improvement scores for identical twins taught music recognition by two
techniques}
\format{A data frame with 12 observations on the following 3 variables.
\describe{ 
\item{Method1}{a numeric vector} 
\item{Method2}{a numeric vector} 
\item{differ}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.59
}
\examples{

str(Music)
attach(Music)
qqnorm(differ)
qqline(differ)
shapiro.test(differ)
t.test(Method1,Method2,paired=TRUE)
detach(Music)

}
\keyword{datasets}

