% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Nascar}
\alias{Nascar}
\title{Efficiency of pit crews for three major NASCAR teams}
\format{A data frame/tibble with 36 observations on six variables
\describe{ 
\item{time}{duration of pit stop (in seconds)} 
\item{team}{a numeric vector representing team 1, 2, or 3} 
\item{ranks}{a numeric vector ranking each pit stop in order of speed} 
}}
\usage{
Nascar
}
\description{
Data for Exercise 10.53
}
\examples{

boxplot(time ~ team, data = Nascar)
model <- lm(time ~ factor(team), data = Nascar)
summary(model)
anova(model)
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
