% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Nervous}
\alias{Nervous}
\title{Reaction effects of 4 drugs on 25 subjects with a nervous disorder}
\format{A data frame with 25 observations on the following 2 variables.
\describe{ 
\item{react}{a numeric vector} 
\item{drug}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 10.3
}
\examples{

str(Nervous)
attach(Nervous)
boxplot(react~drug)
anova(lm(react~as.factor(drug)))
detach(Nervous)

}
\keyword{datasets}

