% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Newsstand}
\alias{Newsstand}
\title{Daily profits for 20 newsstands}
\format{A data frame/tibble with 20 observations on one variable
\describe{ 
\item{profit}{profit of each newsstand (in dollars)} 
}}
\usage{
Newsstand
}
\description{
Data for Exercise 1.43
}
\examples{

stem(Newsstand$profit)
stem(Newsstand$profit, scale = 3)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
