% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Nfldraf2}
\alias{Nfldraf2}
\title{Rating, time in 40-yard dash, and weight of top defensive linemen in the
1994 NFL draft}
\format{A data frame with 47 observations on the following variable.
\describe{
\item{Rating.forty.weight}{a factor with levels
\code{5.000000000e+000 4.870000000e+000 2.810000000e+002}
\code{5.000000000e+000 4.900000000e+000 2.810000000e+002}
\code{5.100000000e+000 5.100000000e+000 2.980000000e+002}
\code{5.200000000e+000 5.000000000e+000 2.920000000e+002}
\code{5.200000000e+000 5.030000000e+000 2.920000000e+002}
\code{5.300000000e+000 4.780000000e+000 2.720000000e+002}
\code{5.300000000e+000 4.800000000e+000 2.720000000e+002}
\code{5.400000000e+000 4.890000000e+000 2.450000000e+002}
\code{5.400000000e+000 4.900000000e+000 2.450000000e+002}
\code{5.400000000e+000 5.000000000e+000 2.650000000e+002}
\code{5.400000000e+000 5.000000000e+000 2.770000000e+002}
\code{5.400000000e+000 5.030000000e+000 2.770000000e+002}
\code{5.500000000e+000 4.800000000e+000 2.640000000e+002}
\code{5.500000000e+000 4.820000000e+000 2.640000000e+002}
\code{5.600000000e+000 4.900000000e+000 2.470000000e+002}
\code{5.600000000e+000 5.100000000e+000 2.870000000e+002}
\code{5.600000000e+000 5.130000000e+000 2.870000000e+002}
\code{5.700000000e+000 4.880000000e+000 2.830000000e+002}
\code{5.700000000e+000 4.900000000e+000 2.830000000e+002}
\code{5.800000000e+000 4.800000000e+000 2.390000000e+002}
\code{5.800000000e+000 4.850000000e+000 2.390000000e+002}
\code{5.800000000e+000 4.900000000e+000 2.710000000e+002}
\code{5.800000000e+000 4.930000000e+000 2.710000000e+002}
\code{5.900000000e+000 4.700000000e+000 2.600000000e+002}
\code{5.900000000e+000 4.720000000e+000 2.600000000e+002}
\code{6.000000000e+000 5.100000000e+000 2.830000000e+002}
\code{6.200000000e+000 5.100000000e+000 3.120000000e+002}
\code{6.400000000e+000 4.800000000e+000 2.840000000e+002}
\code{6.400000000e+000 4.840000000e+000 2.840000000e+002}
\code{6.400000000e+000 5.070000000e+000 2.850000000e+002}
\code{6.400000000e+000 5.100000000e+000 2.850000000e+002}
\code{6.600000000e+000 5.090000000e+000 3.000000000e+002}
\code{6.600000000e+000 5.100000000e+000 3.000000000e+002}
\code{6.900000000e+000 4.680000000e+000 2.550000000e+002}
\code{6.900000000e+000 4.700000000e+000 2.550000000e+002}
\code{7.000000000e+000 5.060000000e+000 2.890000000e+002}
\code{7.000000000e+000 5.100000000e+000 2.890000000e+002}
\code{7.200000000e+000 4.900000000e+000 2.920000000e+002}
\code{7.500000000e+000 4.900000000e+000 2.760000000e+002}
\code{7.500000000e+000 4.940000000e+000 2.760000000e+002}
\code{8.200000000e+000 4.720000000e+000 3.130000000e+002}}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 9.63
}
\examples{
data(Nfldraf2)
}
\keyword{datasets}

