% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Nfldraft}
\alias{Nfldraft}
\title{Rating, time in 40-yard dash, and weight of top offensive linemen in the
1994 NFL draft}
\format{
A data frame/tibble with 29 observations on three variables
\describe{ 
\item{rating}{rating of each player on a scale out of 10}
\item{forty}{forty yard dash time (in seconds)}
\item{weight}{weight of each player (in pounds)}
}
}
\source{
\emph{USA Today}, April 20, 1994.
}
\usage{
Nfldraft
}
\description{
Data for Exercises 9.10 and 9.16
}
\examples{

plot(rating ~ forty, data = Nfldraft)
cor(Nfldraft$rating, Nfldraft$forty)
summary(lm(rating ~ forty, data = Nfldraft))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
