% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Nfldraft}
\alias{Nfldraft}
\title{Rating, time in 40-yard dash, and weight of top offensive linemen in the
1994 NFL draft}
\format{A data frame with 29 observations on the following variable.
\describe{ 
\item{Rating.forty.weight}{a factor with levels
\code{5.000000000e+000 5.300000000e+000 3.100000000e+002}
\code{5.200000000e+000 5.230000000e+000 2.740000000e+002}
\code{5.200000000e+000 5.260000000e+000 2.950000000e+002}
\code{5.300000000e+000 5.180000000e+000 3.010000000e+002}
\code{5.400000000e+000 5.290000000e+000 3.010000000e+002}
\code{5.500000000e+000 5.090000000e+000 2.800000000e+002}
\code{5.500000000e+000 5.260000000e+000 3.050000000e+002}
\code{5.500000000e+000 5.560000000e+000 3.590000000e+002}
\code{5.700000000e+000 5.140000000e+000 2.770000000e+002}
\code{5.700000000e+000 5.290000000e+000 2.860000000e+002}
\code{5.900000000e+000 5.200000000e+000 3.130000000e+002}
\code{5.900000000e+000 5.250000000e+000 2.920000000e+002}
\code{6.000000000e+000 5.030000000e+000 2.740000000e+002}
\code{6.000000000e+000 5.270000000e+000 2.850000000e+002}
\code{6.000000000e+000 5.290000000e+000 3.050000000e+002}
\code{6.100000000e+000 5.270000000e+000 2.850000000e+002}
\code{6.100000000e+000 5.350000000e+000 3.210000000e+002}
\code{6.200000000e+000 5.230000000e+000 2.900000000e+002}
\code{6.200000000e+000 5.290000000e+000 2.890000000e+002}
\code{6.300000000e+000 5.360000000e+000 3.110000000e+002}
\code{6.400000000e+000 5.060000000e+000 3.150000000e+002}
\code{6.400000000e+000 5.260000000e+000 3.020000000e+002}
\code{6.500000000e+000 4.940000000e+000 2.850000000e+002}
\code{6.500000000e+000 5.180000000e+000 3.250000000e+002}
\code{7.000000000e+000 5.200000000e+000 3.250000000e+002}
\code{7.000000000e+000 5.360000000e+000 3.170000000e+002}
\code{7.100000000e+000 5.060000000e+000 2.890000000e+002}
\code{7.200000000e+000 5.200000000e+000 3.150000000e+002}
\code{7.600000000e+000 5.150000000e+000 3.030000000e+002}} 
}}
\description{
Data for Exercises 9.10 and 9.16
}
\examples{

str(Nfldraft)
attach(Nfldraft)
detach(Nfldraft)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

