% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Nicotine}
\alias{Nicotine}
\title{Nicotine content versus sales for 8 major brands of cigarettes}
\format{A data frame with 8 observations on the following 2 variables.
\describe{ 
\item{nicotine}{a numeric vector} 
\item{sales}{a numeric vector} 
}}
\description{
Data for Exercise 9.21
}
\examples{

str(Nicotine)
attach(Nicotine)
model <- lm(sales~nicotine)
summary(model)
detach(Nicotine)
remove(model)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

