% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Nicotine}
\alias{Nicotine}
\title{Nicotine content versus sales for 8 major brands of cigarettes}
\format{A data frame with 8 observations on the following 2 variables.
\describe{
\item{nicotine}{a numeric vector}
\item{sales}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 9.21
}
\examples{
str(Nicotine)
attach(Nicotine)
model <- lm(sales~nicotine)
summary(model)
detach(Nicotine)
remove(model)
}
\keyword{datasets}

