% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Nicotine}
\alias{Nicotine}
\title{Nicotine content versus sales for 8 major brands of cigarettes}
\format{A data frame/tibble with 8 observations on the following 2 variables.
\describe{ 
\item{nicotine}{nicotine content in milligrams} 
\item{sales}{sales figures in $100,000} 
}}
\usage{
Nicotine
}
\description{
Data for Exercise 9.21
}
\examples{

model <- lm(sales ~ nicotine, data = Nicotine)
plot(sales ~ nicotine, data = Nicotine)
abline(model)
summary(model)
predict(model, newdata = data.frame(nicotine = 1), 
     interval = "confidence", level = 0.99)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
