% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Orange}
\alias{Orange}
\title{Price of oranges versus size of the harvest}
\format{
A data frame/tibble with six observations on two variables
\describe{ 
\item{harvest}{harvest in millions of boxes} 
\item{price}{average price charged by California growers 
for a 75-pound box of navel oranges} 
}
}
\usage{
Orange
}
\description{
Data for Exercise 9.61
}
\examples{

plot(price ~ harvest, data = Orange)
model <- lm(price ~ harvest, data = Orange)
abline(model, col = "red")
summary(model)
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
