% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Oxytocin}
\alias{Oxytocin}
\title{Arterial blood pressure of 11 subjects before and after receiving oxytocin}
\format{A data frame with 11 observations on the following 3 variables.
\describe{ 
\item{Subject}{a numeric vector} 
\item{Before}{a numeric vector} 
\item{After}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.86
}
\examples{

str(Oxytocin)
attach(Oxytocin)
DIF <- Before - After
qqnorm(DIF)
qqline(DIF)
shapiro.test(DIF)
t.test(Before,After,paired=TRUE)
detach(Oxytocin)

}
\keyword{datasets}

