% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Parented}
\alias{Parented}
\title{Education backgrounds of parents of entering freshmen at a state university}
\format{A data frame with 6 observations on the following 3 variables.
\describe{ 
\item{Educat}{a factor with levels \code{4yr college
degree} \code{Doctoral degree} \code{Grad degree} \code{H.S grad or less}
\code{Some college} \code{Some grad school}} 
\item{Mother}{a numeric vector} 
\item{Father}{a numeric vector} 
}}
\description{
Data for Exercise 1.32
}
\examples{

str(Parented)
attach(Parented)
MAT <- cbind(Mother, Father)
row.names(MAT) <- Educat 
MAT 
barplot(t(MAT),beside=TRUE,legend=TRUE,col=c("blue","red")) 
detach(Parented) 
remove(MAT)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

