% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Patrol}
\alias{Patrol}
\title{Years of experience and number of tickets given by patrolpersons in New York
City}
\format{A data frame with 10 observations on the following 7 variables.
\describe{ 
\item{tickets}{a numeric vector} 
\item{years}{a numeric vector} 
\item{ln.tickets.}{a numeric vector}
\item{SRES1}{a numeric vector} 
\item{FITS1}{a numeric vector} 
\item{SRES2}{a numeric vector} 
\item{FITS2}{a numeric vector} 
}}
\description{
Data for Example 9.3
}
\examples{

str(Patrol)
attach(Patrol)
model <- lm(tickets~years)
summary(model)
detach(Patrol)
remove(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
