% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Patrol}
\alias{Patrol}
\title{Years of experience and number of tickets given by patrolpersons in New York
City}
\format{A data frame/tibble with ten observations on three variables
\describe{ 
\item{tickets}{number of tickets written per week} 
\item{years}{patrolperson's experience (in years)} 
\item{log_tickets}{natural log of \code{tickets}}
}}
\usage{
Patrol
}
\description{
Data for Example 9.3
}
\examples{

model <- lm(tickets ~ years, data = Patrol)
summary(model)
confint(model, level = 0.98)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
