% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Pearson}
\alias{Pearson}
\title{Karl Pearson's data on heights of brothers and sisters}
\format{A data frame with 11 observations on the following 2 variables.
\describe{ 
\item{brother}{a numeric vector} 
\item{sister}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 2.20
}
\examples{

str(Pearson)
attach(Pearson)
plot(brother,sister)
cor(brother,sister)
detach(Pearson)

}
\keyword{datasets}

