% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Phone}
\alias{Phone}
\title{Length of long-distance phone calls for a small business firm}
\format{A data frame/tibble with 20 observations on the following variable.
\describe{ 
\item{time}{duration of long distance phone call (in minutes)} 
}}
\usage{
Phone
}
\description{
Data for Exercise 6.95
}
\examples{

qqnorm(Phone$time)
qqline(Phone$time)
shapiro.test(Phone$time)
BSDA::SIGN.test(Phone$time, md = 5, alternative = "greater")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
