% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Phone}
\alias{Phone}
\title{Length of long-distance phone calls for a small business firm}
\format{A data frame with 20 observations on the following variable.
\describe{ 
\item{time}{a numeric vector} 
}}
\description{
Data for Exercise 6.95
}
\examples{

str(Phone)
attach(Phone)
qqnorm(time)
qqline(time)
shapiro.test(time)
SIGN.test(time,md=5,alternative="greater")
detach(Phone)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

