% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Politic}
\alias{Politic}
\title{Political party and gender in a voting district}
\format{A data frame with 250 observations on the following 2 variables.
\describe{ 
\item{Party}{a numeric vector} 
\item{Gender}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Example 8.3
}
\examples{

str(Politic)
attach(Politic)
table(Party,Gender)
chisq.test(table(Party,Gender))
detach(Politic)

}
\keyword{datasets}

