% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Politic}
\alias{Politic}
\title{Political party and gender in a voting district}
\format{
A data frame/tibble with 250 observations on two variables
\describe{ 
\item{party}{a factor with levels \code{republican}, \code{democrat}, and \code{other}} 
\item{gender}{a factor with levels \code{female} and \code{male}} 
}
}
\usage{
Politic
}
\description{
Data for Example 8.3
}
\examples{

T1 <- xtabs(~party + gender, data = Politic)
T1
chisq.test(T1)
rm(T1)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
