% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Pollutio}
\alias{Pollutio}
\title{Air pollution index for 15 randomly selected days for a major western city}
\format{A data frame with 15 observations on the following variable.
\describe{ 
\item{inde}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 5.59
}
\examples{

str(Pollutio)
attach(Pollutio)
EDA(inde)
t.test(inde,conf.level=.98)$conf
detach(Pollutio)

}
\keyword{datasets}

