% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Pollutio}
\alias{Pollutio}
\title{Air pollution index for 15 randomly selected days for a major western city}
\format{
A data frame/tibble with 15 observations on one variable
\describe{ 
\item{inde}{air pollution index} 
}
}
\usage{
Pollutio
}
\description{
Data for Exercise 5.59
}
\examples{

stem(Pollutio$inde)
t.test(Pollutio$inde, conf.level = 0.98)$conf

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
