% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Porosity}
\alias{Porosity}
\title{Porosity measurements on 20 samples of Tensleep Sandstone, Pennsylvanian
from Bighorn Basin in Wyoming}
\format{A data frame with 20 observations on the following variable.
\describe{ 
\item{porosity}{a numeric vector} 
}}
\description{
Data for Exercise 5.86
}
\examples{

str(Porosity)
attach(Porosity)
stem(porosity)
fivenum(porosity)
boxplot(porosity)
detach(Porosity)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
