% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Porosity}
\alias{Porosity}
\title{Porosity measurements on 20 samples of Tensleep Sandstone, Pennsylvanian
from Bighorn Basin in Wyoming}
\format{A data frame/tibble with 20 observations on the following variable.
\describe{ 
\item{porosity}{porosity measurement (percent)} 
}}
\source{
Davis, J. C. (1986), \emph{Statistics and Data Analysis in Geology}, 2nd edition,
pages 63-65.
}
\usage{
Porosity
}
\description{
Data for Exercise 5.86
}
\examples{

stem(Porosity$porosity)
fivenum(Porosity$porosity)
boxplot(Porosity$porosity)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
