% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Precinct}
\alias{Precinct}
\title{Robbery rates versus percent low income in eight precincts}
\format{A data frame/tibble with eight observations on two variables
\describe{ 
\item{rate}{robbery rate (per 1000 people)} 
\item{income}{percent with low income} 
}}
\usage{
Precinct
}
\description{
Data for Exercise 2.2 and 2.38
}
\examples{

plot(rate ~ income, data = Precinct)
model <- (lm(rate ~ income, data = Precinct))
abline(model, col = "red")
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
