% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Precinct}
\alias{Precinct}
\title{Robbery rates versus percent low income in 8 precincts}
\format{A data frame with 8 observations on the following 2 variables.
\describe{ 
\item{rate}{a numeric vector} 
\item{income}{a numeric vector} 
}}
\description{
Data for Exercise 2.2 and 2.38
}
\examples{

str(Precinct)
attach(Precinct)
plot(rate,income,main="Exercise 2.2")
model <- lm(income~rate)
model
abline(model,col="green")
detach(Precinct)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

