% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Prejudic}
\alias{Prejudic}
\title{Racial prejudice measured on a sample of 25 high school students}
\format{A data frame with 25 observations on the following variable.
\describe{
\item{prejud}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 5.10 and 5.22
}
\examples{
str(Prejudic)
attach(Prejudic)
EDA(prejud)
detach(Prejudic)
}
\keyword{datasets}

