% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Press}
\alias{Press}
\title{Degree of confidence in the press versus education level for 20 randomly
selected persons}
\format{A data frame/tibble with 20 observations on the following 2 variables.
\describe{ 
\item{education_yrs}{years of education} 
\item{confidence}{degree of confidence in the press (the higher the score, the more confidence)} 
}}
\usage{
Press
}
\description{
Data for Exercise 9.55
}
\examples{

plot(confidence ~ education_yrs, data = Press)
model <- lm(confidence ~ education_yrs, data = Press)
abline(model, col = "purple")
summary(model)
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
