% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Press}
\alias{Press}
\title{Degree of confidence in the press versus education level for 20 randomly
selected persons}
\format{A data frame with 20 observations on the following 4 variables.
\describe{ 
\item{educat}{a numeric vector} 
\item{confid}{a numeric vector} 
\item{SRES1}{a numeric vector}
\item{FITS1}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 9.55
}
\examples{

str(Press)
attach(Press)
summary(lm(confid~educat))
detach(Press)

}
\keyword{datasets}

