% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Prognost}
\alias{Prognost}
\title{Klopfer's prognostic rating scale for subjects receiving behavior
modification therapy}
\format{A data frame with 15 observations on the following variable.
\describe{ 
\item{score}{a numeric vector} 
}}
\description{
Data for Exercise 6.61
}
\examples{

str(Prognost)
attach(Prognost)
EDA(score)
t.test(score,mu=9)
detach(Prognost)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
