% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Prognost}
\alias{Prognost}
\title{Klopfer's prognostic rating scale for subjects receiving behavior
modification therapy}
\format{A data frame/tibble with 15 observations on one variable
\describe{ 
\item{kprs_score}{Kloper's Prognostic Rating Scale score} 
}}
\source{
Newmark, C., et al. (1973), Predictive Validity of the Rorschach Prognostic Rating Scale
with Behavior Modification Techniques, \emph{Journal of Clinical Psychology, 29}, 246-248
}
\usage{
Prognost
}
\description{
Data for Exercise 6.61
}
\examples{

BSDA::EDA(Prognost$kprs_score)
t.test(Prognost$kprs_score, mu = 9)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
