% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Program}
\alias{Program}
\title{Effects of four different methods of programmed learning for statistics
students}
\format{A data frame with 11 observations on the following 4 variables.
\describe{ 
\item{Method1}{a numeric vector} 
\item{Method2}{a numeric vector} 
\item{Method3}{a numeric vector}
\item{Method4}{a numeric vector} 
}}
\description{
Data for Exercise 10.17
}
\examples{

str(Program)
attach(Program)
STACKED <-stack(Program)
STACKED[1:5,]
boxplot(values~ind,col=c("red","blue","green","yellow"),data=STACKED)
anova(lm(values~ind,data=STACKED))
remove(STACKED)
detach(Program)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
