% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Psat}
\alias{Psat}
\title{PSAT scores versus SAT scores}
\format{A data frame with 7 observations on the following 4 variables.
\describe{
\item{psat}{a numeric vector}
\item{sat}{a numeric vector}
\item{SRES1}{a numeric vector}
\item{FITS1}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 2.50
}
\examples{
str(Psat)
attach(Psat)
model <- lm(sat~psat)
plot(psat,resid(model))
detach(Psat)
}
\keyword{datasets}

