% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Psat}
\alias{Psat}
\title{PSAT scores versus SAT scores}
\format{A data frame/tibble with seven observations on the two variables
\describe{
\item{psat}{PSAT score} 
\item{sat}{SAT score} 
}}
\usage{
Psat
}
\description{
Data for Exercise 2.50
}
\examples{

model <- lm(sat ~ psat, data = Psat)
par(mfrow = c(1, 2))
plot(Psat$psat, resid(model))
plot(model, which = 1)
rm(model)
par(mfrow = c(1, 1))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
