% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Psych}
\alias{Psych}
\title{Correct responses for 24 students in a psychology experiment}
\format{A data frame/tibble with 23 observations on one variable
\describe{ 
\item{score}{number of correct repsonses in a psychology experiment} 
}}
\usage{
Psych
}
\description{
Data for Exercise 1.42
}
\examples{

stem(Psych$score)
BSDA::EDA(Psych$score)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
