% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Psych}
\alias{Psych}
\title{Correct responses for 24 students in a psychology experiment}
\format{A data frame with 23 observations on the following variable.
\describe{ 
\item{score}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 1.42
}
\examples{

str(Psych)
attach(Psych)
stem(score)
EDA(score)
detach(Psych)

}
\keyword{datasets}

