% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Puerto}
\alias{Puerto}
\title{Weekly incomes of a random sample of 50 Puerto Rican families in Miami}
\format{A data frame with 50 observations on the following variable.
\describe{ 
\item{income}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 5.22 and 5.65
}
\examples{

str(Puerto)
attach(Puerto)
EDA(income)
t.test(income,conf.level=.90)$conf
detach(Puerto)

}
\keyword{datasets}

