% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Puerto}
\alias{Puerto}
\title{Weekly incomes of a random sample of 50 Puerto Rican families in Miami}
\format{
A data frame/tibble with 50 observations on one variable
\describe{ 
\item{income}{weekly family income (in dollars)} 
}
}
\usage{
Puerto
}
\description{
Data for Exercise 5.22 and 5.65
}
\examples{

stem(Puerto$income)
boxplot(Puerto$income, col = "purple")
t.test(Puerto$income,conf.level = .90)$conf

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
