% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Quail}
\alias{Quail}
\title{Plasma LDL levels in two groups of quail}
\format{A data frame with 20 observations on the following 2 variables.
\describe{ 
\item{placebo}{a numeric vector}
\item{treatmen}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 1.53, 1.77, 1.88, 5.66, and 7.50
}
\examples{

str(Quail)
attach(Quail)
boxplot(placebo,treatmen,names=c("Placebo","Treatment"),
horizontal=TRUE,xlab="LDL level",col=c("lightblue","yellow"))
boxplot(placebo,treatmen,names=c("Placebo","Treatment"),
ylab="LDL level",col=c("lightblue","yellow"))
detach(Quail)

}
\keyword{datasets}

