% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Quality}
\alias{Quality}
\title{Quality control test scores on two manufacturing processes}
\format{A data frame/tibble with 15 observations on two variables
\describe{ 
\item{process}{a character variable with values \code{Process1} and \code{Process2}}
\item{score}{results of a quality control test} 
}}
\usage{
Quality
}
\description{
Data for Exercise 7.81
}
\examples{

boxplot(score ~ process, data = Quality)
t.test(score ~ process, data = Quality)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
