% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Quality}
\alias{Quality}
\title{Quality control test scores on two manufacturing processes}
\format{A data frame with 8 observations on the following 2 variables.
\describe{ 
\item{Process1}{a numeric vector}
\item{Process2}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.81
}
\examples{

str(Quality)
attach(Quality)
qqnorm(Process1)
qqline(Process1)
shapiro.test(Process1)
qqnorm(Process2)
qqline(Process2)
shapiro.test(Process2)
t.test(Process1,Process2)
detach(Quality)

}
\keyword{datasets}

