% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Rainks}
\alias{Rainks}
\title{Rainfall in an area of west central Kansas and four surrounding counties}
\format{A data frame with 35 observations on the following 5 variables.
\describe{ 
\item{rain}{a numeric vector} 
\item{x1}{a numeric vector} 
\item{x2}{a numeric vector} 
\item{x3}{a numeric vector} 
\item{x4}{a numeric vector} 
}}
\description{
Data for Exercise 9.8
}
\examples{

str(Rainks)
attach(Rainks)
cor(Rainks)
lm(rain~x2)
detach(Rainks)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

