% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Rainks}
\alias{Rainks}
\title{Rainfall in an area of west central Kansas and four surrounding counties}
\format{A data frame/tibble with 35 observations on the following 5 variables.
\describe{ 
\item{rain}{rainfall (in inches)} 
\item{x1}{rainfall (in inches)} 
\item{x2}{rainfall (in inches)} 
\item{x3}{rainfall (in inches)} 
\item{x4}{rainfall (in inches)} 
}}
\source{
R. Picard, K. Berk (1990), Data Splitting, \emph{The American Statistician, 44}, (2),
140-147.
}
\usage{
Rainks
}
\description{
Data for Exercise 9.8
}
\examples{

cor(Rainks)
model <- lm(rain ~ x2, data = Rainks)
summary(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
