% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Randd}
\alias{Randd}
\title{Research and development expenditures and sales of a large company}
\format{A data frame/tibble with 12 observations on two variables
\describe{ 
\item{rd}{research and development expenditures (in million dollars)} 
\item{sales}{sales (in million dollars)} 
}}
\usage{
Randd
}
\description{
Data for Exercise 9.36 and Example 9.8
}
\examples{

plot(sales ~ rd, data = Randd)
model <- lm(sales ~ rd, data = Randd)
abline(model, col = "purple")
summary(model)
plot(model, which = 1)
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
