% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Randd}
\alias{Randd}
\title{Research and development expenditures and sales of a large company}
\format{A data frame with 12 observations on the following 5 variables.
\describe{ 
\item{rd}{a numeric vector} 
\item{sales}{a numeric vector} 
\item{SRES1}{a numeric vector}
\item{FITS1}{a numeric vector} 
\item{RESI1}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 9.36 and Example 9.8
}
\examples{

str(Randd)
attach(Randd)
plot(rd,sales)
model <- lm(sales~rd)
abline(model)
summary(model)
# plot(model)
detach(Randd)
remove(model)

}
\keyword{datasets}

