% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Rat}
\alias{Rat}
\title{Survival times of 20 rats exposed to high levels of radiation}
\format{A data frame with 20 observations on the following variable.
\describe{ 
\item{survival.time}{a numeric vector} 
}}
\description{
Data for Exercise 1.52, 1.76, 5.62, and 6.44
}
\examples{

str(Rat)
attach(Rat)
hist(survival.time)
qqnorm(survival.time,col="blue")
qqline(survival.time,col="red")
t.test(survival.time)$conf
t.test(survival.time,mu=100,alternative="greater")
detach(Rat)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
