% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Rat}
\alias{Rat}
\title{Survival times of 20 rats exposed to high levels of radiation}
\format{
A data frame/tibble with 20 observations on one variable
\describe{ 
\item{survival_time}{survival time in weeks for rats exposed to a high level of radiation} 
}
}
\source{
J. Lawless, \emph{Statistical Models and Methods for Lifetime Data} (New York: Wiley, 1982).
}
\usage{
Rat
}
\description{
Data for Exercise 1.52, 1.76, 5.62, and 6.44
}
\examples{

hist(Rat$survival_time)
qqnorm(Rat$survival_time)
qqline(Rat$survival_time)
summary(Rat$survival_time)
t.test(Rat$survival_time)
t.test(Rat$survival_time, mu = 100, alternative = "greater")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
