% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Ratings}
\alias{Ratings}
\title{Grade point averages versus teacher's ratings}
\format{A data frame with 250 observations on the following 7 variables.
\describe{ 
\item{F}{a numeric vector} 
\item{D}{a numeric vector} 
\item{C}{a numeric vector} 
\item{B}{a numeric vector} 
\item{A}{a numeric vector} 
\item{Rating}{a factor with levels \code{A} \code{B} \code{C} \code{D} \code{F}}
\item{GPA}{a numeric vector}
}}
\description{
Data for Example 2.6
}
\examples{

str(Ratings)
attach(Ratings)
boxplot(GPA~Rating,xlab="Teacher's Rating",ylab="GPA",main="Example 2.6",col="pink")
detach(Ratings)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

