% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Reaction}
\alias{Reaction}
\title{Threshold reaction time for persons subjected to emotional stress}
\format{A data frame/tibble with 12 observations on one variable
\describe{ 
\item{time}{threshold reaction time (in seconds) for persons subjected to emotional stress} 
}}
\usage{
Reaction
}
\description{
Data for Example 6.11
}
\examples{

stem(Reaction$time)
SIGN.test(Reaction$time, md = 15, alternative = "less")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
