% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Reading}
\alias{Reading}
\title{Standardized reading scores for 30 fifth graders}
\format{
A data frame/tibble with 30 observations on four variables
\describe{ 
\item{score}{standardized reading test score} 
\item{sorted}{sorted values of \code{score}} 
\item{trimmed}{trimmed values of \code{sorted}}
\item{winsoriz}{winsorized values of \code{score}} 
}
}
\usage{
Reading
}
\description{
Data for Exercise 1.72 and 2.10
}
\examples{

hist(Reading$score, main = "Exercise 1.72", 
     col = "lightgreen", xlab = "Standardized reading score")
summary(Reading$score)
sd(Reading$score)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
