% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Readiq}
\alias{Readiq}
\title{Reading scores versus IQ scores}
\format{A data frame with 14 observations on the following 2 variables.
\describe{ 
\item{reading}{a numeric vector} 
\item{IQ}{a numeric vector} 
}}
\description{
Data for Exercises 2.10 and 2.53
}
\examples{

str(Readiq)
attach(Readiq)
plot(IQ,reading)
model <- lm(reading~IQ)
abline(model)
detach(Readiq)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

