% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Readiq}
\alias{Readiq}
\title{Reading scores versus IQ scores}
\format{A data frame/tibble with 14 observations on two variables
\describe{ 
\item{reading}{reading achievement score} 
\item{iq}{IQ score} 
}}
\usage{
Readiq
}
\description{
Data for Exercises 2.10 and 2.53
}
\examples{

plot(reading ~ iq, data = Readiq)
model <- lm(reading ~ iq, data = Readiq)
abline(model)
predict(model, newdata = data.frame(iq = c(100, 120)))
residuals(model)[c(6, 7)]
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
