% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Referend}
\alias{Referend}
\title{Opinion on referendum by view on freedom of the press}
\format{
A data frame with 237 observations on two variables
\describe{ 
\item{choice}{a factor with levels \code{A}, \code{B}, and \code{C}} 
\item{response}{a factor with levels \code{for}, \code{against}, and \code{undecided}} 
}
}
\usage{
Referend
}
\description{
Data for Exercise 8.20
}
\examples{

T1 <- xtabs(~choice + response, data = Referend)
T1
chisq.test(T1)
chisq.test(T1)$expected

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
