% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Referend}
\alias{Referend}
\title{Opinion on referendum by view on freedom of the press}
\format{A data frame with 3 observations on the following 4 variables.
\describe{ 
\item{Response}{a factor with levels \code{A} \code{B}
\code{C}} 
\item{For}{a numeric vector} 
\item{Against}{a numeric vector} 
\item{undecide}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 8.20
}
\examples{

str(Referend)
attach(Referend)
chisq.test(Referend[,2:4])
detach(Referend)

}
\keyword{datasets}

