% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Region}
\alias{Region}
\title{Pollution index taken in three regions of the country}
\format{A data frame with 48 observations on the following 6 variables.
\describe{
 \item{West}{a numeric vector} 
 \item{Central}{a numeric vector} 
 \item{East}{a numeric vector} 
 \item{Index}{a numeric vector} 
 \item{Region}{a numeric vector}
 \item{Ranks}{a numeric vector} 
 }}
\description{
Data for Exercise 10.26
}
\examples{

str(Region)
attach(Region)
boxplot(Index~Region)
anova(lm(Index~as.factor(Region)))
detach(Region)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

