% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Region}
\alias{Region}
\title{Pollution index taken in three regions of the country}
\format{A data frame/tibble with 48 observations on the following 3 variables.
\describe{
 \item{pollution}{pollution index} 
 \item{region}{region of a county (\code{west}, \code{central}, and \code{east})}
 \item{ranks}{ranked values of \code{pollution}} 
 }}
\usage{
Region
}
\description{
Data for Exercise 10.26
}
\examples{

boxplot(pollution ~ region, data = Region)
anova(lm(pollution ~ region, data = Region))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
