% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Register}
\alias{Register}
\title{Maintenance cost versus age of cash registers in a department store}
\format{A data frame/tibble with nine observations on two variables
\describe{ 
\item{age}{age of cash register (in years)} 
\item{cost}{maintenance cost of cash register (in dollars)} 
}}
\usage{
Register
}
\description{
Data for Exercise 2.3, 2.39, and 2.54
}
\examples{

plot(cost ~ age, data = Register)
model <- lm(cost ~ age, data = Register)
abline(model, col = "red")
predict(model, newdata = data.frame(age = c(5, 10)))
plot(model, which = 1)
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
