% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Rehab}
\alias{Rehab}
\title{Rehabilitative potential of 20 prison inmates as judged by two psychiatrists}
\format{A data frame with 20 observations on the following 3 variables.
\describe{ 
\item{Psych1}{a numeric vector} 
\item{Psych2}{a numeric vector} 
\item{differ}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.61
}
\examples{

str(Rehab)
attach(Rehab)
qqnorm(differ)
qqline(differ)
shapiro.test(differ)
boxplot(Psych1,Psych2,names=c("Psychiatrist 1","Psychiatrist 2"),
col=c("pink","lightblue"))
t.test(Psych1,Psych2,paired=TRUE)
detach(Rehab)

}
\keyword{datasets}

