% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Remedial}
\alias{Remedial}
\title{Math placement test score for 35 freshmen females and 42 freshmen males}
\format{A data frame/tibble with 84 observations on two variables
\describe{ 
\item{gender}{a character variable with values \code{female} and \code{male}} 
\item{score}{math placement score} 
}}
\usage{
Remedial
}
\description{
Data for Exercise 7.43
}
\examples{

boxplot(score ~ gender, data = Remedial, 
col = c("purple", "blue"))
t.test(score ~ gender, data = Remedial, conf.level = 0.98)
t.test(score ~ gender, data = Remedial, conf.level = 0.98)$conf
wilcox.test(score ~ gender, data = Remedial, 
            conf.int = TRUE, conf.level = 0.98)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
