% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Remedial}
\alias{Remedial}
\title{Math placement test score for 35 freshmen females and 42 freshmen males}
\format{A data frame with 42 observations on the following 2 variables.
\describe{ 
\item{female}{a numeric vector} 
\item{male}{a numeric vector} 
}}
\description{
Data for Exercise 7.43
}
\examples{

str(Remedial)
attach(Remedial)
boxplot(female,male,col=c("blue","red"))
wilcox.test(female,male,conf.int=TRUE)
t.test(female,male)
detach(Remedial)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
