% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Rentals}
\alias{Rentals}
\title{Weekly rentals for 45 apartments}
\format{A data frame with 45 observations on the following variable.
\describe{ 
\item{rent}{a numeric vector} 
}}
\description{
Data for Exercise 1.122
}
\examples{

str(Rentals)
attach(Rentals)
EDA(rent)
detach(Rentals)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

