% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Repair}
\alias{Repair}
\title{Recorded times for repairing 22 automobiles involved in wrecks}
\format{A data frame with 22 observations on the following variable.
\describe{ 
\item{time}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 5.77
}
\examples{

str(Repair)
attach(Repair)
stem(time)
SIGN.test(time,conf.level=.98)
detach(Repair)

}
\keyword{datasets}

