% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Retail}
\alias{Retail}
\title{Length of employment versus gross sales for 10 employees of a large retail
store}
\format{A data frame/tibble with 10 observations on the following 2 variables.
\describe{ 
\item{months}{length of employment (in months)}
\item{sales}{employee gross sales (in dollars)} 
}}
\usage{
Retail
}
\description{
Data for Exercise 9.59
}
\examples{

plot(sales ~ months, data = Retail)
model <- lm(sales ~ months, data = Retail)
abline(model)
summary(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
