% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Ronbrown1}
\alias{Ronbrown1}
\title{Oceanography data obtained at site 1 by scientist aboard the ship Ron Brown}
\format{A data frame with 75 observations on the following 12 variables.
\describe{ 
\item{depth}{a numeric vector} 
\item{downtemp1}{a numeric vector} 
\item{downtemp2}{a numeric vector}
\item{downsalinity1}{a numeric vector}
\item{downsalinity2}{a numeric vector} 
\item{downdensity}{a numeric vector} 
\item{C7}{a numeric vector} 
\item{uptemp1}{a numeric vector} 
\item{uptemp2}{a numeric vector}
\item{upsalinity1}{a numeric vector} 
\item{upsalinity2}{a numeric vector} 
\item{updensity}{a numeric vector} 
}}
\description{
Data for Exercise 2.9
}
\examples{

str(Ronbrown1)
attach(Ronbrown1)
plot(depth,downtemp1)
detach(Ronbrown1)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

