% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Ronbrown1}
\alias{Ronbrown1}
\title{Oceanography data obtained at site 1 by scientist aboard the ship Ron Brown}
\format{A data frame/tibble with 75 observations on the following 2 variables.
\describe{ 
\item{depth}{ocen depth (in meters)} 
\item{temperature}{ocean temperature (in Celsius)} 
}}
\usage{
Ronbrown1
}
\description{
Data for Exercise 2.9
}
\examples{

plot(temperature ~ depth, data = Ronbrown1, ylab = "Temperature")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
