% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Ronbrown2}
\alias{Ronbrown2}
\title{Oceanography data obtained at site 2 by scientist aboard the ship Ron Brown}
\format{A data frame/tibble with 150 observations on three variables
\describe{ 
\item{depth}{ocean depth (in meters)}
\item{temperature}{ocean temperature (in Celcius)} 
\item{salinity}{ocean salinity level} 
}}
\usage{
Ronbrown2
}
\description{
Data for Exercise 2.56 and Example 2.4
}
\examples{

plot(salinity ~ depth, data = Ronbrown2)
model <- lm(salinity ~ depth, data = Ronbrown2)
summary(model)
plot(model, which = 1)
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
