% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Ronbrown2}
\alias{Ronbrown2}
\title{Oceanography data obtained at site 2 by scientist aboard the ship Ron Brown}
\format{A data frame with 150 observations on the following 6 variables.
\describe{ 
\item{depth}{a numeric vector}
\item{primarytemp}{a numeric vector} 
\item{secondarytemp}{a numeric vector} 
\item{primarysalinity}{a numeric vector}
\item{secondarysalinity}{a numeric vector} 
\item{density}{a numeric vector} 
}}
\description{
Data for Exercise 2.56 and Example 2.4
}
\examples{

str(Ronbrown2)
attach(Ronbrown2)
plot(depth,primarysalinity,xlab="Depth",ylab="Salinity",
main="Example 2.4",col="tomato")
detach(Ronbrown2)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

