% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Rural}
\alias{Rural}
\title{Social adjustment scores for a rural group and a city group of children}
\format{A data frame with 33 observations on the following 4 variables.
\describe{ 
\item{Rural}{a numeric vector} 
\item{City}{a numeric vector} 
\item{score}{a numeric vector} 
\item{code}{a numeric vector} 
}}
\description{
Data for Exercise 7.16
}
\examples{

str(Rural)
attach(Rural)
wilcox.test(score~code)
wilcox.test(Rural,City)
detach(Rural)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
