% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Salary}
\alias{Salary}
\title{Starting salaries for 25 new PhD psychologist}
\format{A data frame with 25 observations on the following variable.
\describe{ 
\item{salary}{a numeric vector} 
}}
\description{
Data for Exercise 3.66
}
\examples{

data(Salary)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

