% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Salary}
\alias{Salary}
\title{Starting salaries for 25 new PhD psychologist}
\format{A data frame with 25 observations on the following variable.
\describe{ 
\item{salary}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 3.66
}
\examples{

data(Salary)

}
\keyword{datasets}

