% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Salary}
\alias{Salary}
\title{Starting salaries for 25 new PhD psychologist}
\format{A data frame/tibble with 25 observations on the following variable.
\describe{ 
\item{salary}{starting salary for Ph.D. psycholgists} 
}}
\usage{
Salary
}
\description{
Data for Exercise 3.66
}
\examples{

qqnorm(Salary$salary, pch = 19, col = "purple")
qqline(Salary$salary, col = "blue")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
