% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Salinity}
\alias{Salinity}
\title{Surface-water salinity measurements from Whitewater Bay, Florida}
\format{A data frame with 48 observations on the following variable.
\describe{ 
\item{salinity}{a numeric vector} 
}}
\description{
Data for Exercise 5.27 and 5.64
}
\examples{

str(Salinity)
attach(Salinity)
EDA(salinity)
t.test(salinity,conf.level=.99)$conf
detach(Salinity)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

