% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Saving}
\alias{Saving}
\title{Problem asset ration for savings and loan companies in California, New York,
and Texas}
\format{A data frame with 75 observations on the following 6 variables.
\describe{ 
\item{calif}{a numeric vector} 
\item{newyork}{a numeric vector} 
\item{texas}{a numeric vector} 
\item{PAR}{a numeric vector} 
\item{state}{a numeric vector}
\item{ranks}{a numeric vector}
 }}
\description{
Data for Exercise 10.34 and 10.49
}
\examples{

str(Saving)
attach(Saving)
boxplot(PAR~state)
kruskal.test(PAR~as.factor(state))
detach(Saving)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

