% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Scales}
\alias{Scales}
\title{Readings obtained from a 100 pound weight placed on four brands of bathroom
scales}
\format{A data frame with 20 observations on the following 2 variables.
\describe{ 
\item{Brand}{a factor with levels \code{A} \code{B}
\code{C} \code{D}} 
\item{reading}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 1.89
}
\examples{

str(Scales)
attach(Scales)
boxplot(reading~Brand,ylab="Reading",xlab="Brand",main="Problem 1.89")
detach(Scales)

}
\keyword{datasets}

