% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Schizop2}
\alias{Schizop2}
\title{Exam scores for 17 patients to assess the learning ability of schizophrenics
after taking a specified does of a tranquilizer}
\format{A data frame/tibble with 17 observations on the following variable.
\describe{ 
\item{score}{schizophrenics score on a second standardized exam} 
}}
\usage{
Schizop2
}
\description{
Data for Exercise 6.99
}
\examples{

hist(Schizop2$score, xlab = "score on standardized test after a tranquilizer", 
main = "Exercise 6.99", breaks = 10, col = "orange")
BSDA::EDA(Schizop2$score)
BSDA::SIGN.test(Schizop2$score, md = 22, alternative = "greater")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
