% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Schizop2}
\alias{Schizop2}
\title{Exam scores for 17 patients to assess the learning ability of schizophrenics
after taking a specified does of a tranquilizer}
\format{A data frame with 17 observations on the following variable.
\describe{ 
\item{score}{a numeric vector} 
}}
\description{
Data for Exercise 6.99
}
\examples{

str(Schizop2)
attach(Schizop2)
EDA(score)
SIGN.test(score,md=22,alternative="greater")
detach(Schizop2)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
