% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Schizoph}
\alias{Schizoph}
\title{Standardized exam scores for 13 patients to investigate the learning ability
of schizophrenics after a specified dose of a tranquilizer}
\format{A data frame with 13 observations on the following variable.
\describe{ 
\item{score}{a numeric vector} 
}}
\description{
Data for Example 6.10
}
\examples{

str(Schizoph)
attach(Schizoph)
EDA(score)
t.test(score,mu=20)
detach(Schizoph)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

