% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Seatbelt}
\alias{Seatbelt}
\title{Injury level versus seatbelt usage}
\format{A data frame/tibble with 86,759 observations on two variables
\describe{ 
\item{seatbelt}{a factor with levels \code{No} and \code{Yes}} 
\item{injuries}{a factor with levels \code{None}, \code{Minimal}, 
\code{Minor}, or \code{Major} indicating the extent of the drivers injuries}
}}
\source{
Jobson, J. (1982), \emph{Applied Multivariate Data Analysis}, Springer-Verlag, 
New York, p. 18.
}
\usage{
Seatbelt
}
\description{
Data for Exercise 8.24
}
\examples{

T1 <- xtabs(~seatbelt + injuries, data = Seatbelt)
T1
chisq.test(T1)
rm(T1)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
