% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Seatbelt}
\alias{Seatbelt}
\title{Injury level versus seatbelt usage}
\format{A data frame with 2 observations on the following 5 variables.
\describe{
\item{seatbelt}{a factor with levels \code{no}
\code{yes}}
\item{None}{a numeric vector}
\item{Minimal}{a numeric vector}
\item{Minor}{a numeric vector}
\item{Major}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 8.24
}
\examples{
str(Seatbelt)
attach(Seatbelt)
Seatbelt
chisq.test(Seatbelt[,2:5])
detach(Seatbelt)
}
\keyword{datasets}

