% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Seatbelt}
\alias{Seatbelt}
\title{Injury level versus seatbelt usage}
\format{A data frame with 2 observations on the following 5 variables.
\describe{ 
\item{seatbelt}{a factor with levels \code{no}
\code{yes}} 
\item{None}{a numeric vector} 
\item{Minimal}{a numeric vector} 
\item{Minor}{a numeric vector}
\item{Major}{a numeric vector} 
}}
\description{
Data for Exercise 8.24
}
\examples{

str(Seatbelt)
attach(Seatbelt)
Seatbelt
chisq.test(Seatbelt[,2:5])
detach(Seatbelt)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

