% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Selfdefe}
\alias{Selfdefe}
\title{Self-confidence scores for 9 women before and after instructions on
self-defense}
\format{A data frame with 9 observations on the following 3 variables.
\describe{ 
\item{Woman}{a numeric vector} 
\item{Before}{a numeric vector} 
\item{After}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Example 7.19
}
\examples{

str(Selfdefe)
attach(Selfdefe)
DIF <- After-Before
qqnorm(DIF)
qqline(DIF)
shapiro.test(DIF)
t.test(After,Before,paired=TRUE,alternative="greater")
detach(Selfdefe)
remove(DIF)

}
\keyword{datasets}

