% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Selfdefe}
\alias{Selfdefe}
\title{Self-confidence scores for 9 women before and after instructions on
self-defense}
\format{A data frame/tibble with 9 observations on the following 3 variables.
\describe{ 
\item{woman}{number identifying the woman} 
\item{before}{before the course self-confidence score} 
\item{after}{after the course self-confidence score} 
}}
\usage{
Selfdefe
}
\description{
Data for Example 7.19
}
\examples{

Selfdefe$differ <- Selfdefe$after - Selfdefe$before
Selfdefe
t.test(Selfdefe$differ, alternative = "greater")
t.test(Selfdefe$after, Selfdefe$before, 
      paired = TRUE, alternative = "greater")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
