% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Senior}
\alias{Senior}
\title{Reaction times of 30 senior citizens applying for drivers license renewals}
\format{A data frame/tibble with 31 observations on the following variable.
\describe{ 
\item{reaction}{reaction time for senior citizens applying for a driver's license renewal} 
}}
\usage{
Senior
}
\description{
Data for Exercise 1.83 and 3.67
}
\examples{

stem(Senior$reaction)
fivenum(Senior$reaction)
boxplot(Senior$reaction, main = "Problem 1.83, part d",
        horizontal = TRUE, col = "purple")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
