% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Senior}
\alias{Senior}
\title{Reaction times of 30 senior citizens applying for drivers license renewals}
\format{A data frame with 31 observations on the following variable.
\describe{ 
\item{reaction}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 1.83 and 3.67
}
\examples{

str(Senior)
attach(Senior)
fivenum(reaction)
boxplot(reaction,horizontal=TRUE,main="Problem 1.83 Part d.",col="orange")
detach(Senior)

}
\keyword{datasets}

