% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Sentence}
\alias{Sentence}
\title{Sentences of 41 prisoners convicted of a homicide offense}
\format{A data frame/tibble with 41 observations on one variable
\describe{ 
\item{months}{sentence length (in months) for prisoners convicted of homocide} 
}}
\source{
U.S. Department of Justice, Bureau of Justice Statistics, \emph{Prison Sentences
and Time Served for Violence}, NCJ-153858, April 1995.
}
\usage{
Sentence
}
\description{
Data for Exercise 1.123
}
\examples{

stem(Sentence$months)
ll <- mean(Sentence$months)-2*sd(Sentence$months)
ul <- mean(Sentence$months)+2*sd(Sentence$months)
limits <- c(ll, ul)
limits
rm(ul, ll, limits)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
