% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Sentence}
\alias{Sentence}
\title{Sentences of 41 prisoners convicted of a homicide offense}
\format{A data frame with 41 observations on the following variable.
\describe{ 
\item{months}{a numeric vector} 
}}
\description{
Data for Exercise 1.123
}
\examples{

str(Sentence)
attach(Sentence)
stem(months)
EDA(months)
ll <- mean(months)-2*sd(months)
ul <- mean(months)+2*sd(months)
limits <- c(ll,ul)
limits
detach(Sentence)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

