% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Shkdrug}
\alias{Shkdrug}
\title{Effects of a drug and electroshock therapy on the ability to solve simple
tasks}
\format{A data frame/tibble with 64 observations on two variables
\describe{ 
\item{treatment}{type of treament \code{Drug/NoS}, \code{Drug/Shk}, \code{NoDg/NoS}, or \code{NoDrug/S}}
\item{response}{number of tasks completed in a 10-minute period}
}}
\usage{
Shkdrug
}
\description{
Data for Exercises 10.11 and 10.12
}
\examples{

boxplot(response ~ treatment, data = Shkdrug, col = "gray")
model <- lm(response ~ treatment, data = Shkdrug)
anova(model)
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
