% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Shock}
\alias{Shock}
\title{Effect of experimental shock on time to complete difficult task}
\format{A data frame/tibble with 27 observations on two variables
\describe{ 
\item{group}{grouping variable with values of \code{Group1} (no shock), 
\code{Group2} (medium shock), and \code{Group3} (severe shock)} 
\item{attempts}{number of attempts to complete a task}
}}
\usage{
Shock
}
\description{
Data for Exercise 10.50
}
\examples{

boxplot(attempts ~ group, data = Shock, col = "violet")
model <- lm(attempts ~ group, data = Shock)
anova(model)
rm(model)


}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
