% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Shock}
\alias{Shock}
\title{Effect of experimental shock on time to complete difficult task}
\format{A data frame with 9 observations on the following 3 variables.
\describe{ 
\item{Group1}{a numeric vector} 
\item{Group2}{a numeric vector} 
\item{Group3}{a numeric vector} 
}}
\description{
Data for Exercise 10.50
}
\examples{

str(Shock)
attach(Shock)
STACKED <-stack(Shock)
STACKED[1:5,]
boxplot(values~ind,col=c("red","blue","green"),data=STACKED)
anova(lm(values~ind,data=STACKED))
remove(STACKED)
detach(Shock)


}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
