% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Shoplift}
\alias{Shoplift}
\title{Sales receipts versus shoplifting losses for a department store}
\format{A data frame with 8 observations on the following 2 variables.
\describe{ 
\item{sales}{a numeric vector} 
\item{loss}{a numeric vector} 
}}
\description{
Data for Exercise 9.58
}
\examples{

str(Shoplift)
attach(Shoplift)
summary(lm(loss~sales))
detach(Shoplift)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

