% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Shoplift}
\alias{Shoplift}
\title{Sales receipts versus shoplifting losses for a department store}
\format{A data frame/tibble with eight observations on two variables
\describe{ 
\item{sales}{sales (in 1000 dollars)} 
\item{loss}{loss (in 100 dollars)} 
}}
\usage{
Shoplift
}
\description{
Data for Exercise 9.58
}
\examples{

plot(loss ~ sales, data = Shoplift)
model <- lm(loss ~ sales, data = Shoplift)
summary(model)
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
