% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Short}
\alias{Short}
\title{James Short's measurements of the parallax of the sun}
\format{A data frame with 158 observations on the following 10 variables.
\describe{ 
\item{Sample.1}{a numeric vector}
\item{Sample.2}{a numeric vector} 
\item{Sample.3}{a numeric vector} 
\item{Sample.4}{a numeric vector} 
\item{Sample.5}{a numeric vector} 
\item{Sample.6}{a numeric vector}
\item{Sample.7}{a numeric vector} 
\item{Sample.8}{a numeric vector} 
\item{Parallax}{a numeric vector} 
\item{Sample}{a numeric vector} 
}}
\description{
Data for Exercise 6.65
}
\examples{

str(Short)
attach(Short)
hist(Parallax)
EDA(Parallax)
SIGN.test(Parallax,md=8.798)
t.test(Parallax,mu=8.798)
detach(Short)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
