% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Short}
\alias{Short}
\title{James Short's measurements of the parallax of the sun}
\format{
A data frame/tibble with 158 observations on two variables
\describe{
\item{sample}{sample number} 
\item{parallax}{parallax measurements (seconds of a degree)} 
}
}
\usage{
Short
}
\description{
Data for Exercise 6.65
}
\examples{

hist(Short$parallax, main = "Problem 6.65", 
xlab = "", col = "orange")
SIGN.test(Short$parallax, md = 8.798)
t.test(Short$parallax, mu = 8.798)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
