% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Shuttle}
\alias{Shuttle}
\title{Number of people riding shuttle versus number of automobiles in the downtown
area}
\format{
A data frame/tibble with 15 observations on two variables
\describe{ 
\item{users}{number of shuttle riders} 
\item{autos}{number of automobiles in the downtown area} 
}
}
\usage{
Shuttle
}
\description{
Data for Exercise 9.20
}
\examples{

plot(autos ~ users, data = Shuttle)
model <- lm(autos ~ users, data = Shuttle)
summary(model)
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
