% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Shuttle}
\alias{Shuttle}
\title{Number of people riding shuttle versus number of automobiles in the downtown
area}
\format{A data frame with 15 observations on the following 2 variables.
\describe{
\item{shuttle}{a numeric vector}
\item{autos}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 9.20
}
\examples{
str(Shuttle)
attach(Shuttle)
model <- lm(autos~shuttle)
summary(model)
detach(Shuttle)
remove(model)
}
\keyword{datasets}

