% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Situp}
\alias{Situp}
\title{Maximum number of situps by participants in an exercise class}
\format{A data frame with 20 observations on the following variable.
\describe{ 
\item{number}{a numeric vector} 
}}
\description{
Data for Exercise 1.47
}
\examples{

str(Situp)
attach(Situp)
stem(number)
hist(number,breaks=seq(0,70,10))
hist(number,breaks=seq(0,70,10),right=FALSE,col="blue",prob=TRUE,
main="Problems 1.46 & 1.47")
lines(density(number),col="red",lwd=3)
detach(Situp)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

