% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Skewed}
\alias{Skewed}
\title{Illustrates the Wilcoxon Rank Sum test}
\format{A data frame with 21 observations on the following 2 variables.
\describe{
\item{C1}{a numeric vector}
\item{C2}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.65
}
\examples{
str(Skewed)
attach(Skewed)
boxplot(C1,C2)
wilcox.test(C1,C2)
detach(Skewed)
}
\keyword{datasets}

