% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Skewed}
\alias{Skewed}
\title{Illustrates the Wilcoxon Rank Sum test}
\format{A data frame/tibble with 21 observations on two variables
\describe{ 
\item{C1}{values from a sample of size 16 from a particular population} 
\item{C2}{values from a sample of size 14 from a particular population} 
}}
\usage{
Skewed
}
\description{
Data for Exercise 7.65
}
\examples{

boxplot(Skewed$C1, Skewed$C2, col = c("pink", "lightblue"))
wilcox.test(Skewed$C1, Skewed$C2)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
