% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Skin}
\alias{Skin}
\title{Survival times of closely and poorly matched skin grafts on burn patients}
\format{A data frame with 11 observations on the following 2 variables.
\describe{ 
\item{close}{a numeric vector} 
\item{poor}{a numeric vector} 
}}
\description{
Data for Exercise 5.20
}
\examples{

str(Skin)
attach(Skin)
DIFF <- close-poor
stem(DIFF)
EDA(DIFF)
remove(DIFF)
detach(Skin)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

