% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Slc}
\alias{Slc}
\title{Sodium-lithium countertransport activity on 190 individuals from six large
English kindred}
\format{A data frame with 190 observations on the following variable.
\describe{
\item{SLC}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 5.116
}
\examples{
str(Slc)
attach(Slc)
EDA(SLC)
detach(Slc)
}
\keyword{datasets}

