% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Slc}
\alias{Slc}
\title{Sodium-lithium countertransport activity on 190 individuals from six large
English kindred}
\format{A data frame/tibble with 190 observations on the following variable.
\describe{ 
\item{slc}{Red blood cell sodium-lithium countertransport} 
}}
\source{
Roeder, K., (1994), "A Graphical Technique for Determining the Number of Components
in a Mixture of Normals," \emph{Journal of the American Statistical Association, 89}, 497-495.
}
\usage{
Slc
}
\description{
Data for Exercise 5.116
}
\examples{

BSDA::EDA(Slc$slc)
hist(Slc$slc, freq = FALSE, xlab = "sodium lithium countertransport",
     main = "", col = "lightblue")
     lines(density(Slc$slc), col = "purple")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
