% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Smokyph}
\alias{Smokyph}
\title{Water pH levels of 75 water samples taken in the Great Smoky Mountains}
\format{A data frame with 75 observations on the following 5 variables.
\describe{ 
\item{waterph}{a numeric vector} 
\item{code}{a factor with levels \code{high} \code{low}} 
\item{elev}{a numeric vector} 
\item{SRES1}{a numeric vector} 
\item{FITS1}{a numeric vector} 
}}
\description{
Data for Exercises 6.40, 6.59, 7.10, and 7.35
}
\examples{

str(Smokyph)
attach(Smokyph)
t.test(waterph,mu=7)
SIGN.test(waterph,md=7)
tapply(waterph,code,mean)
stripchart(waterph~code,method="stack",pch=19,col=c("red","blue"))
qqnorm(waterph[code=="low"])
qqnorm(waterph[code=="high"])
t.test(waterph[code=="low"],waterph[code=="high"])
t.test(waterph[code=="low"],waterph[code=="high"],conf.level=.90)$conf
detach(Smokyph)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

