% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Snore}
\alias{Snore}
\title{Snoring versus heart disease}
\format{A data frame/tibble with 2,484 observations on the following 2 variables.
\describe{ 
\item{snore}{factor with levels \code{nonsnorer}, \code{ocassional snorer}, \code{nearly every night}, and \code{snores every night}}
\item{heartdisease}{factor indicating whether the indiviudal has heart disease (\code{no} or \code{yes})} 
}}
\source{
Norton, P. and Dunn, E. (1985), Snoring as a Risk Factor for Disease, \emph{British Medical Journal, 291},
630-632.
}
\usage{
Snore
}
\description{
Data for Exercise 8.21
}
\examples{

T1 <- xtabs(~ heartdisease + snore, data = Snore)
T1
chisq.test(T1)
rm(T1)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
