% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Snow}
\alias{Snow}
\title{Concentration of microparticles in snowfields of Greenland and Antarctica}
\format{A data frame with 34 observations on the following 4 variables.
\describe{
\item{antarc}{a numeric vector}
\item{greenld}{a numeric vector}
\item{concent}{a numeric vector}
\item{site}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.87
}
\examples{
str(Snow)
attach(Snow)
boxplot(concent~site)
detach(Snow)
}
\keyword{datasets}

