% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Snow}
\alias{Snow}
\title{Concentration of microparticles in snowfields of Greenland and Antarctica}
\format{A data frame/tibble with 34 observations on two variables
\describe{ 
\item{concent}{concentration of microparticles from melted snow (in parts per billion)}
\item{site}{location of snow sample (\code{Antarctica} or \code{Greenland})} 
}}
\source{
Davis, J., \emph{Statistics and Data Analysis in Geology}, John Wiley, New York.
}
\usage{
Snow
}
\description{
Data for Exercise 7.87
}
\examples{

boxplot(concent ~ site, data = Snow, col = c("lightblue", "lightgreen"))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
