% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Soccer}
\alias{Soccer}
\title{Weights of 25 soccer players}
\format{A data frame with 25 observations on the following variable.
\describe{ 
\item{weight}{a numeric vector} 
}}
\description{
Data for Exercise 1.46
}
\examples{

str(Soccer)
attach(Soccer)
stem(weight,scale=2)
hist(weight,breaks=seq(110,210,10),col="orange",
main="Problem 1.46 \\n Weights of Soccer Players",right=FALSE)
detach(Soccer)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
