% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Social}
\alias{Social}
\title{Median income level for 25 social workers from North Carolina}
\format{A data frame with 25 observations on the following variable.
\describe{ 
\item{income}{a numeric vector} 
}}
\description{
Data for Exercise 6.63
}
\examples{

str(Social)
attach(Social)
SIGN.test(income,md=27500,alternative="less")
detach(Social)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

