% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Social}
\alias{Social}
\title{Median income level for 25 social workers from North Carolina}
\format{A data frame/tibble with 25 observations on one variable
\describe{ 
\item{income}{annual income (in dollars) of North Carolina social workers with less than five years experience.} 
}}
\usage{
Social
}
\description{
Data for Exercise 6.63
}
\examples{

BSDA::SIGN.test(Social$income, md = 27500, alternative = "less")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
