% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Sophomor}
\alias{Sophomor}
\title{Grade point averages, SAT scores and final grade in college algebra for 20
sophomores}
\format{A data frame/tibble with 20 observations on four variables
\describe{ 
\item{student}{identification number} 
\item{gpa}{grade point average} 
\item{sat}{SAT math score} 
\item{exam}{final exam grade in college algebra} 
}}
\usage{
Sophomor
}
\description{
Data for Exercise 2.42
}
\examples{

cor(Sophomor)
plot(exam ~ gpa, data = Sophomor)
\dontrun{
library(ggplot2)
ggplot2::ggplot(data = Sophomor, aes(x = gpa, y = exam)) + 
           geom_point()
           ggplot2::ggplot(data = Sophomor, aes(x = sat, y = exam)) + 
           geom_point()
}

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
