% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Sophomor}
\alias{Sophomor}
\title{Grade point averages, SAT scores and final grade in college algebra for 20
sophomores}
\format{A data frame with 20 observations on the following 4 variables.
\describe{
\item{Student}{a numeric vector}
\item{GPA}{a numeric vector}
\item{SAT}{a numeric vector}
\item{Exam}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 2.42
}
\examples{
str(Sophomor)
attach(Sophomor)
cor(Sophomor)
detach(Sophomor)
}
\keyword{datasets}

