% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{South}
\alias{South}
\title{Murder rates for 30 cities in the South}
\format{A data frame with 31 observations on the following variable.
\describe{ 
\item{rate}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 1.84
}
\examples{

str(South)
attach(South)
boxplot(rate,col="yellow",main="Problem 1.83")
detach(South)

}
\keyword{datasets}

