% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{South}
\alias{South}
\title{Murder rates for 30 cities in the South}
\format{A data frame/tibble with 31 observations on the following variable.
\describe{ 
\item{rate}{murder rate per 100,000 people} 
}}
\usage{
South
}
\description{
Data for Exercise 1.84
}
\examples{

boxplot(South$rate, col = "gray", ylab = "Murder rate per 100,000 people")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
