% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Speed}
\alias{Speed}
\title{Speed reading scores before and after a course on speed reading}
\format{
A data frame/tibble with 15 observations on four variables
\describe{ 
\item{before}{reading comprehension score before taking a speed-reading course} 
\item{after}{reading comprehension score after taking a speed-reading course} 
\item{differ}{after - before (comprehension reading scores)}
\item{signranks}{signed ranked differences} 
}
}
\usage{
Speed
}
\description{
Data for Exercise 7.58
}
\examples{

t.test(Speed$differ, alternative = "greater")
t.test(Speed$signranks, alternative = "greater")
wilcox.test(Pair(Speed$after, Speed$before) ~ 1, data = Speed, alternative = "greater")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
