% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Speed}
\alias{Speed}
\title{Speed reading scores before and after a course on speed reading}
\format{A data frame with 15 observations on the following 4 variables.
\describe{ 
\item{Before}{a numeric vector} 
\item{After}{a numeric vector} 
\item{differ}{a numeric vector}
\item{signrnks}{a numeric vector} 
}}
\description{
Data for Exercise 7.58
}
\examples{

str(Speed)
attach(Speed)
qqnorm(differ)
qqline(differ)
shapiro.test(differ)
t.test(After,Before,paired=TRUE,alternative="greater")
wilcox.test(After,Before,paired=TRUE,alternative="greater")
detach(Speed)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
