% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Spellers}
\alias{Spellers}
\title{Standardized spelling test scores for two fourth grade classes}
\format{A data frame/tibble with 10 observations on the following 2 variables.
\describe{ 
\item{teacher}{character variable with values \code{Fourth} and \code{Colleague}} 
\item{score}{score on a standardized spelling test} 
}}
\usage{
Spellers
}
\description{
Data for Exercise 7.82
}
\examples{

boxplot(score ~ teacher, data = Spellers)
t.test(score ~ teacher, data = Spellers)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
