% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Spellers}
\alias{Spellers}
\title{Standardized spelling test scores for two fourth grade classes}
\format{A data frame with 10 observations on the following 2 variables.
\describe{ 
\item{Fourth}{a numeric vector} 
\item{Colleag}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.82
}
\examples{

str(Spellers)
attach(Spellers)
t.test(Fourth,Colleag,alternative="greater")
detach(Spellers)

}
\keyword{datasets}

