% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Spellers}
\alias{Spellers}
\title{Standardized spelling test scores for two fourth grade classes}
\format{A data frame/tibble with ten observations on two variables
\describe{ 
\item{teacher}{character variable with values \code{Fourth} and \code{Colleague}} 
\item{score}{score on a standardized spelling test} 
}}
\usage{
Spellers
}
\description{
Data for Exercise 7.82
}
\examples{

boxplot(score ~ teacher, data = Spellers, col = "pink")
t.test(score ~ teacher, data = Spellers)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
