% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Spelling}
\alias{Spelling}
\title{Spelling scores for 9 eighth graders before and after a 2-week course of
instruction}
\format{A data frame/tibble with 9 observations on the following 3 variables.
\describe{ 
\item{before}{spelling score before a 2-week course of instruction} 
\item{after}{spelling score after a 2-week course of instruction} 
\item{differ}{after - before (spelling score)} 
}}
\usage{
Spelling
}
\description{
Data for Exercise 7.56
}
\examples{

qqnorm(Spelling$differ)
qqline(Spelling$differ)
shapiro.test(Spelling$differ)
t.test(Spelling$before, Spelling$after, paired = TRUE)
t.test(Spelling$differ)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
