% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Spelling}
\alias{Spelling}
\title{Spelling scores for 9 eighth graders before and after a 2-week course of
instruction}
\format{A data frame with 9 observations on the following 3 variables.
\describe{ 
\item{Before}{a numeric vector} 
\item{After}{a numeric vector} 
\item{differ}{a numeric vector} 
}}
\description{
Data for Exercise 7.56
}
\examples{

str(Spelling)
attach(Spelling)
qqnorm(differ)
qqline(differ)
shapiro.test(differ)
t.test(After,Before,paired=TRUE,alternative="greater")
detach(Spelling)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
