% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Sports}
\alias{Sports}
\title{Favorite sport by gender}
\format{
A data frame/tibble with 200 observations on two variables
\describe{ 
\item{gender}{a factor with levels \code{male} and \code{female}}
\item{sport}{a factor with levels \code{football}, \code{basketball}, 
\code{baseball}, and \code{tennis}} 
}
}
\usage{
Sports
}
\description{
Data for Exercise 8.32
}
\examples{

T1 <- xtabs(~gender + sport, data = Sports)
T1
chisq.test(T1)
rm(T1)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
