% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Sports}
\alias{Sports}
\title{Favorite sport by gender}
\format{A data frame with 2 observations on the following variable.
\describe{
\item{gender.football.basketbl.baseball.tennis}{a factor
with levels \code{female 3.800000000e+001 2.100000000e+001 1.500000000e+001
2.600000000e+001} \code{male 3.300000000e+001 3.800000000e+001
2.400000000e+001 5.000000000e+000}}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 8.32
}
\examples{
str(Sports)
attach(Sports)
Sports
detach(Sports)
}
\keyword{datasets}

