% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Spouse}
\alias{Spouse}
\title{Convictions in spouse murder cases by gender}
\format{A data frame with 4 observations on the following 3 variables.
\describe{ 
\item{result}{a factor with levels \code{acquitted}
\code{convicted} \code{not prosecuted} \code{pleaded guilty}}
\item{husband}{a numeric vector} 
\item{wife}{a numeric vector} 
}}
\description{
Data for Exercise 8.33
}
\examples{

str(Spouse)
attach(Spouse)
Spouse
chisq.test(Spouse[,2:3])
detach(Spouse)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

