% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Stable}
\alias{Stable}
\title{Times of a 2-year old stallion on a one mile run}
\format{A data frame with 9 observations on the following variable.
\describe{ 
\item{time}{a numeric vector} 
}}
\description{
Data for Exercise 6.93
}
\examples{

str(Stable)
attach(Stable)
EDA(time)
SIGN.test(time,md=98.5,alternative="greater")
detach(Stable)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

