% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Stamp}
\alias{Stamp}
\title{Thicknesses of 1872 Hidalgo stamps issued in Mexico}
\format{A data frame with 485 observations on the following 3 variables.
\describe{ 
\item{thickness}{a numeric vector}
\item{thick}{a numeric vector} 
\item{freq}{a numeric vector} 
}}
\description{
Data for Statistical Insight Chapter 1 and Exercise 5.110
}
\examples{

str(Stamp)
attach(Stamp)
hist(thickness,prob=TRUE,col="lightblue")
lines(density(thickness),lwd=2,col="blue")
t.test(thickness,conf.level=.99)$conf
detach(Stamp)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

