% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Statclas}
\alias{Statclas}
\title{Grades for two introductory statistics classes}
\format{
A data frame/tibble with 72 observations on two variables
\describe{ 
\item{class}{class meeting time (9am or 2pm)} 
\item{score}{grade for an introductory statistics class} 
}
}
\usage{
Statclas
}
\description{
Data for Exercise 7.30
}
\examples{

str(Statclas)
boxplot(score ~ class, data = Statclas, col = "red")
t.test(score ~ class, data = Statclas)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
