% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Statclas}
\alias{Statclas}
\title{Grades for two introductory statistics classes}
\format{A data frame with 36 observations on the following 2 variables.
\describe{ 
\item{X9am}{a numeric vector} 
\item{X2pm}{a numeric vector} 
}}
\description{
Data for Exercise 7.30
}
\examples{

str(Statclas)
attach(Statclas)
t.test(X9am,X2pm)
detach(Statclas)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
