% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Statisti}
\alias{Statisti}
\title{Test scores for two beginning statistics classes}
\format{A data frame with 31 observations on the following 2 variables.
\describe{ 
\item{Class1}{a numeric vector} 
\item{Class2}{a numeric vector} 
}}
\description{
Data for Exercises 1.70 and 1.87
}
\examples{

str(Statisti)
attach(Statisti)
boxplot(Class1,Class2,names=c("Class 1","Class 2"),col=c("red","blue"),
main="Problem 1.87")
detach(Statisti)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

