% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Step}
\alias{Step}
\title{STEP science test scores for a class of ability-grouped students}
\format{
A data frame/tibble with 12 observations on one variable
\describe{ 
\item{score}{State test of educational progress (STEP) science test score} 
}
}
\usage{
Step
}
\description{
Data for Exercise 6.79
}
\examples{

EDA(Step$score)
t.test(Step$score, mu = 80, alternative = "less")
wilcox.test(Step$score, mu = 80, alternative = "less")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
