% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Step}
\alias{Step}
\title{STEP science test scores for a class of ability-grouped students}
\format{A data frame with 12 observations on the following variable.
\describe{
\item{score}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 6.79
}
\examples{
str(Step)
attach(Step)
EDA(score)
t.test(score,mu=80,alternative="less")
detach(Step)
}
\keyword{datasets}

