% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Stress}
\alias{Stress}
\title{Short-term memory test scores on 12 subjects before and after a stressful
situation}
\format{A data frame with 12 observations on the following 2 variables.
\describe{ 
\item{Prestre}{a numeric vector}
\item{Poststre}{a numeric vector} 
}}
\description{
Data for Example 7.20
}
\examples{

str(Stress)
attach(Stress)
DIF <- Poststre -Prestre
qqnorm(DIF)
qqline(DIF)
shapiro.test(DIF)
t.test(Poststre,Prestre,paired=TRUE,alternative="less")
detach(Stress)
remove(DIF)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

