% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Study}
\alias{Study}
\title{Number of hours studied per week by a sample of 50 freshmen}
\format{A data frame with 50 observations on the following variable.
\describe{ 
\item{hours}{a numeric vector} 
}}
\description{
Data for Exercise 5.25
}
\examples{

str(Study)
attach(Study)
stem(hours)
EDA(hours)
detach(Study)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

