% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Study}
\alias{Study}
\title{Number of hours studied per week by a sample of 50 freshmen}
\format{A data frame/tibble with 50 observations on one variable
\describe{ 
\item{hours}{number of hours a week freshmen reported studying for their courses} 
}}
\usage{
Study
}
\description{
Data for Exercise 5.25
}
\examples{

stem(Study$hours)
hist(Study$hours, col = "violet")
summary(Study$hours)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
