% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Submarin}
\alias{Submarin}
\title{Number of German submarines sunk by U.S. Navy in World War II}
\format{A data frame with 16 observations on the following 3 variables.
\describe{ 
\item{Month}{a numeric vector} 
\item{reported}{a numeric vector} 
\item{actual}{a numeric vector} 
}}
\description{
Data for Exercises 2.16, 2.45, and 2.59
}
\examples{

str(Submarin)
attach(Submarin)
plot(reported,actual)
model <- lm(actual~reported)
abline(model)
anova(model)
summary(model)
detach(Submarin)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

