% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Submarin}
\alias{Submarin}
\title{Number of German submarines sunk by U.S. Navy in World War II}
\format{A data frame/tibble with 16 observations on the following 3 variables.
\describe{ 
\item{month}{month} 
\item{reported}{number of submarines reported sunk by U.S. Navy} 
\item{actual}{number of submarines actually sunk by U.S. Navy} 
}}
\source{
F. Mosteller, S. Fienberg, and R. Rourke, \emph{Beginning Statistics with Data Analysis}
(Reading, MA: Addison-Wesley, 1983)
}
\usage{
Submarin
}
\description{
Data for Exercises 2.16, 2.45, and 2.59
}
\examples{

model <- lm(actual ~ reported, data = Submarin)
summary(model)
plot(actual ~ reported, data = Submarin)
abline(model, col = "red")
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
