% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Subway}
\alias{Subway}
\title{Time it takes a subway to travel from the airport to downtown}
\format{A data frame with 30 observations on the following variable.
\describe{ 
\item{time}{a numeric vector} 
}}
\description{
Data for Exercise 5.19
}
\examples{

str(Subway)
attach(Subway)
EDA(time)
detach(Subway)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

