% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Subway}
\alias{Subway}
\title{Time it takes a subway to travel from the airport to downtown}
\format{A data frame/tibble with 30 observations on the following variable.
\describe{ 
\item{time}{time (in minutes) it takes a subway to travel from the airport to downtown} 
}}
\usage{
Subway
}
\description{
Data for Exercise 5.19
}
\examples{

hist(Subway$time, main = "Exercise 5.19", 
xlab = "Time (in minutes)", col = "purple")
summary(Subway$time)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
