% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Sunspot}
\alias{Sunspot}
\title{Wolfer sunspot numbers from 1700 through 2000}
\format{A data frame with 301 observations on the following 2 variables.
\describe{ 
\item{year}{a numeric vector} 
\item{sunspots}{a numeric vector} 
}}
\description{
Data for Example 1.7
}
\examples{

str(Sunspot)
attach(Sunspot)
plot(year,sunspots,type="l",main="Yearly Sunspots")    # Using standard plot
library(lattice)
xyplot(sunspots ~ 1700:2000, xlab = "", type = "l",main="Yearly Sunspots")
xyplot(sunspots ~ 1700:2000, xlab = "", type = "l", aspect="xy",
main="Yearly Sunspots")
detach(Sunspot)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

