% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Superbowl}
\alias{Superbowl}
\title{Margin of victory in Superbowls I to XXXV}
\format{A data frame/tibble with 35 observations on five variables
\describe{ 
\item{winning_team}{name of Suberbowl winning team}
\item{winner_score}{winning score for the Superbowl} 
\item{losing_team}{name of Suberbowl losing team}
\item{loser_score}{score of losing teama numeric vector} 
\item{victory_margin}{winner_score - loser_score} 
}}
\usage{
Superbowl
}
\description{
Data for Exercise 1.54
}
\examples{

stem(Superbowl$victory_margin)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
