% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Supercar}
\alias{Supercar}
\title{Top speeds attained by five makes of supercars}
\format{A data frame with 30 observations on the following 7 variables.
\describe{
\item{Acura}{a numeric vector}
\item{Ferrari}{a numeric vector}
\item{Lotus}{a numeric vector}
\item{Porsche}{a numeric vector}
\item{Viper}{a numeric vector}
\item{speed}{a numeric vector}
\item{car}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Statistical Insight Chapter 10
}
\examples{
str(Supercar)
attach(Supercar)
boxplot(speed~car)
anova(lm(speed~as.factor(car)))
detach(Supercar)
}
\keyword{datasets}

