% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Supercar}
\alias{Supercar}
\title{Top speeds attained by five makes of supercars}
\format{A data frame/tibble with 30 observations on two variables
\describe{
\item{speed}{top speed (in miles per hour) of car without redlining} 
\item{car}{name of sports car} 
}}
\source{
\emph{Car and Drvier} (July 1995).
}
\usage{
Supercar
}
\description{
Data for Statistical Insight Chapter 10
}
\examples{

boxplot(speed ~ car, data = Supercar, col = rainbow(6),
        ylab = "Speed (mph)")
summary(aov(speed ~ car, data = Supercar))
anova(lm(speed ~ car, data = Supercar))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
