% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Tablrock}
\alias{Tablrock}
\title{Ozone concentrations at Mt. Mitchell, North Carolina}
\format{A data frame with 719 observations on the following 16 variables.
\describe{ 
\item{hour}{a factor with levels \code{00:00}
\code{01:00} \code{02:00} \code{03:00} \code{04:00} \code{05:00}
\code{06:00} \code{07:00} \code{08:00} \code{09:00} \code{10:00}
\code{11:00} \code{12:00} \code{13:00} \code{14:00} \code{15:00}
\code{16:00} \code{17:00} \code{18:00} \code{19:00} \code{20:00}
\code{21:00} \code{22:00} \code{23:00}} 
\item{X03}{a numeric vector}
\item{tmp}{a numeric vector} 
\item{vdc}{a numeric vector}
\item{wd}{a numeric vector} 
\item{ws}{a numeric vector}
\item{amb}{a numeric vector} 
\item{dew}{a numeric vector}
\item{so2}{a numeric vector} 
\item{no}{a numeric vector}
\item{no2}{a numeric vector} 
\item{nox}{a numeric vector}
\item{co}{a numeric vector} 
\item{co2}{a numeric vector}
\item{gas}{a numeric vector} 
\item{air}{a numeric vector} 
}}
\description{
Data for Exercise 5.63
}
\examples{

str(Tablrock)
attach(Tablrock)
EDA(X03)
t.test(X03,conf.level=.99)$conf
detach(Tablrock)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

