% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Tenness}
\alias{Tenness}
\title{Tennessee self concept scores for 20 gifted high school students}
\format{A data frame with 20 observations on the following variable.
\describe{ 
\item{score}{a numeric vector} 
}}
\description{
Data for Exercise 6.56
}
\examples{

str(Tenness)
attach(Tenness)
EDA(score)
t.test(score,mu=30,alternative="less")
SIGN.test(score,md=30,alternative="less")
detach(Tenness)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

