% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Tensile}
\alias{Tensile}
\title{Tensile strength of plastic bags from two production runs}
\format{A data frame/tibble with 72 observations on the following 2 variables.
\describe{
\item{tensile}{plastic bag tensile strength (pounds per square inch)}
\item{run}{factor with run number (1 or 2)} 
}}
\usage{
Tensile
}
\description{
Data for Example 7.11
}
\examples{

boxplot(tensile ~ run, data = Tensile, 
col = c("purple", "cyan"))
t.test(tensile ~ run, data = Tensile)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
