% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Tensile}
\alias{Tensile}
\title{Tensile strength of plastic bags from two production runs}
\format{A data frame with 72 observations on the following 4 variables.
\describe{
\item{Run.1}{a numeric vector}
\item{Run.2}{a numeric vector}
\item{Tensile}{a numeric vector}
\item{Run}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Example 7.11
}
\examples{
str(Tensile)
attach(Tensile)
boxplot(Run.1,Run.2,names=c("Run 1","Run 2"),col=c("red","Blue"))
boxplot(Tensile~Run,names=c("Run 1","Run 2"),col=c("red","Blue"))
t.test(Tensile~Run)
t.test(Run.1,Run.2)
detach(Tensile)
}
\keyword{datasets}

