% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Test1}
\alias{Test1}
\title{Grades on the first test in a statistics class}
\format{A data frame with 25 observations on the following variable.
\describe{
\item{test1}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 5.80
}
\examples{
str(Test1)
attach(Test1)
EDA(test1)
detach(Test1)
}
\keyword{datasets}

