% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Test1}
\alias{Test1}
\title{Grades on the first test in a statistics class}
\format{A data frame with 25 observations on the following variable.
\describe{ 
\item{test1}{a numeric vector} 
}}
\description{
Data for Exercise 5.80
}
\examples{

str(Test1)
attach(Test1)
EDA(test1)
detach(Test1)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

