% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Test1}
\alias{Test1}
\title{Grades on the first test in a statistics class}
\format{A data frame/tibble with 25 observations on one variable
\describe{ 
\item{score}{score on first statistics exam} 
}}
\usage{
Test1
}
\description{
Data for Exercise 5.80
}
\examples{

stem(Test1$score)
boxplot(Test1$score)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
