% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Thermal}
\alias{Thermal}
\title{Heat loss of thermal pane windows versus outside temperature}
\format{A data frame with 12 observations on the following 3 variables.
\describe{ 
\item{temp}{a numeric vector} 
\item{loss}{a numeric vector} 
\item{x}{a numeric vector} 
}}
\description{
Data for Example 9.5
}
\examples{

str(Thermal)
attach(Thermal)
model <- lm(loss~temp)
summary(model)
detach(Thermal)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

