% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Thermal}
\alias{Thermal}
\title{Heat loss of thermal pane windows versus outside temperature}
\format{A data frame/tibble with 12 observations on the two variables
\describe{ 
\item{temp}{temperature (degrees Celcius)} 
\item{loss}{heat loss (BTUs)} 
}}
\usage{
Thermal
}
\description{
Data for Example 9.5
}
\examples{

model <- lm(loss ~ temp, data = Thermal)
summary(model)
plot(loss ~ temp, data = Thermal)
abline(model, col = "red")
rm(model)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
