% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Tiaa}
\alias{Tiaa}
\title{1999-2000 closing prices for TIAA-CREF stocks}
\format{A data frame with 365 observations on the following 4 variables.
\describe{ 
\item{crefstk}{a numeric vector} 
\item{crefgwt}{a numeric vector} 
\item{tiaa}{a numeric vector} 
\item{date}{a factor with levels \code{1/01/2000} \code{1/02/2000} \code{1/03/2000}
\code{1/04/2000} \code{1/05/2000} \code{1/06/2000} \code{1/07/2000}
\code{1/08/2000} \code{1/09/2000} \code{1/10/2000} \code{1/11/2000}
\code{1/12/2000} \code{1/13/2000} \code{1/14/2000} \code{1/15/2000}
\code{1/16/2000} \code{1/17/2000} \code{1/18/2000} \code{1/19/2000}
\code{1/20/2000} \code{1/21/2000} \code{1/22/2000} \code{1/23/2000}
\code{1/24/2000} \code{1/25/2000} \code{1/26/2000} \code{1/27/2000}
\code{1/28/2000} \code{1/29/2000} \code{1/30/2000} \code{1/31/2000}
\code{10/01/1999} \code{10/02/1999} \code{10/03/1999} \code{10/04/1999}
\code{10/05/1999} \code{10/06/1999} \code{10/07/1999} \code{10/08/1999}
\code{10/09/1999} \code{10/10/1999} \code{10/11/1999} \code{10/12/1999}
\code{10/13/1999} \code{10/14/1999} \code{10/15/1999} \code{10/16/1999}
\code{10/17/1999} \code{10/18/1999} \code{10/19/1999} \code{10/20/1999}
\code{10/21/1999} \code{10/22/1999} \code{10/23/1999} \code{10/24/1999}
\code{10/25/1999} \code{10/26/1999} \code{10/27/1999} \code{10/28/1999}
\code{10/29/1999} \code{10/30/1999} \code{10/31/1999} \code{11/01/1999}
\code{11/02/1999} \code{11/03/1999} \code{11/04/1999} \code{11/05/1999}
\code{11/06/1999} \code{11/07/1999} \code{11/08/1999} \code{11/09/1999}
\code{11/10/1999} \code{11/11/1999} \code{11/12/1999} \code{11/13/1999}
\code{11/14/1999} \code{11/15/1999} \code{11/16/1999} \code{11/17/1999}
\code{11/18/1999} \code{11/19/1999} \code{11/20/1999} \code{11/21/1999}
\code{11/22/1999} \code{11/23/1999} \code{11/24/1999} \code{11/25/1999}
\code{11/26/1999} \code{11/27/1999} \code{11/28/1999} \code{11/29/1999}
\code{11/30/1999} \code{12/01/1999} \code{12/02/1999} \code{12/03/1999}
\code{12/04/1999} \code{12/05/1999} \code{12/06/1999} \code{12/07/1999}
\code{12/08/1999} \code{12/09/1999} \code{12/10/1999} \code{12/11/1999}
\code{12/12/1999} \code{12/13/1999} \code{12/14/1999} \code{12/15/1999}
\code{12/16/1999} \code{12/17/1999} \code{12/18/1999} \code{12/19/1999}
\code{12/20/1999} \code{12/21/1999} \code{12/22/1999} \code{12/23/1999}
\code{12/24/1999} \code{12/25/1999} \code{12/26/1999} \code{12/27/1999}
\code{12/28/1999} \code{12/29/1999} \code{12/30/1999} \code{12/31/1999}
\code{2/01/2000} \code{2/02/2000} \code{2/03/2000} \code{2/04/2000}
\code{2/05/2000} \code{2/06/2000} \code{2/07/2000} \code{2/08/2000}
\code{2/09/2000} \code{2/10/2000} \code{2/11/2000} \code{2/12/2000}
\code{2/13/2000} \code{2/14/2000} \code{2/15/2000} \code{2/16/2000}
\code{2/17/2000} \code{2/18/2000} \code{2/19/2000} \code{2/20/2000}
\code{2/21/2000} \code{2/22/2000} \code{2/23/2000} \code{2/24/2000}
\code{2/25/2000} \code{2/26/2000} \code{2/27/2000} \code{2/28/2000}
\code{2/29/2000} \code{3/01/2000} \code{3/02/2000} \code{3/03/2000}
\code{3/04/2000} \code{3/05/2000} \code{3/06/2000} \code{3/07/2000}
\code{3/08/2000} \code{3/09/2000} \code{3/10/2000} \code{3/11/2000}
\code{3/12/2000} \code{3/13/2000} \code{3/14/2000} \code{3/15/2000}
\code{3/16/2000} \code{3/17/2000} \code{3/18/2000} \code{3/19/2000}
\code{3/20/2000} \code{3/21/2000} \code{3/22/2000} \code{3/23/2000}
\code{3/24/2000} \code{3/25/2000} \code{3/26/2000} \code{3/27/2000}
\code{3/28/2000} \code{3/29/2000} \code{3/30/2000} \code{3/31/2000}
\code{4/01/2000} \code{4/02/2000} \code{4/03/2000} \code{4/04/2000}
\code{4/05/2000} \code{4/06/2000} \code{4/07/2000} \code{4/08/2000}
\code{4/09/2000} \code{4/10/2000} \code{4/11/2000} \code{4/12/2000}
\code{4/13/2000} \code{4/14/2000} \code{4/16/1999} \code{4/17/1999}
\code{4/18/1999} \code{4/19/1999} \code{4/20/1999} \code{4/21/1999}
\code{4/22/1999} \code{4/23/1999} \code{4/24/1999} \code{4/25/1999}
\code{4/26/1999} \code{4/27/1999} \code{4/28/1999} \code{4/29/1999}
\code{4/30/1999} \code{5/01/1999} \code{5/02/1999} \code{5/03/1999}
\code{5/04/1999} \code{5/05/1999} \code{5/06/1999} \code{5/07/1999}
\code{5/08/1999} \code{5/09/1999} \code{5/10/1999} \code{5/11/1999}
\code{5/12/1999} \code{5/13/1999} \code{5/14/1999} \code{5/15/1999}
\code{5/16/1999} \code{5/17/1999} \code{5/18/1999} \code{5/19/1999}
\code{5/20/1999} \code{5/21/1999} \code{5/22/1999} \code{5/23/1999}
\code{5/24/1999} \code{5/25/1999} \code{5/26/1999} \code{5/27/1999}
\code{5/28/1999} \code{5/29/1999} \code{5/30/1999} \code{5/31/1999}
\code{6/01/1999} \code{6/02/1999} \code{6/03/1999} \code{6/04/1999}
\code{6/05/1999} \code{6/06/1999} \code{6/07/1999} \code{6/08/1999}
\code{6/09/1999} \code{6/10/1999} \code{6/11/1999} \code{6/12/1999}
\code{6/13/1999} \code{6/14/1999} \code{6/15/1999} \code{6/16/1999}
\code{6/17/1999} \code{6/18/1999} \code{6/19/1999} \code{6/20/1999}
\code{6/21/1999} \code{6/22/1999} \code{6/23/1999} \code{6/24/1999}
\code{6/25/1999} \code{6/26/1999} \code{6/27/1999} \code{6/28/1999}
\code{6/29/1999} \code{6/30/1999} \code{7/01/1999} \code{7/02/1999}
\code{7/03/1999} \code{7/04/1999} \code{7/05/1999} \code{7/06/1999}
\code{7/07/1999} \code{7/08/1999} \code{7/09/1999} \code{7/10/1999}
\code{7/11/1999} \code{7/12/1999} \code{7/13/1999} \code{7/14/1999}
\code{7/15/1999} \code{7/16/1999} \code{7/17/1999} \code{7/18/1999}
\code{7/19/1999} \code{7/20/1999} \code{7/21/1999} \code{7/22/1999}
\code{7/23/1999} \code{7/24/1999} \code{7/25/1999} \code{7/26/1999}
\code{7/27/1999} \code{7/28/1999} \code{7/29/1999} \code{7/30/1999}
\code{7/31/1999} \code{8/01/1999} \code{8/02/1999} \code{8/03/1999}
\code{8/04/1999} \code{8/05/1999} \code{8/06/1999} \code{8/07/1999}
\code{8/08/1999} \code{8/09/1999} \code{8/10/1999} \code{8/11/1999}
\code{8/12/1999} \code{8/13/1999} \code{8/14/1999} \code{8/15/1999}
\code{8/16/1999} \code{8/17/1999} \code{8/18/1999} \code{8/19/1999}
\code{8/20/1999} \code{8/21/1999} \code{8/22/1999} \code{8/23/1999}
\code{8/24/1999} \code{8/25/1999} \code{8/26/1999} \code{8/27/1999}
\code{8/28/1999} \code{8/29/1999} \code{8/30/1999} \code{8/31/1999}
\code{9/01/1999} \code{9/02/1999} \code{9/03/1999} \code{9/04/1999}
\code{9/05/1999} \code{9/06/1999} \code{9/07/1999} \code{9/08/1999}
\code{9/09/1999} \code{9/10/1999} \code{9/11/1999} \code{9/12/1999}
\code{9/13/1999} \code{9/14/1999} \code{9/15/1999} \code{9/16/1999}
\code{9/17/1999} \code{9/18/1999} \code{9/19/1999} \code{9/20/1999}
\code{9/21/1999} \code{9/22/1999} \code{9/23/1999} \code{9/24/1999}
\code{9/25/1999} \code{9/26/1999} \code{9/27/1999} \code{9/28/1999}
\code{9/29/1999} \code{9/30/1999}} 
}}
\description{
Data for your enjoyment
}
\examples{

data(Tiaa)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

