% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Ticket}
\alias{Ticket}
\title{Time to complete an airline ticket reservation}
\format{A data frame with 20 observations on the following variable.
\describe{ 
\item{time}{a numeric vector}
}}
\description{
Data for Exercise 5.18
}
\examples{

str(Ticket)
attach(Ticket)
EDA(time)
detach(Ticket)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
