% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Ticket}
\alias{Ticket}
\title{Time to complete an airline ticket reservation}
\format{
A data frame/tibble with 20 observations on one variable
\describe{ 
\item{time}{time (in seconds) to check out a reservation}
}
}
\usage{
Ticket
}
\description{
Data for Exercise 5.18
}
\examples{

EDA(Ticket$time)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
