% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Toaster}
\alias{Toaster}
\title{Consumer Reports (Oct 94) rating of toaster ovens versus the cost}
\format{A data frame/tibble with 17 observations on three variables
\describe{ 
\item{toaster}{name of toaster} 
\item{score}{Consumer Reports score}
\item{cost}{price of toaster (in dollars)} 
}}
\source{
\emph{Consumer Reports} (October 1994)
}
\usage{
Toaster
}
\description{
Data for Exercise 9.36
}
\examples{

plot(cost ~ score, data = Toaster)
model <- lm(cost ~ score, data = Toaster)
summary(model)
names(summary(model))
summary(model)$r.squared
plot(model, which = 1)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
