% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Tonsils}
\alias{Tonsils}
\title{Size of tonsils collected from 1,398 children}
\format{A data frame with 3 observations on the following 3 variables.
\describe{ 
\item{Size}{a factor with levels \code{Large}
\code{Normal} \code{Very Large}} 
\item{Carrier}{a numeric vector}
\item{Non.carrier}{a numeric vector} 
}}
\description{
Data for Exercise 2.78
}
\examples{

str(Tonsils)
attach(Tonsils)
TON <- as.matrix(Tonsils[,2:3])
rownames(TON) <- Size
TON
barplot(t(TON),beside=TRUE,legend=TRUE)
remove(TON)
detach(Tonsils)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

