% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Track15}
\alias{Track15}
\title{Olympic winning times for the men's 1500-meter run}
\format{A data frame with 26 observations on the following 2 variables.
\describe{ 
\item{Year}{a numeric vector} 
\item{X1500m}{a numeric vector} 
}}
\description{
Data for Exercise 1.36
}
\examples{

str(Track15)
attach(Track15)
plot(Year,X1500m,type="l",lwd=2,lty=2,col="red",xlab="Year",
ylab="1500m Time (seconds)",main="Problem 1.36")       
detach(Track15) 

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

