% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Track15}
\alias{Track15}
\title{Olympic winning times for the men's 1500-meter run}
\format{A data frame/tibble with 26 observations on two variables
\describe{ 
\item{year}{Olympic year} 
\item{time}{Olympic winning time (in seconds) for the 1500-meter run} 
}}
\source{
\emph{The World Almanac and Book of Facts}, 2000
}
\usage{
Track15
}
\description{
Data for Exercise 1.36
}
\examples{

plot(time~ year, data = Track15, type = "b", pch = 19,
ylab = "1500m time in seconds", col = "green") 

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
