% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Treatments}
\alias{Treatments}
\title{Illustrates analysis of variance for three treatment groups}
\format{A data frame/tibble with 24 observations on the following 2 variables.
\describe{
\item{score}{score from an experiment} 
\item{group}{factor with levels 1, 2, and 3} 
}}
\usage{
Treatments
}
\description{
Data for Exercise 10.44
}
\examples{

boxplot(score ~ group, data = Treatments)
summary(aov(score ~ group, data = Treatments))
summary(lm(score ~ group, data = Treatments))
anova(lm(score ~ group, data = Treatments))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
