% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Treatments}
\alias{Treatments}
\title{Illustrates analysis of variance for three treatment groups}
\format{A data frame with 24 observations on the following 5 variables.
\describe{ 
\item{treat1}{a numeric vector} 
\item{treat2}{a numeric vector} 
\item{treat3}{a numeric vector}
\item{Treatmnt}{a numeric vector} 
\item{Group}{a numeric vector} 
}}
\description{
Data for Exercise 10.44
}
\examples{

str(Treatments)
attach(Treatments)
anova(lm(Treatmnt~as.factor(Group)))
detach(Treatments)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
