% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Trees}
\alias{Trees}
\title{Number of trees in 20 grids}
\format{A data frame/tibble with 20 observations on one variable
\describe{ 
\item{number}{number of trees in a grid} 
}}
\usage{
Trees
}
\description{
Data for Exercise 1.50
}
\examples{

stem(Trees$number)
hist(Trees$number, main = "Exercise 1.50", xlab = "number",
     col = "brown")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
