% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Trucks}
\alias{Trucks}
\title{Miles per gallon for standard 4-wheel drive trucks manufactured by
Chevrolet, Dodge and Ford}
\format{A data frame/tibble with 15 observations on two variables
\describe{ 
\item{mpg}{miles per gallon} 
\item{truck}{a factor with levels \code{chevy}, \code{dodge}, and \code{ford}} 
}}
\usage{
Trucks
}
\description{
Data for Example 10.2
}
\examples{

boxplot(mpg ~ truck, data = Trucks, horizontal = TRUE, las = 1)
summary(aov(mpg ~ truck, data = Trucks))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
