% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Tv}
\alias{Tv}
\title{Percent of students that watch more than 6 hours of TV per day versus
national math test scores}
\format{A data frame/tibble with 53 observations on the following 3 variables.
\describe{ 
\item{state}{U.S. state}
\item{percent}{percent of students who watch more than six hours of TV a day} 
\item{test}{state average on national math test} 
}}
\source{
Educational Testing Services
}
\usage{
Tv
}
\description{
Data for Examples 2.1 and 2.7
}
\examples{

plot(test ~ percent, data = Tv, col = "blue")
cor(Tv$test, Tv$percent)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
