% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Twin}
\alias{Twin}
\title{Intelligence test scores for identical twins in which one twin is given a
drug}
\format{A data frame with 9 observations on the following 3 variables.
\describe{ 
\item{TwinA}{a numeric vector} 
\item{TwinB}{a numeric vector} 
\item{differ}{a numeric vector} 
}}
\description{
Data for Exercise 7.54
}
\examples{

str(Twin)
attach(Twin)
qqnorm(differ)
qqline(differ)
shapiro.test(differ)
t.test(TwinA,TwinB,paired=TRUE)
detach(Twin)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

