% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Twin}
\alias{Twin}
\title{Intelligence test scores for identical twins in which one twin is given a
drug}
\format{A data frame/tibble with 9 observations on the following 3 variables.
\describe{ 
\item{twinA}{score on intelligence test without drug} 
\item{twinB}{score on intelligence test after taking drug} 
\item{differ}{twinA - twinB} 
}}
\usage{
Twin
}
\description{
Data for Exercise 7.54
}
\examples{

qqnorm(Twin$differ)
qqline(Twin$differ)
shapiro.test(Twin$differ)
t.test(Twin$twinA, Twin$twinB, paired = TRUE)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
