% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Vacation}
\alias{Vacation}
\title{Number of days of paid holidays and vacation leave for sample of 35 textile
workers}
\format{
A data frame/tibble with 35 observations on one variable
\describe{ 
\item{number}{number of days of paid holidays and vacation leave taken} 
}
}
\usage{
Vacation
}
\description{
Data for Exercise 6.46 and 6.98
}
\examples{

boxplot(Vacation$number, col = "violet")
hist(Vacation$number, main = "Exercise 6.46", col = "blue",
     xlab = "number of days of paid holidays and vacation leave taken")
     t.test(Vacation$number, mu = 24)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
