% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Vacation}
\alias{Vacation}
\title{Number of days of paid holidays and vacation leave for sample of 35 textile
workers}
\format{A data frame with 35 observations on the following variable.
\describe{ 
\item{number}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 6.46 and 6.98
}
\examples{

str(Vacation)
attach(Vacation)
EDA(number)
t.test(number,mu=24)
detach(Vacation)

}
\keyword{datasets}

