% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Vaccine}
\alias{Vaccine}
\title{Reported serious reactions due to vaccines in 11 southern states}
\format{A data frame with 11 observations on the following 2 variables.
\describe{ 
\item{State}{a factor with levels \code{Alabama}
\code{Arkansas} \code{Florida} \code{Georgia} \code{Louisiana}
\code{Mississippi} \code{North Carolina} \code{Oklahoma} \code{South
Carolina} \code{Tennessee} \code{Texas}} 
\item{number}{a numeric vector} 
}}
\description{
Data for Exercise 1.111
}
\examples{

str(Vaccine)
attach(Vaccine)
fn <- fivenum(number)
fn
iqr <- IQR(number)
ll <- fn[2]-1.5*iqr
ul <- fn[4]+1.5*iqr
limits <- c(ll,ul)
limits
boxplot(number)
detach(Vaccine)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

