% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Vaccine}
\alias{Vaccine}
\title{Reported serious reactions due to vaccines in 11 southern states}
\format{A data frame/tibble with 11 observations on two variables
\describe{ 
\item{state}{U.S. state} 
\item{number}{number of reported serious reactions per million doses of a vaccine} 
}}
\source{
Center for Disease Control, Atlanta, Georgia.
}
\usage{
Vaccine
}
\description{
Data for Exercise 1.111
}
\examples{

stem(Vaccine$number, scale = 2) 
fn <- fivenum(Vaccine$number)
fn
iqr <- IQR(Vaccine$number)
iqr

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
