% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Vehicle}
\alias{Vehicle}
\title{Fatality ratings for foreign and domestic vehicles}
\format{A data frame with 2 observations on the following 6 variables.
\describe{ 
\item{make}{a factor with levels \code{domestic}
\code{foreign}} 
\item{A}{a numeric vector} 
\item{B}{a numeric vector} 
\item{C}{a numeric vector} 
\item{D}{a numeric vector} 
\item{F}{a numeric vector} 
}}
\description{
Data for Exercise 8.34
}
\examples{

str(Vehicle)
attach(Vehicle)
Vehicle
chisq.test(Vehicle[,2:6])
detach(Vehicle)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

