% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Vehicle}
\alias{Vehicle}
\title{Fatality ratings for foreign and domestic vehicles}
\format{A data frame/tibble with 151 observations on two variables
\describe{ 
\item{make}{a factor with levels \code{domestic} and \code{foreign}} 
\item{rating}{a factor with levels \code{Much better than average}, \code{Above average}, \code{Average}, \code{Below average}, and \code{Much worse than average}} 
}}
\source{
Insurance Institute for Highway Safety and the Highway Loss Data Institute, 1995
}
\usage{
Vehicle
}
\description{
Data for Exercise 8.34
}
\examples{

T1 <- xtabs(~make + rating, data = Vehicle)
T1
chisq.test(T1)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
