% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Verbal}
\alias{Verbal}
\title{Verbal test scores and number of library books checked out for 15 eighth
graders}
\format{
A data frame/tibble with 15 observations on two variables
\describe{ 
\item{number}{number of library books checked out} 
\item{verbal}{verbal test score}
}
}
\usage{
Verbal
}
\description{
Data for Exercise 9.30
}
\examples{

plot(verbal ~ number, data = Verbal)
abline(lm(verbal ~ number, data = Verbal), col = "red")
summary(lm(verbal ~ number, data = Verbal))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
