% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Verbal}
\alias{Verbal}
\title{Verbal test scores and number of library books checked out for 15 eighth
graders}
\format{A data frame with 15 observations on the following 4 variables.
\describe{ 
\item{number}{a numeric vector} 
\item{verbal}{a numeric vector} 
\item{SRES1}{a numeric vector}
\item{FITS1}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 9.30
}
\examples{

str(Verbal)
attach(Verbal)
model <- lm(verbal~number)
summary(model)
detach(Verbal)
remove(model)

}
\keyword{datasets}

