% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Victoria}
\alias{Victoria}
\title{Number of sunspots versus mean annual level of Lake Victoria Nyanza from
1902 to 1921}
\format{A data frame with 20 observations on the following 3 variables.
\describe{ 
\item{year}{a numeric vector} 
\item{level}{a numeric vector} 
\item{sunspot}{a numeric vector} 
}}
\description{
Data for Exercise 2.98
}
\examples{

str(Victoria)
attach(Victoria)
plot(sunspot,level)
model <- lm(level~sunspot)
abline(model)
cor(sunspot,level)
model
detach(Victoria)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
