% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Viscosit}
\alias{Viscosit}
\title{Viscosity measurements of a substance on two different days}
\format{A data frame with 11 observations on the following 2 variables.
\describe{ 
\item{first}{a numeric vector} 
\item{second}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.44
}
\examples{

str(Viscosit)
attach(Viscosit)
t.test(first,second)
detach(Viscosit)

}
\keyword{datasets}

