% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Viscosit}
\alias{Viscosit}
\title{Viscosity measurements of a substance on two different days}
\format{
A data frame/tibble with 11 observations on two variables
\describe{ 
\item{first}{viscosity measurement for a certain substance on day one} 
\item{second}{viscosity measurement for a certain substance on day two} 
}
}
\usage{
Viscosit
}
\description{
Data for Exercise 7.44
}
\examples{

boxplot(Viscosit$first, Viscosit$second, col = "blue")
t.test(Viscosit$first, Viscosit$second, var.equal = TRUE)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
