% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Visual}
\alias{Visual}
\title{Visual acuity of a group of subjects tested under a specified dose of a drug}
\format{A data frame with 18 observations on the following variable.
\describe{
\item{visual}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 5.6
}
\examples{
str(Visual)
attach(Visual)
stem(visual)
fivenum(visual)
boxplot(visual)
detach(Visual)
}
\keyword{datasets}

