% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Visual}
\alias{Visual}
\title{Visual acuity of a group of subjects tested under a specified dose of a drug}
\format{
A data frame/tibble with 18 observations on one variable
\describe{ 
\item{visual}{visual acuity measurement} 
}
}
\usage{
Visual
}
\description{
Data for Exercise 5.6
}
\examples{

stem(Visual$visual)
boxplot(Visual$visual, col = "purple")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
