% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Vocab}
\alias{Vocab}
\title{Reading scores before and after vocabulary training for 14 employees who did
not complete high school}
\format{A data frame/tibble with 14 observations on two variables
\describe{ 
\item{first}{reading test score before formal vocabulary training} 
\item{second}{reading test score after formal vocabulary training} 
}}
\usage{
Vocab
}
\description{
Data for Exercise 7.80
}
\examples{

t.test(Vocab$first, Vocab$second, paired = TRUE)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
