% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Vocab}
\alias{Vocab}
\title{Reading scores before and after vocabulary training for 14 employees who did
not complete high school}
\format{A data frame with 14 observations on the following 2 variables.
\describe{ 
\item{First}{a numeric vector} 
\item{Second}{a numeric vector} 
}}
\description{
Data for Exercise 7.80
}
\examples{

str(Vocab)
attach(Vocab)
DIF <- Second - First
qqnorm(DIF)
qqline(DIF)
shapiro.test(DIF)
t.test(Second,First,paired=TRUE)
detach(Vocab)
remove(DIF)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

