% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Wastewat}
\alias{Wastewat}
\title{Volume of injected waste water from Rocky Mountain Arsenal and number of
earthquakes near Denver}
\format{A data frame with 44 observations on the following 4 variables.
\describe{ 
\item{gallons}{a numeric vector} 
\item{number}{a numeric vector} 
\item{ln.no.}{a numeric vector}
\item{index}{a numeric vector} 
}}
\description{
Data for Exercise 9.18
}
\examples{

str(Wastewat)
attach(Wastewat)
model <- lm(number~gallons)
summary(model)
detach(Wastewat)
remove(model)

<<<<<<< HEAD
=======
}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051
}
\keyword{datasets}

