% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Weather94}
\alias{Weather94}
\title{Weather casualties in 1994}
\format{A data frame with 11 observations on the following 2 variables.
\describe{ 
\item{Weather.Type}{a factor with levels \code{Extreme
Temp} \code{Flash flood} \code{Fog} \code{High wind} \code{Hurricane}
\code{Lightning} \code{Other} \code{River flood} \code{Thunderstorm}
\code{Tornado} \code{Winter weather}} 
\item{Number}{a numeric vector} 
}}
\description{
Data for Exercise 1.30
}
\examples{

str(Weather94)
attach(Weather94)
names(Number) <- Weather.Type
barplot(Number,col="lightblue",las=2,cex.names=.65,main="Problem 1.30") 
# las=2 places bar names vertically
detach(Weather94)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}

