% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Wheat}
\alias{Wheat}
\title{Price of a bushel of wheat versus the national weekly earnings of production
workers}
\format{A data frame/tibble with 19 observations on three variables
\describe{ 
\item{year}{year} 
\item{earnings}{national weekly earnings (in dollars) for production workers} 
\item{price}{price for a bushel of wheat (in dollars)} 
}}
\source{
\emph{The World Almanac and Book of Facts}, 2000.
}
\usage{
Wheat
}
\description{
Data for Exercise 2.11
}
\examples{

par(mfrow = c(1, 2))
plot(earnings ~ year, data = Wheat)
plot(price ~ year, data = Wheat)
par(mfrow = c(1, 1))

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
