% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Windmill}
\alias{Windmill}
\title{Direct current produced by different wind velocities}
\format{A data frame with 25 observations on the following 7 variables.
\describe{ 
\item{velocity}{a numeric vector} 
\item{output}{a numeric vector} 
\item{SRES1}{a numeric vector}
\item{FITS1}{a numeric vector} 
\item{X1.velocity}{a numeric vector} 
\item{SRES2}{a numeric vector} 
\item{FITS2}{a numeric vector} 
}}
\description{
Data for Exercise 9.34
}
\examples{

str(Windmill)
attach(Windmill)
summary(lm(output~velocity))
detach(Windmill)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
