% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Window}
\alias{Window}
\title{Wind leakage for storm windows exposed to a 50 mph wind}
\format{A data frame/tibble with 9 observations on the following 2 variables.
\describe{ 
\item{window}{window number} 
\item{leakage}{percent leakage from a 50 mph wind} 
}}
\usage{
Window
}
\description{
Data for Exercise 6.54
}
\examples{

BSDA::SIGN.test(Window$leakage, md = 0.125, alternative = "greater")

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
