% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Window}
\alias{Window}
\title{Wind leakage for storm windows exposed to a 50 mph wind}
\format{A data frame with 9 observations on the following 2 variables.
\describe{
\item{Window}{a numeric vector}
\item{Leakage}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 6.54
}
\examples{
str(Window)
attach(Window)
SIGN.test(Leakage,md=.125,alternative="greater")
detach(Window)
}
\keyword{datasets}

