% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Wins}
\alias{Wins}
\title{Baseball team wins versus 7 independent variables for National league teams
in 1990}
\format{A data frame with 12 observations on the following 9 variables.
\describe{ 
\item{team}{a factor with levels \code{Atlanta}
\code{Chicago} \code{Cincinnati} \code{Houston} \code{Los Angeles}
\code{Montreal} \code{New York} \code{Philadelphia} \code{Pittsburgh}
\code{San Diego} \code{San Francisco} \code{St. Louis}}
\item{wins}{a numeric vector} 
\item{batavg}{a numeric vector} 
\item{rbi}{a numeric vector} 
\item{stole}{a numeric vector} 
\item{strkout}{a numeric vector} 
\item{caught}{a numeric vector} 
\item{errors}{a numeric vector} 
\item{era}{a numeric vector} 
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 9.23
}
\examples{

str(Wins)
attach(Wins)
plot(era,wins)
model <- lm(wins~era)
abline(model)
summary(model)
detach(Wins)
remove(model)

}
\keyword{datasets}

