% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Wool}
\alias{Wool}
\title{Strength tests of two types of wool fabric}
\format{A data frame/tibble with 20 observations on two variables
\describe{ 
\item{type}{type of wool (\code{Type I}, \code{Type 2})} 
\item{strength}{strength of wool} 
}}
\usage{
Wool
}
\description{
Data for Exercise 7.42
}
\examples{

boxplot(strength ~ type, data = Wool, col = c("blue", "purple"))
t.test(strength ~ type, data = Wool, var.equal = TRUE)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\keyword{datasets}
