% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Wool}
\alias{Wool}
\title{Strength tests of two types of wool fabric}
\format{A data frame with 10 observations on the following 2 variables.
\describe{
\item{Type.1}{a numeric vector}
\item{Type.2}{a numeric vector}
}}
\source{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Duxbury
}
\description{
Data for Exercise 7.42
}
\examples{
str(Wool)
attach(Wool)
t.test(Type.1,Type.2,var.equal=TRUE)
detach(Wool)
}
\keyword{datasets}

