% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSDA-package.R
\docType{data}
\name{Yearsunspot}
\alias{Yearsunspot}
\title{Monthly sunspot activity from 1974 to 2000}
\format{A data frame/tibble with 252 observations on two variables
\describe{ 
\item{number}{average number of sunspots} 
\item{year}{date} 
}}
\source{
NASA/Marshall Space Flight Center, Huntsville, AL 35812.
}
\usage{
Yearsunspot
}
\description{
Data for Exercise 2.7
}
\examples{

plot(number ~ year, data = Yearsunspot)

}
\references{
Kitchens, L. J. (2003) \emph{Basic Statistics and Data Analysis}.
Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.
}
\keyword{datasets}
